feat(react-dev-overlay): accept preventDisplay prop to allow config of which errors are display (used by next live)
