+++ packages/next/compiled/@next/react-dev-overlay/client.js
@@ -4,4 +4,4 @@
  * Copyright 2014-2020 Benjamin Tan
  * Copyright 2011-2013 John-David Dalton
  * Available under MIT license
- */(function(){"use strict";var t={function:true,object:true};var r=t[typeof window]&&window||this;var a=r;var i=t[typeof n]&&n;var o=t["object"]&&e&&!e.nodeType&&e;var l=i&&o&&typeof global=="object"&&global;if(l&&(l.global===l||l.window===l||l.self===l)){r=l}var s=Math.pow(2,53)-1;var u=/\bOpera/;var c=this;var f=Object.prototype;var d=f.hasOwnProperty;var p=f.toString;function capitalize(e){e=String(e);return e.charAt(0).toUpperCase()+e.slice(1)}function cleanupOS(e,n,t){var r={"10.0":"10",6.4:"10 Technical Preview",6.3:"8.1",6.2:"8",6.1:"Server 2008 R2 / 7","6.0":"Server 2008 / Vista",5.2:"Server 2003 / XP 64-bit",5.1:"XP",5.01:"2000 SP1","5.0":"2000","4.0":"NT","4.90":"ME"};if(n&&t&&/^Win/i.test(e)&&!/^Windows Phone /i.test(e)&&(r=r[/[\d.]+$/.exec(e)])){e="Windows "+r}e=String(e);if(n&&t){e=e.replace(RegExp(n,"i"),t)}e=format(e.replace(/ ce$/i," CE").replace(/\bhpw/i,"web").replace(/\bMacintosh\b/,"Mac OS").replace(/_PowerPC\b/i," OS").replace(/\b(OS X) [^ \d]+/i,"$1").replace(/\bMac (OS X)\b/,"$1").replace(/\/(\d)/," $1").replace(/_/g,".").replace(/(?: BePC|[ .]*fc[ \d.]+)$/i,"").replace(/\bx86\.64\b/gi,"x86_64").replace(/\b(Windows Phone) OS\b/,"$1").replace(/\b(Chrome OS \w+) [\d.]+\b/,"$1").split(" on ")[0]);return e}function each(e,n){var t=-1,r=e?e.length:0;if(typeof r=="number"&&r>-1&&r<=s){while(++t<r){n(e[t],t,e)}}else{forOwn(e,n)}}function format(e){e=trim(e);return/^(?:webOS|i(?:OS|P))/.test(e)?e:capitalize(e)}function forOwn(e,n){for(var t in e){if(d.call(e,t)){n(e[t],t,e)}}}function getClassOf(e){return e==null?capitalize(e):p.call(e).slice(8,-1)}function isHostType(e,n){var t=e!=null?typeof e[n]:"number";return!/^(?:boolean|number|string|undefined)$/.test(t)&&(t=="object"?!!e[n]:true)}function qualify(e){return String(e).replace(/([ -])(?!$)/g,"$1?")}function reduce(e,n){var t=null;each(e,(function(r,a){t=n(t,r,a,e)}));return t}function trim(e){return String(e).replace(/^ +| +$/g,"")}function parse(e){var n=r;var t=e&&typeof e=="object"&&getClassOf(e)!="String";if(t){n=e;e=null}var i=n.navigator||{};var o=i.userAgent||"";e||(e=o);var l=t||c==a;var s=t?!!i.likeChrome:/\bChrome\b/.test(e)&&!/internal|\n/i.test(p.toString());var f="Object",d=t?f:"ScriptBridgingProxyObject",v=t?f:"Environment",b=t&&n.java?"JavaPackage":getClassOf(n.java),m=t?f:"RuntimeObject";var h=/\bJava/.test(b)&&n.java;var g=h&&getClassOf(n.environment)==v;var y=h?"a":"α";var x=h?"b":"β";var w=n.document||{};var _=n.operamini||n.opera;var E=u.test(E=t&&_?_["[[Class]]"]:getClassOf(_))?E:_=null;var O;var j=e;var S=[];var k=null;var T=e==o;var C=T&&_&&typeof _.version=="function"&&_.version();var P;var M=getLayout([{label:"EdgeHTML",pattern:"Edge"},"Trident",{label:"WebKit",pattern:"AppleWebKit"},"iCab","Presto","NetFront","Tasman","KHTML","Gecko"]);var A=getName(["Adobe AIR","Arora","Avant Browser","Breach","Camino","Electron","Epiphany","Fennec","Flock","Galeon","GreenBrowser","iCab","Iceweasel","K-Meleon","Konqueror","Lunascape","Maxthon",{label:"Microsoft Edge",pattern:"(?:Edge|Edg|EdgA|EdgiOS)"},"Midori","Nook Browser","PaleMoon","PhantomJS","Raven","Rekonq","RockMelt",{label:"Samsung Internet",pattern:"SamsungBrowser"},"SeaMonkey",{label:"Silk",pattern:"(?:Cloud9|Silk-Accelerated)"},"Sleipnir","SlimBrowser",{label:"SRWare Iron",pattern:"Iron"},"Sunrise","Swiftfox","Vivaldi","Waterfox","WebPositive",{label:"Yandex Browser",pattern:"YaBrowser"},{label:"UC Browser",pattern:"UCBrowser"},"Opera Mini",{label:"Opera Mini",pattern:"OPiOS"},"Opera",{label:"Opera",pattern:"OPR"},"Chromium","Chrome",{label:"Chrome",pattern:"(?:HeadlessChrome)"},{label:"Chrome Mobile",pattern:"(?:CriOS|CrMo)"},{label:"Firefox",pattern:"(?:Firefox|Minefield)"},{label:"Firefox for iOS",pattern:"FxiOS"},{label:"IE",pattern:"IEMobile"},{label:"IE",pattern:"MSIE"},"Safari"]);var R=getProduct([{label:"BlackBerry",pattern:"BB10"},"BlackBerry",{label:"Galaxy S",pattern:"GT-I9000"},{label:"Galaxy S2",pattern:"GT-I9100"},{label:"Galaxy S3",pattern:"GT-I9300"},{label:"Galaxy S4",pattern:"GT-I9500"},{label:"Galaxy S5",pattern:"SM-G900"},{label:"Galaxy S6",pattern:"SM-G920"},{label:"Galaxy S6 Edge",pattern:"SM-G925"},{label:"Galaxy S7",pattern:"SM-G930"},{label:"Galaxy S7 Edge",pattern:"SM-G935"},"Google TV","Lumia","iPad","iPod","iPhone","Kindle",{label:"Kindle Fire",pattern:"(?:Cloud9|Silk-Accelerated)"},"Nexus","Nook","PlayBook","PlayStation Vita","PlayStation","TouchPad","Transformer",{label:"Wii U",pattern:"WiiU"},"Wii","Xbox One",{label:"Xbox 360",pattern:"Xbox"},"Xoom"]);var N=getManufacturer({Apple:{iPad:1,iPhone:1,iPod:1},Alcatel:{},Archos:{},Amazon:{Kindle:1,"Kindle Fire":1},Asus:{Transformer:1},"Barnes & Noble":{Nook:1},BlackBerry:{PlayBook:1},Google:{"Google TV":1,Nexus:1},HP:{TouchPad:1},HTC:{},Huawei:{},Lenovo:{},LG:{},Microsoft:{Xbox:1,"Xbox One":1},Motorola:{Xoom:1},Nintendo:{"Wii U":1,Wii:1},Nokia:{Lumia:1},Oppo:{},Samsung:{"Galaxy S":1,"Galaxy S2":1,"Galaxy S3":1,"Galaxy S4":1},Sony:{PlayStation:1,"PlayStation Vita":1},Xiaomi:{Mi:1,Redmi:1}});var D=getOS(["Windows Phone","KaiOS","Android","CentOS",{label:"Chrome OS",pattern:"CrOS"},"Debian",{label:"DragonFly BSD",pattern:"DragonFly"},"Fedora","FreeBSD","Gentoo","Haiku","Kubuntu","Linux Mint","OpenBSD","Red Hat","SuSE","Ubuntu","Xubuntu","Cygwin","Symbian OS","hpwOS","webOS ","webOS","Tablet OS","Tizen","Linux","Mac OS X","Macintosh","Mac","Windows 98;","Windows "]);function getLayout(n){return reduce(n,(function(n,t){return n||RegExp("\\b"+(t.pattern||qualify(t))+"\\b","i").exec(e)&&(t.label||t)}))}function getManufacturer(n){return reduce(n,(function(n,t,r){return n||(t[R]||t[/^[a-z]+(?: +[a-z]+\b)*/i.exec(R)]||RegExp("\\b"+qualify(r)+"(?:\\b|\\w*\\d)","i").exec(e))&&r}))}function getName(n){return reduce(n,(function(n,t){return n||RegExp("\\b"+(t.pattern||qualify(t))+"\\b","i").exec(e)&&(t.label||t)}))}function getOS(n){return reduce(n,(function(n,t){var r=t.pattern||qualify(t);if(!n&&(n=RegExp("\\b"+r+"(?:/[\\d.]+|[ \\w.]*)","i").exec(e))){n=cleanupOS(n,r,t.label||t)}return n}))}function getProduct(n){return reduce(n,(function(n,t){var r=t.pattern||qualify(t);if(!n&&(n=RegExp("\\b"+r+" *\\d+[.\\w_]*","i").exec(e)||RegExp("\\b"+r+" *\\w+-[\\w]*","i").exec(e)||RegExp("\\b"+r+"(?:; *(?:[a-z]+[_-])?[a-z]+\\d+|[^ ();-]*)","i").exec(e))){if((n=String(t.label&&!RegExp(r,"i").test(t.label)?t.label:n).split("/"))[1]&&!/[\d.]+/.test(n[0])){n[0]+=" "+n[1]}t=t.label||t;n=format(n[0].replace(RegExp(r,"i"),t).replace(RegExp("; *(?:"+t+"[_-])?","i")," ").replace(RegExp("("+t+")[-_.]?(\\w)","i"),"$1 $2"))}return n}))}function getVersion(n){return reduce(n,(function(n,t){return n||(RegExp(t+"(?:-[\\d.]+/|(?: for [\\w-]+)?[ /-])([\\d.]+[^ ();/_-]*)","i").exec(e)||0)[1]||null}))}function toStringPlatform(){return this.description||""}M&&(M=[M]);if(/\bAndroid\b/.test(D)&&!R&&(O=/\bAndroid[^;]*;(.*?)(?:Build|\) AppleWebKit)\b/i.exec(e))){R=trim(O[1]).replace(/^[a-z]{2}-[a-z]{2};\s*/i,"")||null}if(N&&!R){R=getProduct([N])}else if(N&&R){R=R.replace(RegExp("^("+qualify(N)+")[-_.\\s]","i"),N+" ").replace(RegExp("^("+qualify(N)+")[-_.]?(\\w)","i"),N+" $2")}if(O=/\bGoogle TV\b/.exec(R)){R=O[0]}if(/\bSimulator\b/i.test(e)){R=(R?R+" ":"")+"Simulator"}if(A=="Opera Mini"&&/\bOPiOS\b/.test(e)){S.push("running in Turbo/Uncompressed mode")}if(A=="IE"&&/\blike iPhone OS\b/.test(e)){O=parse(e.replace(/like iPhone OS/,""));N=O.manufacturer;R=O.product}else if(/^iP/.test(R)){A||(A="Safari");D="iOS"+((O=/ OS ([\d_]+)/i.exec(e))?" "+O[1].replace(/_/g,"."):"")}else if(A=="Konqueror"&&/^Linux\b/i.test(D)){D="Kubuntu"}else if(N&&N!="Google"&&(/Chrome/.test(A)&&!/\bMobile Safari\b/i.test(e)||/\bVita\b/.test(R))||/\bAndroid\b/.test(D)&&/^Chrome/.test(A)&&/\bVersion\//i.test(e)){A="Android Browser";D=/\bAndroid\b/.test(D)?D:"Android"}else if(A=="Silk"){if(!/\bMobi/i.test(e)){D="Android";S.unshift("desktop mode")}if(/Accelerated *= *true/i.test(e)){S.unshift("accelerated")}}else if(A=="UC Browser"&&/\bUCWEB\b/.test(e)){S.push("speed mode")}else if(A=="PaleMoon"&&(O=/\bFirefox\/([\d.]+)\b/.exec(e))){S.push("identifying as Firefox "+O[1])}else if(A=="Firefox"&&(O=/\b(Mobile|Tablet|TV)\b/i.exec(e))){D||(D="Firefox OS");R||(R=O[1])}else if(!A||(O=!/\bMinefield\b/i.test(e)&&/\b(?:Firefox|Safari)\b/.exec(A))){if(A&&!R&&/[\/,]|^[^(]+?\)/.test(e.slice(e.indexOf(O+"/")+8))){A=null}if((O=R||N||D)&&(R||N||/\b(?:Android|Symbian OS|Tablet OS|webOS)\b/.test(D))){A=/[a-z]+(?: Hat)?/i.exec(/\bAndroid\b/.test(D)?D:O)+" Browser"}}else if(A=="Electron"&&(O=(/\bChrome\/([\d.]+)\b/.exec(e)||0)[1])){S.push("Chromium "+O)}if(!C){C=getVersion(["(?:Cloud9|CriOS|CrMo|Edge|Edg|EdgA|EdgiOS|FxiOS|HeadlessChrome|IEMobile|Iron|Opera ?Mini|OPiOS|OPR|Raven|SamsungBrowser|Silk(?!/[\\d.]+$)|UCBrowser|YaBrowser)","Version",qualify(A),"(?:Firefox|Minefield|NetFront)"])}if(O=M=="iCab"&&parseFloat(C)>3&&"WebKit"||/\bOpera\b/.test(A)&&(/\bOPR\b/.test(e)?"Blink":"Presto")||/\b(?:Midori|Nook|Safari)\b/i.test(e)&&!/^(?:Trident|EdgeHTML)$/.test(M)&&"WebKit"||!M&&/\bMSIE\b/i.test(e)&&(D=="Mac OS"?"Tasman":"Trident")||M=="WebKit"&&/\bPlayStation\b(?! Vita\b)/i.test(A)&&"NetFront"){M=[O]}if(A=="IE"&&(O=(/; *(?:XBLWP|ZuneWP)(\d+)/i.exec(e)||0)[1])){A+=" Mobile";D="Windows Phone "+(/\+$/.test(O)?O:O+".x");S.unshift("desktop mode")}else if(/\bWPDesktop\b/i.test(e)){A="IE Mobile";D="Windows Phone 8.x";S.unshift("desktop mode");C||(C=(/\brv:([\d.]+)/.exec(e)||0)[1])}else if(A!="IE"&&M=="Trident"&&(O=/\brv:([\d.]+)/.exec(e))){if(A){S.push("identifying as "+A+(C?" "+C:""))}A="IE";C=O[1]}if(T){if(isHostType(n,"global")){if(h){O=h.lang.System;j=O.getProperty("os.arch");D=D||O.getProperty("os.name")+" "+O.getProperty("os.version")}if(g){try{C=n.require("ringo/engine").version.join(".");A="RingoJS"}catch(e){if((O=n.system)&&O.global.system==n.system){A="Narwhal";D||(D=O[0].os||null)}}if(!A){A="Rhino"}}else if(typeof n.process=="object"&&!n.process.browser&&(O=n.process)){if(typeof O.versions=="object"){if(typeof O.versions.electron=="string"){S.push("Node "+O.versions.node);A="Electron";C=O.versions.electron}else if(typeof O.versions.nw=="string"){S.push("Chromium "+C,"Node "+O.versions.node);A="NW.js";C=O.versions.nw}}if(!A){A="Node.js";j=O.arch;D=O.platform;C=/[\d.]+/.exec(O.version);C=C?C[0]:null}}}else if(getClassOf(O=n.runtime)==d){A="Adobe AIR";D=O.flash.system.Capabilities.os}else if(getClassOf(O=n.phantom)==m){A="PhantomJS";C=(O=O.version||null)&&O.major+"."+O.minor+"."+O.patch}else if(typeof w.documentMode=="number"&&(O=/\bTrident\/(\d+)/i.exec(e))){C=[C,w.documentMode];if((O=+O[1]+4)!=C[1]){S.push("IE "+C[1]+" mode");M&&(M[1]="");C[1]=O}C=A=="IE"?String(C[1].toFixed(1)):C[0]}else if(typeof w.documentMode=="number"&&/^(?:Chrome|Firefox)\b/.test(A)){S.push("masking as "+A+" "+C);A="IE";C="11.0";M=["Trident"];D="Windows"}D=D&&format(D)}if(C&&(O=/(?:[ab]|dp|pre|[ab]\d+pre)(?:\d+\+?)?$/i.exec(C)||/(?:alpha|beta)(?: ?\d)?/i.exec(e+";"+(T&&i.appMinorVersion))||/\bMinefield\b/i.test(e)&&"a")){k=/b/i.test(O)?"beta":"alpha";C=C.replace(RegExp(O+"\\+?$"),"")+(k=="beta"?x:y)+(/\d+\+?/.exec(O)||"")}if(A=="Fennec"||A=="Firefox"&&/\b(?:Android|Firefox OS|KaiOS)\b/.test(D)){A="Firefox Mobile"}else if(A=="Maxthon"&&C){C=C.replace(/\.[\d.]+/,".x")}else if(/\bXbox\b/i.test(R)){if(R=="Xbox 360"){D=null}if(R=="Xbox 360"&&/\bIEMobile\b/.test(e)){S.unshift("mobile mode")}}else if((/^(?:Chrome|IE|Opera)$/.test(A)||A&&!R&&!/Browser|Mobi/.test(A))&&(D=="Windows CE"||/Mobi/i.test(e))){A+=" Mobile"}else if(A=="IE"&&T){try{if(n.external===null){S.unshift("platform preview")}}catch(e){S.unshift("embedded")}}else if((/\bBlackBerry\b/.test(R)||/\bBB10\b/.test(e))&&(O=(RegExp(R.replace(/ +/g," *")+"/([.\\d]+)","i").exec(e)||0)[1]||C)){O=[O,/BB10/.test(e)];D=(O[1]?(R=null,N="BlackBerry"):"Device Software")+" "+O[0];C=null}else if(this!=forOwn&&R!="Wii"&&(T&&_||/Opera/.test(A)&&/\b(?:MSIE|Firefox)\b/i.test(e)||A=="Firefox"&&/\bOS X (?:\d+\.){2,}/.test(D)||A=="IE"&&(D&&!/^Win/.test(D)&&C>5.5||/\bWindows XP\b/.test(D)&&C>8||C==8&&!/\bTrident\b/.test(e)))&&!u.test(O=parse.call(forOwn,e.replace(u,"")+";"))&&O.name){O="ing as "+O.name+((O=O.version)?" "+O:"");if(u.test(A)){if(/\bIE\b/.test(O)&&D=="Mac OS"){D=null}O="identify"+O}else{O="mask"+O;if(E){A=format(E.replace(/([a-z])([A-Z])/g,"$1 $2"))}else{A="Opera"}if(/\bIE\b/.test(O)){D=null}if(!T){C=null}}M=["Presto"];S.push(O)}if(O=(/\bAppleWebKit\/([\d.]+\+?)/i.exec(e)||0)[1]){O=[parseFloat(O.replace(/\.(\d)$/,".0$1")),O];if(A=="Safari"&&O[1].slice(-1)=="+"){A="WebKit Nightly";k="alpha";C=O[1].slice(0,-1)}else if(C==O[1]||C==(O[2]=(/\bSafari\/([\d.]+\+?)/i.exec(e)||0)[1])){C=null}O[1]=(/\b(?:Headless)?Chrome\/([\d.]+)/i.exec(e)||0)[1];if(O[0]==537.36&&O[2]==537.36&&parseFloat(O[1])>=28&&M=="WebKit"){M=["Blink"]}if(!T||!s&&!O[1]){M&&(M[1]="like Safari");O=(O=O[0],O<400?1:O<500?2:O<526?3:O<533?4:O<534?"4+":O<535?5:O<537?6:O<538?7:O<601?8:O<602?9:O<604?10:O<606?11:O<608?12:"12")}else{M&&(M[1]="like Chrome");O=O[1]||(O=O[0],O<530?1:O<532?2:O<532.05?3:O<533?4:O<534.03?5:O<534.07?6:O<534.1?7:O<534.13?8:O<534.16?9:O<534.24?10:O<534.3?11:O<535.01?12:O<535.02?"13+":O<535.07?15:O<535.11?16:O<535.19?17:O<536.05?18:O<536.1?19:O<537.01?20:O<537.11?"21+":O<537.13?23:O<537.18?24:O<537.24?25:O<537.36?26:M!="Blink"?"27":"28")}M&&(M[1]+=" "+(O+=typeof O=="number"?".x":/[.+]/.test(O)?"":"+"));if(A=="Safari"&&(!C||parseInt(C)>45)){C=O}else if(A=="Chrome"&&/\bHeadlessChrome/i.test(e)){S.unshift("headless")}}if(A=="Opera"&&(O=/\bzbov|zvav$/.exec(D))){A+=" ";S.unshift("desktop mode");if(O=="zvav"){A+="Mini";C=null}else{A+="Mobile"}D=D.replace(RegExp(" *"+O+"$"),"")}else if(A=="Safari"&&/\bChrome\b/.exec(M&&M[1])){S.unshift("desktop mode");A="Chrome Mobile";C=null;if(/\bOS X\b/.test(D)){N="Apple";D="iOS 4.3+"}else{D=null}}else if(/\bSRWare Iron\b/.test(A)&&!C){C=getVersion("Chrome")}if(C&&C.indexOf(O=/[\d.]+$/.exec(D))==0&&e.indexOf("/"+O+"-")>-1){D=trim(D.replace(O,""))}if(D&&D.indexOf(A)!=-1&&!RegExp(A+" OS").test(D)){D=D.replace(RegExp(" *"+qualify(A)+" *"),"")}if(M&&!/\b(?:Avant|Nook)\b/.test(A)&&(/Browser|Lunascape|Maxthon/.test(A)||A!="Safari"&&/^iOS/.test(D)&&/\bSafari\b/.test(M[1])||/^(?:Adobe|Arora|Breach|Midori|Opera|Phantom|Rekonq|Rock|Samsung Internet|Sleipnir|SRWare Iron|Vivaldi|Web)/.test(A)&&M[1])){(O=M[M.length-1])&&S.push(O)}if(S.length){S=["("+S.join("; ")+")"]}if(N&&R&&R.indexOf(N)<0){S.push("on "+N)}if(R){S.push((/^on /.test(S[S.length-1])?"":"on ")+R)}if(D){O=/ ([\d.+]+)$/.exec(D);P=O&&D.charAt(D.length-O[0].length-1)=="/";D={architecture:32,family:O&&!P?D.replace(O[0],""):D,version:O?O[1]:null,toString:function(){var e=this.version;return this.family+(e&&!P?" "+e:"")+(this.architecture==64?" 64-bit":"")}}}if((O=/\b(?:AMD|IA|Win|WOW|x86_|x)64\b/i.exec(j))&&!/\bi686\b/i.test(j)){if(D){D.architecture=64;D.family=D.family.replace(RegExp(" *"+O),"")}if(A&&(/\bWOW64\b/i.test(e)||T&&/\w(?:86|32)$/.test(i.cpuClass||i.platform)&&!/\bWin64; x64\b/i.test(e))){S.unshift("32-bit")}}else if(D&&/^OS X/.test(D.family)&&A=="Chrome"&&parseFloat(C)>=39){D.architecture=64}e||(e=null);var z={};z.description=e;z.layout=M&&M[0];z.manufacturer=N;z.name=A;z.prerelease=k;z.product=R;z.ua=e;z.version=A&&C;z.os=D||{architecture:null,family:null,version:null,toString:function(){return"null"}};z.parse=parse;z.toString=toStringPlatform;if(z.version){S.unshift(C)}if(z.name){S.unshift(A)}if(D&&A&&!(D==String(D).split(" ")[0]&&(D==A.split(" ")[0]||R))){S.push(R?"("+D+")":"on "+D)}if(S.length){z.description=S.join(" ")}return z}var v=parse();if(i&&o){forOwn(v,(function(e,n){i[n]=e}))}else{r.platform=v}}).call(this)},922:function(e,n){"use strict";Object.defineProperty(n,"__esModule",{value:true});var t="<unknown>";function parse(e){var n=e.split("\n");return n.reduce((function(e,n){var t=parseChrome(n)||parseWinjs(n)||parseGecko(n)||parseNode(n)||parseJSC(n);if(t){e.push(t)}return e}),[])}var r=/^\s*at (.*?) ?\(((?:file|https?|blob|chrome-extension|native|eval|webpack|<anonymous>|\/|[a-z]:\\|\\\\).*?)(?::(\d+))?(?::(\d+))?\)?\s*$/i;var a=/\((\S*)(?::(\d+))(?::(\d+))\)/;function parseChrome(e){var n=r.exec(e);if(!n){return null}var i=n[2]&&n[2].indexOf("native")===0;var o=n[2]&&n[2].indexOf("eval")===0;var l=a.exec(n[2]);if(o&&l!=null){n[2]=l[1];n[3]=l[2];n[4]=l[3]}return{file:!i?n[2]:null,methodName:n[1]||t,arguments:i?[n[2]]:[],lineNumber:n[3]?+n[3]:null,column:n[4]?+n[4]:null}}var i=/^\s*at (?:((?:\[object object\])?.+) )?\(?((?:file|ms-appx|https?|webpack|blob):.*?):(\d+)(?::(\d+))?\)?\s*$/i;function parseWinjs(e){var n=i.exec(e);if(!n){return null}return{file:n[2],methodName:n[1]||t,arguments:[],lineNumber:+n[3],column:n[4]?+n[4]:null}}var o=/^\s*(.*?)(?:\((.*?)\))?(?:^|@)((?:file|https?|blob|chrome|webpack|resource|\[native).*?|[^@]*bundle)(?::(\d+))?(?::(\d+))?\s*$/i;var l=/(\S+) line (\d+)(?: > eval line \d+)* > eval/i;function parseGecko(e){var n=o.exec(e);if(!n){return null}var r=n[3]&&n[3].indexOf(" > eval")>-1;var a=l.exec(n[3]);if(r&&a!=null){n[3]=a[1];n[4]=a[2];n[5]=null}return{file:n[3],methodName:n[1]||t,arguments:n[2]?n[2].split(","):[],lineNumber:n[4]?+n[4]:null,column:n[5]?+n[5]:null}}var s=/^\s*(?:([^@]*)(?:\((.*?)\))?@)?(\S.*?):(\d+)(?::(\d+))?\s*$/i;function parseJSC(e){var n=s.exec(e);if(!n){return null}return{file:n[3],methodName:n[1]||t,arguments:[],lineNumber:+n[4],column:n[5]?+n[5]:null}}var u=/^\s*at (?:((?:\[object object\])?[^\\/]+(?: \[as \S+\])?) )?\(?(.*?):(\d+)(?::(\d+))?\)?\s*$/i;function parseNode(e){var n=u.exec(e);if(!n){return null}return{file:n[2],methodName:n[1]||t,arguments:[],lineNumber:+n[3],column:n[4]?+n[4]:null}}n.parse=parse},977:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.onRefresh=n.unregister=n.register=n.onFullRefreshNeeded=n.onBuildError=n.onBuildOk=n.ReactDevOverlay=n.getNodeError=void 0;var o=i(t(591));var l=t(525);var s=false;var u=undefined;function onUnhandledError(e){var n=e===null||e===void 0?void 0:e.error;if(!n||!(n instanceof Error)||typeof n.stack!=="string"){return}var t=n;o.emit({type:o.TYPE_UNHANDLED_ERROR,reason:n,frames:(0,l.parseStack)(t.stack)})}function onUnhandledRejection(e){var n=e===null||e===void 0?void 0:e.reason;if(!n||!(n instanceof Error)||typeof n.stack!=="string"){return}var t=n;o.emit({type:o.TYPE_UNHANDLED_REJECTION,reason:n,frames:(0,l.parseStack)(t.stack)})}function register(){if(s){return}s=true;try{var e=Error.stackTraceLimit;Error.stackTraceLimit=50;u=e}catch(e){}window.addEventListener("error",onUnhandledError);window.addEventListener("unhandledrejection",onUnhandledRejection)}n.register=register;function unregister(){if(!s){return}s=false;if(u!==undefined){try{Error.stackTraceLimit=u}catch(e){}u=undefined}window.removeEventListener("error",onUnhandledError);window.removeEventListener("unhandledrejection",onUnhandledRejection)}n.unregister=unregister;function onBuildOk(){o.emit({type:o.TYPE_BUILD_OK})}n.onBuildOk=onBuildOk;function onBuildError(e){o.emit({type:o.TYPE_BUILD_ERROR,message:e})}n.onBuildError=onBuildError;function onFullRefreshNeeded(e){o.emit({type:o.TYPE_FULL_REFRESH_NEEDED,reason:e!==null&&e!==void 0?e:null})}n.onFullRefreshNeeded=onFullRefreshNeeded;function onRefresh(){o.emit({type:o.TYPE_REFRESH})}n.onRefresh=onRefresh;var c=t(915);r(n,c,"getNodeError");var f=t(943);r(n,f,"default","ReactDevOverlay")},198:function(e,n,t){"use strict";var r=this&&this.__extends||function(){var extendStatics=function(e,n){extendStatics=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,n){e.__proto__=n}||function(e,n){for(var t in n)if(Object.prototype.hasOwnProperty.call(n,t))e[t]=n[t]};return extendStatics(e,n)};return function(e,n){if(typeof n!=="function"&&n!==null)throw new TypeError("Class extends value "+String(n)+" is not a constructor or null");extendStatics(e,n);function __(){this.constructor=e}e.prototype=n===null?Object.create(n):(__.prototype=n.prototype,new __)}}();var a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.ErrorBoundary=void 0;var i=a(t(522));var o=function(e){r(ErrorBoundary,e);function ErrorBoundary(){var n=e!==null&&e.apply(this,arguments)||this;n.state={error:null};return n}ErrorBoundary.prototype.componentDidCatch=function(e,n){this.props.onError(e,(n===null||n===void 0?void 0:n.componentStack)||null);this.setState({error:e})};ErrorBoundary.prototype.render=function(){return this.state.error?null:this.props.children};return ErrorBoundary}(i["default"].PureComponent);n.ErrorBoundary=o},943:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};var s=this&&this.__spreadArray||function(e,n,t){if(t||arguments.length===2)for(var r=0,a=n.length,i;r<a;r++){if(i||!(r in n)){if(!i)i=Array.prototype.slice.call(n,0,r);i[r]=n[r]}}return e.concat(i||Array.prototype.slice.call(n))};n.__esModule=true;var u=o(t(522));var c=o(t(591));var f=t(395);var d=t(71);var p=t(610);var v=t(710);var b=t(198);var m=t(796);var h=t(174);var g=t(803);function reducer(e,n){switch(n.type){case c.TYPE_BUILD_OK:{return r(r({},e),{buildError:null})}case c.TYPE_BUILD_ERROR:{return r(r({},e),{buildError:n.message})}case c.TYPE_FULL_REFRESH_NEEDED:{var t=r(r({},e),{fullRefreshReason:null,isAboutToFullRefresh:true});if(n.reason===null){return t}return r(r({},t),{fullRefreshReason:n.reason})}case c.TYPE_REFRESH:{return r(r({},e),{buildError:null,errors:[]})}case c.TYPE_UNHANDLED_ERROR:case c.TYPE_UNHANDLED_REJECTION:{return r(r({},e),{nextId:e.nextId+1,errors:s(s([],l(e.errors),false),[{id:e.nextId,event:n}],false)})}default:{var a=n;return e}}}var y=function ReactDevOverlay(e){var n=e.children;var t=l(u.useReducer(reducer,{nextId:1,buildError:null,errors:[],fullRefreshReason:null,isAboutToFullRefresh:false}),2),r=t[0],a=t[1];u.useEffect((function(){c.on(a);return function(){c.off(a)}}),[a]);var i=u.useCallback((function(e,n){}),[]);var o=r.buildError!=null;var s=Boolean(r.errors.length);var y=r.isAboutToFullRefresh;var x=o||s||y;return u.createElement(u.Fragment,null,u.createElement(b.ErrorBoundary,{onError:i},n!==null&&n!==void 0?n:null),x?u.createElement(f.ShadowPortal,null,u.createElement(g.CssReset,null),u.createElement(m.Base,null),u.createElement(h.ComponentStyles,null),o?u.createElement(d.BuildError,{message:r.buildError}):s?u.createElement(p.Errors,{errors:r.errors}):y?u.createElement(v.FullRefreshWarning,{reason:r.fullRefreshReason}):undefined):undefined)};n["default"]=y},591:function(e,n){"use strict";var t=this&&this.__assign||function(){t=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return t.apply(this,arguments)};n.__esModule=true;n.off=n.on=n.emit=n.TYPE_UNHANDLED_REJECTION=n.TYPE_UNHANDLED_ERROR=n.TYPE_REFRESH=n.TYPE_FULL_REFRESH_NEEDED=n.TYPE_BUILD_ERROR=n.TYPE_BUILD_OK=void 0;n.TYPE_BUILD_OK="build-ok";n.TYPE_BUILD_ERROR="build-error";n.TYPE_FULL_REFRESH_NEEDED="full-refresh-needed";n.TYPE_REFRESH="fast-refresh";n.TYPE_UNHANDLED_ERROR="unhandled-error";n.TYPE_UNHANDLED_REJECTION="unhandled-rejection";var r=new Set;var a=[];function drain(){setTimeout((function(){var _loop_1=function(){var e=a.shift();r.forEach((function(n){return n(e)}))};while(Boolean(a.length)&&Boolean(r.size)){_loop_1()}}),1)}function emit(e){a.push(Object.freeze(t({},e)));drain()}n.emit=emit;function on(e){if(r.has(e)){return false}r.add(e);drain();return true}n.on=on;function off(e){if(r.has(e)){r["delete"](e);return true}return false}n.off=off},7:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.CodeFrame=void 0;var s=l(t(194));var u=o(t(522));var c=l(t(518));var f=t(392);var d=function CodeFrame(e){var n=e.stackFrame,t=e.codeFrame;var a=u.useMemo((function(){var e=t.split(/\r?\n/g);var n=e.map((function(e){return/^>? +\d+ +\| ( *)/.exec((0,c["default"])(e))})).filter(Boolean).map((function(e){return e.pop()})).reduce((function(e,n){return isNaN(e)?n.length:Math.min(e,n.length)}),NaN);if(n>1){var r=" ".repeat(n);return e.map((function(e,n){return~(n=e.indexOf("|"))?e.substring(0,n)+e.substring(n).replace(r,""):e})).join("\n")}return e.join("\n")}),[t]);var i=u.useMemo((function(){return s["default"].ansiToJson(a,{json:true,use_classes:true,remove_empty:true})}),[a]);var o=u.useCallback((function(){var e;var t=new URLSearchParams;for(var r in n){t.append(r,((e=n[r])!==null&&e!==void 0?e:"").toString())}self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_launch-editor?"+t.toString()).then((function(){}),(function(){console.error("There was an issue opening this code in your editor.")}))}),[n]);return u.createElement("div",{"data-nextjs-codeframe":true},u.createElement("div",null,u.createElement("p",{role:"link",onClick:o,tabIndex:1,title:"Click to open in your editor"},u.createElement("span",null,(0,f.getFrameSource)(n)," @ ",n.methodName),u.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},u.createElement("path",{d:"M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"}),u.createElement("polyline",{points:"15 3 21 3 21 9"}),u.createElement("line",{x1:"10",y1:"14",x2:"21",y2:"3"})))),u.createElement("pre",null,i.map((function(e,n){return u.createElement("span",{key:"frame-"+n,style:r({color:e.fg?"var(--color-"+e.fg+")":undefined},e.decoration==="bold"?{fontWeight:800}:e.decoration==="italic"?{fontStyle:"italic"}:undefined)},e.content)}))))};n.CodeFrame=d},983:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.CodeFrame=void 0;var a=t(7);r(n,a,"CodeFrame")},761:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-codeframe] {\n    overflow: auto;\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-codeframe]::selection,\n  [data-nextjs-codeframe] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-codeframe] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n\n  [data-nextjs-codeframe] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  [data-nextjs-codeframe] > div {\n    display: inline-block;\n    width: auto;\n    min-width: 100%;\n    border-bottom: 1px solid var(--color-ansi-bright-black);\n  }\n  [data-nextjs-codeframe] > div > p {\n    display: flex;\n    align-items: center;\n    justify-content: space-between;\n    cursor: pointer;\n    margin: 0;\n  }\n  [data-nextjs-codeframe] > div > p:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-codeframe] div > p > svg {\n    width: auto;\n    height: 1em;\n    margin-left: 8px;\n  }\n  [data-nextjs-codeframe] div > pre {\n    overflow: hidden;\n    display: inline-block;\n  }\n"],["\n  [data-nextjs-codeframe] {\n    overflow: auto;\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-codeframe]::selection,\n  [data-nextjs-codeframe] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-codeframe] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n\n  [data-nextjs-codeframe] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  [data-nextjs-codeframe] > div {\n    display: inline-block;\n    width: auto;\n    min-width: 100%;\n    border-bottom: 1px solid var(--color-ansi-bright-black);\n  }\n  [data-nextjs-codeframe] > div > p {\n    display: flex;\n    align-items: center;\n    justify-content: space-between;\n    cursor: pointer;\n    margin: 0;\n  }\n  [data-nextjs-codeframe] > div > p:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-codeframe] div > p > svg {\n    width: auto;\n    height: 1em;\n    margin-left: 8px;\n  }\n  [data-nextjs-codeframe] div > pre {\n    overflow: hidden;\n    display: inline-block;\n  }\n"])));n.styles=i;var o},870:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__rest||function(e,n){var t={};for(var r in e)if(Object.prototype.hasOwnProperty.call(e,r)&&n.indexOf(r)<0)t[r]=e[r];if(e!=null&&typeof Object.getOwnPropertySymbols==="function")for(var a=0,r=Object.getOwnPropertySymbols(e);a<r.length;a++){if(n.indexOf(r[a])<0&&Object.prototype.propertyIsEnumerable.call(e,r[a]))t[r[a]]=e[r[a]]}return t};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.Dialog=void 0;var s=i(t(522));var u=t(814);var c=function Dialog(e){var n=e.children,t=e.type,r=e.onClose,a=o(e,["children","type","onClose"]);var i=l(s.useState(null),2),c=i[0],f=i[1];var d=s.useCallback((function(e){f(e)}),[]);(0,u.useOnClickOutside)(c,r);s.useEffect((function(){if(c==null){return}var e=c.getRootNode();if(!(e instanceof ShadowRoot)){return}var n=e;function handler(e){var t=n.activeElement;if(e.key==="Enter"&&t instanceof HTMLElement&&t.getAttribute("role")==="link"){e.preventDefault();e.stopPropagation();t.click()}}n.addEventListener("keydown",handler);return function(){return n.removeEventListener("keydown",handler)}}),[c]);return s.createElement("div",{ref:d,"data-nextjs-dialog":true,tabIndex:-1,role:"dialog","aria-labelledby":a["aria-labelledby"],"aria-describedby":a["aria-describedby"],"aria-modal":"true"},s.createElement("div",{"data-nextjs-dialog-banner":true,className:"banner-"+t}),n)};n.Dialog=c},244:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogBody=void 0;var o=i(t(522));var l=function DialogBody(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-body":true,className:t},n)};n.DialogBody=l},519:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogContent=void 0;var o=i(t(522));var l=function DialogContent(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-content":true,className:t},n)};n.DialogContent=l},734:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogHeader=void 0;var o=i(t(522));var l=function DialogHeader(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-header":true,className:t},n)};n.DialogHeader=l},824:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.styles=n.DialogHeader=n.DialogContent=n.DialogBody=n.Dialog=void 0;var a=t(870);r(n,a,"Dialog");var i=t(244);r(n,i,"DialogBody");var o=t(519);r(n,o,"DialogContent");var l=t(734);r(n,l,"DialogHeader");var s=t(629);r(n,s,"styles")},629:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog] {\n    display: flex;\n    flex-direction: column;\n    width: 100%;\n    margin-right: auto;\n    margin-left: auto;\n    outline: none;\n    background: white;\n    border-radius: var(--size-gap);\n    box-shadow: 0 var(--size-gap-half) var(--size-gap-double)\n      rgba(0, 0, 0, 0.25);\n    max-height: calc(100% - 56px);\n    overflow-y: hidden;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      max-height: calc(100% - 15px);\n    }\n  }\n\n  @media (min-width: 576px) {\n    [data-nextjs-dialog] {\n      max-width: 540px;\n      box-shadow: 0 var(--size-gap) var(--size-gap-quad) rgba(0, 0, 0, 0.25);\n    }\n  }\n\n  @media (min-width: 768px) {\n    [data-nextjs-dialog] {\n      max-width: 720px;\n    }\n  }\n\n  @media (min-width: 992px) {\n    [data-nextjs-dialog] {\n      max-width: 960px;\n    }\n  }\n\n  [data-nextjs-dialog-banner] {\n    position: relative;\n  }\n  [data-nextjs-dialog-banner].banner-warning {\n    border-color: var(--color-ansi-yellow);\n  }\n  [data-nextjs-dialog-banner].banner-error {\n    border-color: var(--color-ansi-red);\n  }\n\n  [data-nextjs-dialog-banner]::after {\n    z-index: 2;\n    content: '';\n    position: absolute;\n    top: 0;\n    right: 0;\n    width: 100%;\n    /* banner width: */\n    border-top-width: var(--size-gap-half);\n    border-bottom-width: 0;\n    border-top-style: solid;\n    border-bottom-style: solid;\n    border-top-color: inherit;\n    border-bottom-color: transparent;\n  }\n\n  [data-nextjs-dialog-content] {\n    overflow-y: auto;\n    border: none;\n    margin: 0;\n    /* calc(padding + banner width offset) */\n    padding: calc(var(--size-gap-double) + var(--size-gap-half))\n      var(--size-gap-double);\n    height: 100%;\n    display: flex;\n    flex-direction: column;\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-header] {\n    flex-shrink: 0;\n    margin-bottom: var(--size-gap-double);\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-body] {\n    position: relative;\n    flex: 1 1 auto;\n  }\n"],["\n  [data-nextjs-dialog] {\n    display: flex;\n    flex-direction: column;\n    width: 100%;\n    margin-right: auto;\n    margin-left: auto;\n    outline: none;\n    background: white;\n    border-radius: var(--size-gap);\n    box-shadow: 0 var(--size-gap-half) var(--size-gap-double)\n      rgba(0, 0, 0, 0.25);\n    max-height: calc(100% - 56px);\n    overflow-y: hidden;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      max-height: calc(100% - 15px);\n    }\n  }\n\n  @media (min-width: 576px) {\n    [data-nextjs-dialog] {\n      max-width: 540px;\n      box-shadow: 0 var(--size-gap) var(--size-gap-quad) rgba(0, 0, 0, 0.25);\n    }\n  }\n\n  @media (min-width: 768px) {\n    [data-nextjs-dialog] {\n      max-width: 720px;\n    }\n  }\n\n  @media (min-width: 992px) {\n    [data-nextjs-dialog] {\n      max-width: 960px;\n    }\n  }\n\n  [data-nextjs-dialog-banner] {\n    position: relative;\n  }\n  [data-nextjs-dialog-banner].banner-warning {\n    border-color: var(--color-ansi-yellow);\n  }\n  [data-nextjs-dialog-banner].banner-error {\n    border-color: var(--color-ansi-red);\n  }\n\n  [data-nextjs-dialog-banner]::after {\n    z-index: 2;\n    content: '';\n    position: absolute;\n    top: 0;\n    right: 0;\n    width: 100%;\n    /* banner width: */\n    border-top-width: var(--size-gap-half);\n    border-bottom-width: 0;\n    border-top-style: solid;\n    border-bottom-style: solid;\n    border-top-color: inherit;\n    border-bottom-color: transparent;\n  }\n\n  [data-nextjs-dialog-content] {\n    overflow-y: auto;\n    border: none;\n    margin: 0;\n    /* calc(padding + banner width offset) */\n    padding: calc(var(--size-gap-double) + var(--size-gap-half))\n      var(--size-gap-double);\n    height: 100%;\n    display: flex;\n    flex-direction: column;\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-header] {\n    flex-shrink: 0;\n    margin-bottom: var(--size-gap-double);\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-body] {\n    position: relative;\n    flex: 1 1 auto;\n  }\n"])));n.styles=i;var o},727:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.LeftRightDialogHeader=void 0;var l=i(t(522));var s=function LeftRightDialogHeader(e){var n=e.children,t=e.className,r=e.previous,a=e.next,i=e.close;var s=l.useRef(null);var u=l.useRef(null);var c=l.useRef(null);var f=o(l.useState(null),2),d=f[0],p=f[1];var v=l.useCallback((function(e){p(e)}),[]);l.useEffect((function(){if(d==null){return}var e=d.getRootNode();var n=self.document;function handler(n){if(n.key==="ArrowLeft"){n.stopPropagation();if(s.current){s.current.focus()}r&&r()}else if(n.key==="ArrowRight"){n.stopPropagation();if(u.current){u.current.focus()}a&&a()}else if(n.key==="Escape"){n.stopPropagation();if(e instanceof ShadowRoot){var t=e.activeElement;if(t&&t!==c.current&&t instanceof HTMLElement){t.blur();return}}if(i){i()}}}e.addEventListener("keydown",handler);if(e!==n){n.addEventListener("keydown",handler)}return function(){e.removeEventListener("keydown",handler);if(e!==n){n.removeEventListener("keydown",handler)}}}),[i,d,a,r]);l.useEffect((function(){if(d==null){return}var e=d.getRootNode();if(e instanceof ShadowRoot){var n=e.activeElement;if(r==null){if(s.current&&n===s.current){s.current.blur()}}else if(a==null){if(u.current&&n===u.current){u.current.blur()}}}}),[d,a,r]);return l.createElement("div",{"data-nextjs-dialog-left-right":true,className:t},l.createElement("nav",{ref:v},l.createElement("button",{ref:s,type:"button",disabled:r==null?true:undefined,"aria-disabled":r==null?true:undefined,onClick:r!==null&&r!==void 0?r:undefined},l.createElement("svg",{viewBox:"0 0 14 14",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M6.99996 1.16666L1.16663 6.99999L6.99996 12.8333M12.8333 6.99999H1.99996H12.8333Z",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"}))),l.createElement("button",{ref:u,type:"button",disabled:a==null?true:undefined,"aria-disabled":a==null?true:undefined,onClick:a!==null&&a!==void 0?a:undefined},l.createElement("svg",{viewBox:"0 0 14 14",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M6.99996 1.16666L12.8333 6.99999L6.99996 12.8333M1.16663 6.99999H12H1.16663Z",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"})))," ",n),i?l.createElement("button",{ref:c,type:"button",onClick:i,"aria-label":"Close"},l.createElement("span",{"aria-hidden":"true"},l.createElement("svg",{width:"24",height:"24",viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M18 6L6 18",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"}),l.createElement("path",{d:"M6 6L18 18",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"})))):null)};n.LeftRightDialogHeader=s},830:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.styles=n.LeftRightDialogHeader=void 0;var a=t(727);r(n,a,"LeftRightDialogHeader");var i=t(448);r(n,i,"styles")},448:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog-left-right] {\n    display: flex;\n    flex-direction: row;\n    align-content: center;\n    align-items: center;\n    justify-content: space-between;\n  }\n  [data-nextjs-dialog-left-right] > nav > button {\n    display: inline-flex;\n    align-items: center;\n    justify-content: center;\n\n    width: calc(var(--size-gap-double) + var(--size-gap));\n    height: calc(var(--size-gap-double) + var(--size-gap));\n    font-size: 0;\n    border: none;\n    background-color: rgba(255, 85, 85, 0.1);\n    color: var(--color-ansi-red);\n    cursor: pointer;\n    transition: background-color 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > nav > button > svg {\n    width: auto;\n    height: calc(var(--size-gap) + var(--size-gap-half));\n  }\n  [data-nextjs-dialog-left-right] > nav > button:hover {\n    background-color: rgba(255, 85, 85, 0.2);\n  }\n  [data-nextjs-dialog-left-right] > nav > button:disabled {\n    background-color: rgba(255, 85, 85, 0.1);\n    color: rgba(255, 85, 85, 0.4);\n    cursor: not-allowed;\n  }\n\n  [data-nextjs-dialog-left-right] > nav > button:first-of-type {\n    border-radius: var(--size-gap-half) 0 0 var(--size-gap-half);\n    margin-right: 1px;\n  }\n  [data-nextjs-dialog-left-right] > nav > button:last-of-type {\n    border-radius: 0 var(--size-gap-half) var(--size-gap-half) 0;\n  }\n\n  [data-nextjs-dialog-left-right] > button:last-of-type {\n    border: 0;\n    padding: 0;\n\n    background-color: transparent;\n    appearance: none;\n\n    opacity: 0.4;\n    transition: opacity 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > button:last-of-type:hover {\n    opacity: 0.7;\n  }\n"],["\n  [data-nextjs-dialog-left-right] {\n    display: flex;\n    flex-direction: row;\n    align-content: center;\n    align-items: center;\n    justify-content: space-between;\n  }\n  [data-nextjs-dialog-left-right] > nav > button {\n    display: inline-flex;\n    align-items: center;\n    justify-content: center;\n\n    width: calc(var(--size-gap-double) + var(--size-gap));\n    height: calc(var(--size-gap-double) + var(--size-gap));\n    font-size: 0;\n    border: none;\n    background-color: rgba(255, 85, 85, 0.1);\n    color: var(--color-ansi-red);\n    cursor: pointer;\n    transition: background-color 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > nav > button > svg {\n    width: auto;\n    height: calc(var(--size-gap) + var(--size-gap-half));\n  }\n  [data-nextjs-dialog-left-right] > nav > button:hover {\n    background-color: rgba(255, 85, 85, 0.2);\n  }\n  [data-nextjs-dialog-left-right] > nav > button:disabled {\n    background-color: rgba(255, 85, 85, 0.1);\n    color: rgba(255, 85, 85, 0.4);\n    cursor: not-allowed;\n  }\n\n  [data-nextjs-dialog-left-right] > nav > button:first-of-type {\n    border-radius: var(--size-gap-half) 0 0 var(--size-gap-half);\n    margin-right: 1px;\n  }\n  [data-nextjs-dialog-left-right] > nav > button:last-of-type {\n    border-radius: 0 var(--size-gap-half) var(--size-gap-half) 0;\n  }\n\n  [data-nextjs-dialog-left-right] > button:last-of-type {\n    border: 0;\n    padding: 0;\n\n    background-color: transparent;\n    appearance: none;\n\n    opacity: 0.4;\n    transition: opacity 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > button:last-of-type:hover {\n    opacity: 0.7;\n  }\n"])));n.styles=i;var o},434:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.Overlay=void 0;var s=l(t(426));var u=i(t(522));var c=t(253);var f=function Overlay(e){var n=e.className,t=e.children,r=e.fixed;u.useEffect((function(){(0,c.lock)();return function(){(0,c.unlock)()}}),[]);var a=o(u.useState(null),2),i=a[0],l=a[1];var f=u.useCallback((function(e){l(e)}),[]);u.useEffect((function(){if(i==null){return}var e=(0,s["default"])({context:i});return function(){e.disengage()}}),[i]);return u.createElement("div",{"data-nextjs-dialog-overlay":true,className:n,ref:f},u.createElement("div",{"data-nextjs-dialog-backdrop":true,"data-nextjs-dialog-backdrop-fixed":r?true:undefined}),t)};n.Overlay=f},253:function(e,n){"use strict";n.__esModule=true;n.unlock=n.lock=void 0;var t;var r;var a=0;function lock(){setTimeout((function(){if(a++>0){return}var e=window.innerWidth-document.documentElement.clientWidth;if(e>0){t=document.body.style.paddingRight;document.body.style.paddingRight=e+"px"}r=document.body.style.overflow;document.body.style.overflow="hidden"}))}n.lock=lock;function unlock(){setTimeout((function(){if(a===0||--a!==0){return}if(t!==undefined){document.body.style.paddingRight=t;t=undefined}if(r!==undefined){document.body.style.overflow=r;r=undefined}}))}n.unlock=unlock},382:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Overlay=void 0;var a=t(434);r(n,a,"Overlay")},426:function(e,n,t){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;var a=r(t(36));var i=r(t(730));function nodeArray(e){if(!e){return[]}if(Array.isArray(e)){return e}if(e.nodeType!==undefined){return[e]}if(typeof e==="string"){e=document.querySelectorAll(e)}if(e.length!==undefined){return[].slice.call(e,0)}throw new TypeError("unexpected input "+String(e))}function contextToElement(e){var n=e.context,t=e.label,r=t===undefined?"context-to-element":t,a=e.resolveDocument,i=e.defaultToDocument;var o=nodeArray(n)[0];if(a&&o&&o.nodeType===Node.DOCUMENT_NODE){o=o.documentElement}if(!o&&i){return document.documentElement}if(!o){throw new TypeError(r+" requires valid options.context")}if(o.nodeType!==Node.ELEMENT_NODE&&o.nodeType!==Node.DOCUMENT_FRAGMENT_NODE){throw new TypeError(r+" requires options.context to be an Element")}return o}function getShadowHost(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context;var t=contextToElement({label:"get/shadow-host",context:n});var r=null;while(t){r=t;t=t.parentNode}if(r.nodeType===r.DOCUMENT_FRAGMENT_NODE&&r.host){return r.host}return null}function getDocument(e){if(!e){return document}if(e.nodeType===Node.DOCUMENT_NODE){return e}return e.ownerDocument||document}function isActiveElement(e){var n=contextToElement({label:"is/active-element",resolveDocument:true,context:e});var t=getDocument(n);if(t.activeElement===n){return true}var r=getShadowHost({context:n});if(r&&r.shadowRoot.activeElement===n){return true}return false}function getParents(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context;var t=[];var r=contextToElement({label:"get/parents",context:n});while(r){t.push(r);r=r.parentNode;if(r&&r.nodeType!==Node.ELEMENT_NODE){r=null}}return t}var o=["matches","webkitMatchesSelector","mozMatchesSelector","msMatchesSelector"];var l=null;function findMethodName(e){o.some((function(n){if(!e[n]){return false}l=n;return true}))}function elementMatches(e,n){if(!l){findMethodName(e)}return e[l](n)}var s=JSON.parse(JSON.stringify(a["default"]));var u=s.os.family||"";var c=u==="Android";var f=u.slice(0,7)==="Windows";var d=u==="OS X";var p=u==="iOS";var v=s.layout==="Blink";var b=s.layout==="Gecko";var m=s.layout==="Trident";var h=s.layout==="EdgeHTML";var g=s.layout==="WebKit";var y=parseFloat(s.version);var x=Math.floor(y);s.majorVersion=x;s.is={ANDROID:c,WINDOWS:f,OSX:d,IOS:p,BLINK:v,GECKO:b,TRIDENT:m,EDGE:h,WEBKIT:g,IE9:m&&x===9,IE10:m&&x===10,IE11:m&&x===11};function before(){var e={activeElement:document.activeElement,windowScrollTop:window.scrollTop,windowScrollLeft:window.scrollLeft,bodyScrollTop:document.body.scrollTop,bodyScrollLeft:document.body.scrollLeft};var n=document.createElement("iframe");n.setAttribute("style","position:absolute; position:fixed; top:0; left:-2px; width:1px; height:1px; overflow:hidden;");n.setAttribute("aria-live","off");n.setAttribute("aria-busy","true");n.setAttribute("aria-hidden","true");document.body.appendChild(n);var t=n.contentWindow;var r=t.document;r.open();r.close();var a=r.createElement("div");r.body.appendChild(a);e.iframe=n;e.wrapper=a;e.window=t;e.document=r;return e}function test(e,n){e.wrapper.innerHTML="";var t=typeof n.element==="string"?e.document.createElement(n.element):n.element(e.wrapper,e.document);var r=n.mutate&&n.mutate(t,e.wrapper,e.document);if(!r&&r!==false){r=t}!t.parentNode&&e.wrapper.appendChild(t);r&&r.focus&&r.focus();return n.validate?n.validate(t,r,e.document):e.document.activeElement===r}function after(e){if(e.activeElement===document.body){document.activeElement&&document.activeElement.blur&&document.activeElement.blur();if(s.is.IE10){document.body.focus()}}else{e.activeElement&&e.activeElement.focus&&e.activeElement.focus()}document.body.removeChild(e.iframe);window.scrollTop=e.windowScrollTop;window.scrollLeft=e.windowScrollLeft;document.body.scrollTop=e.bodyScrollTop;document.body.scrollLeft=e.bodyScrollLeft}function detectFocus(e){var n=before();var t={};Object.keys(e).map((function(r){t[r]=test(n,e[r])}));after(n);return t}var w="1.4.1";function readLocalStorage(e){var n=void 0;try{n=window.localStorage&&window.localStorage.getItem(e);n=n?JSON.parse(n):{}}catch(e){n={}}return n}function writeLocalStorage(e,n){if(!document.hasFocus()){try{window.localStorage&&window.localStorage.removeItem(e)}catch(e){}return}try{window.localStorage&&window.localStorage.setItem(e,JSON.stringify(n))}catch(e){}}var _=typeof window!=="undefined"&&window.navigator.userAgent||"";var E="ally-supports-cache";var O=readLocalStorage(E);if(O.userAgent!==_||O.version!==w){O={}}O.userAgent=_;O.version=w;var j={get:function get(){return O},set:function set(e){Object.keys(e).forEach((function(n){O[n]=e[n]}));O.time=(new Date).toISOString();writeLocalStorage(E,O)}};function cssShadowPiercingDeepCombinator(){var e=void 0;try{document.querySelector("html >>> :first-child");e=">>>"}catch(n){try{document.querySelector("html /deep/ :first-child");e="/deep/"}catch(n){e=""}}return e}var S="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";var k={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test">'+'<area shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" tabindex="-1" alt="" src="'+S+'">';return e.querySelector("area")}};var T={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test">'+'<area href="#void" tabindex="-1" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" alt="" src="'+S+'">';return false},validate:function validate(e,n,t){if(s.is.GECKO){return true}var r=e.querySelector("area");r.focus();return t.activeElement===r}};var C={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-area-href-test">'+'<area shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-area-href-test" alt="" src="'+S+'">';return e.querySelector("area")},validate:function validate(e,n,t){if(s.is.GECKO){return true}return t.activeElement===n}};var P={name:"can-focus-audio-without-controls",element:"audio",mutate:function mutate(e){try{e.setAttribute("src",S)}catch(e){}}};var M="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ";var A={element:"div",mutate:function mutate(e){e.innerHTML='<map name="broken-image-map-test"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#broken-image-map-test" alt="" src="'+M+'">';return e.querySelector("area")}};var R={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","-1");e.setAttribute("style","display: -webkit-flex; display: -ms-flexbox; display: flex;");e.innerHTML='<span style="display: block;">hello</span>';return e.querySelector("span")}};var N={element:"fieldset",mutate:function mutate(e){e.setAttribute("tabindex",0);e.setAttribute("disabled","disabled")}};var D={element:"fieldset",mutate:function mutate(e){e.innerHTML="<legend>legend</legend><p>content</p>"}};var z={element:"span",mutate:function mutate(e){e.setAttribute("style","display: -webkit-flex; display: -ms-flexbox; display: flex;");e.innerHTML='<span style="display: block;">hello</span>'}};var F={element:"form",mutate:function mutate(e){e.setAttribute("tabindex",0);e.setAttribute("disabled","disabled")}};var I={element:"a",mutate:function mutate(e){e.href="#void";e.innerHTML='<img ismap src="'+S+'" alt="">';return e.querySelector("img")}};var L={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" tabindex="-1" alt="" '+'src="'+S+'">';return e.querySelector("img")}};var B={element:function element(e,n){var t=n.createElement("iframe");e.appendChild(t);var r=t.contentWindow.document;r.open();r.close();return t},mutate:function mutate(e){e.style.visibility="hidden";var n=e.contentWindow.document;var t=n.createElement("input");n.body.appendChild(t);return t},validate:function validate(e){var n=e.contentWindow.document;var t=n.querySelector("input");return n.activeElement===t}};var H=!s.is.WEBKIT;function focusInZeroDimensionObject(){return H}var W={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","invalid-value")}};var q={element:"label",mutate:function mutate(e){e.setAttribute("tabindex","-1")},validate:function validate(e,n,t){var r=e.offsetHeight;e.focus();return t.activeElement===e}};var U="data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHhtb"+"G5zOnhsaW5rPSJodHRwOi8vd3d3LnczLm9yZy8xOTk5L3hsaW5rIiBpZD0ic3ZnIj48dGV4dCB4PSIxMCIgeT0iMjAiIGlkPSJ"+"zdmctbGluay10ZXh0Ij50ZXh0PC90ZXh0Pjwvc3ZnPg==";var G={element:"object",mutate:function mutate(e){e.setAttribute("type","image/svg+xml");e.setAttribute("data",U);e.setAttribute("width","200");e.setAttribute("height","50");e.style.visibility="hidden"}};var V={name:"can-focus-object-svg",element:"object",mutate:function mutate(e){e.setAttribute("type","image/svg+xml");e.setAttribute("data",U);e.setAttribute("width","200");e.setAttribute("height","50")},validate:function validate(e,n,t){if(s.is.GECKO){return true}return t.activeElement===e}};var K=!s.is.IE9;function focusObjectSwf(){return K}var Z={element:"div",mutate:function mutate(e){e.innerHTML='<map name="focus-redirect-img-usemap"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#focus-redirect-img-usemap" alt="" '+'src="'+S+'">';return e.querySelector("img")},validate:function validate(e,n,t){var r=e.querySelector("area");return t.activeElement===r}};var $={element:"fieldset",mutate:function mutate(e){e.innerHTML='<legend>legend</legend><input tabindex="-1"><input tabindex="0">';return false},validate:function validate(e,n,t){var r=e.querySelector('input[tabindex="-1"]');var a=e.querySelector('input[tabindex="0"]');e.focus();e.querySelector("legend").focus();return t.activeElement===r&&"focusable"||t.activeElement===a&&"tabbable"||""}};var Y={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px; overflow: auto;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>';return e.querySelector("div")}};var X={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>'}};var J={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px; overflow: auto;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>'}};var Q={element:"details",mutate:function mutate(e){e.innerHTML="<summary>foo</summary><p>content</p>";return e.firstElementChild}};function makeFocusableForeignObject(){var e=document.createElement("div");e.innerHTML='<svg><foreignObject width="30" height="30">\n      <input type="text"/>\n  </foreignObject></svg>';return e.firstChild.firstChild}function focusSvgForeignObjectHack(e){var n=e.ownerSVGElement||e.nodeName.toLowerCase()==="svg";if(!n){return false}var t=makeFocusableForeignObject();e.appendChild(t);var r=t.querySelector("input");r.focus();r.disabled=true;e.removeChild(t);return true}function generate(e){return'<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">'+e+"</svg>"}function focus(e){if(e.focus){return}try{HTMLElement.prototype.focus.call(e)}catch(n){focusSvgForeignObjectHack(e)}}function validate(e,n,t){focus(n);return t.activeElement===n}var ee={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text focusable="true">a</text>');return e.querySelector("text")},validate:validate};var ne={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text tabindex="0">a</text>');return e.querySelector("text")},validate:validate};var te={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text tabindex="-1">a</text>');return e.querySelector("text")},validate:validate};var re={element:"div",mutate:function mutate(e){e.innerHTML=generate(['<g id="ally-test-target"><a xlink:href="#void"><text>link</text></a></g>','<use xlink:href="#ally-test-target" x="0" y="0" tabindex="-1" />'].join(""));return e.querySelector("use")},validate:validate};var ae={element:"div",mutate:function mutate(e){e.innerHTML=generate('<foreignObject tabindex="-1"><input type="text" /></foreignObject>');return e.querySelector("foreignObject")||e.getElementsByTagName("foreignObject")[0]},validate:validate};var ie=Boolean(s.is.GECKO&&typeof SVGElement!=="undefined"&&SVGElement.prototype.focus);function focusSvgInIframe(){return ie}var oe={element:"div",mutate:function mutate(e){e.innerHTML=generate("");return e.firstChild},validate:validate};var le={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","3x")}};var se={element:"table",mutate:function mutate(e,n,t){var r=t.createDocumentFragment();r.innerHTML="<tr><td>cell</td></tr>";e.appendChild(r)}};var ue={element:"video",mutate:function mutate(e){try{e.setAttribute("src",S)}catch(e){}}};var ce=s.is.GECKO||s.is.TRIDENT||s.is.EDGE;function tabsequenceAreaAtImgPosition(){return ce}var fe={cssShadowPiercingDeepCombinator:cssShadowPiercingDeepCombinator,focusInZeroDimensionObject:focusInZeroDimensionObject,focusObjectSwf:focusObjectSwf,focusSvgInIframe:focusSvgInIframe,tabsequenceAreaAtImgPosition:tabsequenceAreaAtImgPosition};var de={focusAreaImgTabindex:k,focusAreaTabindex:T,focusAreaWithoutHref:C,focusAudioWithoutControls:P,focusBrokenImageMap:A,focusChildrenOfFocusableFlexbox:R,focusFieldsetDisabled:N,focusFieldset:D,focusFlexboxContainer:z,focusFormDisabled:F,focusImgIsmap:I,focusImgUsemapTabindex:L,focusInHiddenIframe:B,focusInvalidTabindex:W,focusLabelTabindex:q,focusObjectSvg:V,focusObjectSvgHidden:G,focusRedirectImgUsemap:Z,focusRedirectLegend:$,focusScrollBody:Y,focusScrollContainerWithoutOverflow:X,focusScrollContainer:J,focusSummary:Q,focusSvgFocusableAttribute:ee,focusSvgTabindexAttribute:ne,focusSvgNegativeTabindexAttribute:te,focusSvgUseTabindex:re,focusSvgForeignobjectTabindex:ae,focusSvg:oe,focusTabindexTrailingCharacters:le,focusTable:se,focusVideoWithoutControls:ue};function executeTests(){var e=detectFocus(de);Object.keys(fe).forEach((function(n){e[n]=fe[n]()}));return e}var pe=null;function _supports(){if(pe){return pe}pe=j.get();if(!pe.time){j.set(executeTests());pe=j.get()}return pe}var ve=void 0;var be=/^\s*(-|\+)?[0-9]+\s*$/;var me=/^\s*(-|\+)?[0-9]+.*$/;function isValidTabindex(e){if(!ve){ve=_supports()}var n=ve.focusTabindexTrailingCharacters?me:be;var t=contextToElement({label:"is/valid-tabindex",resolveDocument:true,context:e});var r=t.hasAttribute("tabindex");var a=t.hasAttribute("tabIndex");if(!r&&!a){return false}var i=t.ownerSVGElement||t.nodeName.toLowerCase()==="svg";if(i&&!ve.focusSvgTabindexAttribute){return false}if(ve.focusInvalidTabindex){return true}var o=t.getAttribute(r?"tabindex":"tabIndex");if(o==="-32768"){return false}return Boolean(o&&n.test(o))}function tabindexValue(e){if(!isValidTabindex(e)){return null}var n=e.hasAttribute("tabindex");var t=n?"tabindex":"tabIndex";var r=parseInt(e.getAttribute(t),10);return isNaN(r)?-1:r}function isUserModifyWritable(e){var n=e.webkitUserModify||"";return Boolean(n&&n.indexOf("write")!==-1)}function hasCssOverflowScroll(e){return[e.getPropertyValue("overflow"),e.getPropertyValue("overflow-x"),e.getPropertyValue("overflow-y")].some((function(e){return e==="auto"||e==="scroll"}))}function hasCssDisplayFlex(e){return e.display.indexOf("flex")>-1}function isScrollableContainer(e,n,t,r){if(n!=="div"&&n!=="span"){return false}if(t&&t!=="div"&&t!=="span"&&!hasCssOverflowScroll(r)){return false}return e.offsetHeight<e.scrollHeight||e.offsetWidth<e.scrollWidth}var he=void 0;function isFocusRelevantRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{flexbox:false,scrollable:false,shadow:false}:t;if(!he){he=_supports()}var a=contextToElement({label:"is/focus-relevant",resolveDocument:true,context:n});if(!r.shadow&&a.shadowRoot){return true}var i=a.nodeName.toLowerCase();if(i==="input"&&a.type==="hidden"){return false}if(i==="input"||i==="select"||i==="button"||i==="textarea"){return true}if(i==="legend"&&he.focusRedirectLegend){return true}if(i==="label"){return true}if(i==="area"){return true}if(i==="a"&&a.hasAttribute("href")){return true}if(i==="object"&&a.hasAttribute("usemap")){return false}if(i==="object"){var o=a.getAttribute("type");if(!he.focusObjectSvg&&o==="image/svg+xml"){return false}else if(!he.focusObjectSwf&&o==="application/x-shockwave-flash"){return false}}if(i==="iframe"||i==="object"){return true}if(i==="embed"||i==="keygen"){return true}if(a.hasAttribute("contenteditable")){return true}if(i==="audio"&&(he.focusAudioWithoutControls||a.hasAttribute("controls"))){return true}if(i==="video"&&(he.focusVideoWithoutControls||a.hasAttribute("controls"))){return true}if(he.focusSummary&&i==="summary"){return true}var l=isValidTabindex(a);if(i==="img"&&a.hasAttribute("usemap")){return l&&he.focusImgUsemapTabindex||he.focusRedirectImgUsemap}if(he.focusTable&&(i==="table"||i==="td")){return true}if(he.focusFieldset&&i==="fieldset"){return true}var s=i==="svg";var u=a.ownerSVGElement;var c=a.getAttribute("focusable");var f=tabindexValue(a);if(i==="use"&&f!==null&&!he.focusSvgUseTabindex){return false}if(i==="foreignobject"){return f!==null&&he.focusSvgForeignobjectTabindex}if(elementMatches(a,"svg a")&&a.hasAttribute("xlink:href")){return true}if((s||u)&&a.focus&&!he.focusSvgNegativeTabindexAttribute&&f<0){return false}if(s){return l||he.focusSvg||he.focusSvgInIframe||Boolean(he.focusSvgFocusableAttribute&&c&&c==="true")}if(u){if(he.focusSvgTabindexAttribute&&l){return true}if(he.focusSvgFocusableAttribute){return c==="true"}}if(l){return true}var d=window.getComputedStyle(a,null);if(isUserModifyWritable(d)){return true}if(he.focusImgIsmap&&i==="img"&&a.hasAttribute("ismap")){var p=getParents({context:a}).some((function(e){return e.nodeName.toLowerCase()==="a"&&e.hasAttribute("href")}));if(p){return true}}if(!r.scrollable&&he.focusScrollContainer){if(he.focusScrollContainerWithoutOverflow){if(isScrollableContainer(a,i)){return true}}else if(hasCssOverflowScroll(d)){return true}}if(!r.flexbox&&he.focusFlexboxContainer&&hasCssDisplayFlex(d)){return true}var v=a.parentElement;if(!r.scrollable&&v){var b=v.nodeName.toLowerCase();var m=window.getComputedStyle(v,null);if(he.focusScrollBody&&isScrollableContainer(v,i,b,m)){return true}if(he.focusChildrenOfFocusableFlexbox){if(hasCssDisplayFlex(m)){return true}}}return false}isFocusRelevantRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isFocusRelevant(n){return isFocusRelevantRules({context:n,except:e})};n.rules=isFocusRelevantRules;return n};var ge=isFocusRelevantRules.except({});function findIndex(e,n){if(e.findIndex){return e.findIndex(n)}var t=e.length;if(t===0){return-1}for(var r=0;r<t;r++){if(n(e[r],r,e)){return r}}return-1}function getContentDocument(e){try{return e.contentDocument||e.contentWindow&&e.contentWindow.document||e.getSVGDocument&&e.getSVGDocument()||null}catch(e){return null}}function getWindow(e){var n=getDocument(e);return n.defaultView||window}var ye=void 0;function selectInShadows(e){if(typeof ye!=="string"){var n=cssShadowPiercingDeepCombinator();if(n){ye=", html "+n+" "}}if(!ye){return e}return e+ye+e.replace(/\s*,\s*/g,",").split(",").join(ye)}var xe=void 0;function findDocumentHostElement(e){if(!xe){xe=selectInShadows("object, iframe")}if(e._frameElement!==undefined){return e._frameElement}e._frameElement=null;var n=e.parent.document.querySelectorAll(xe);[].some.call(n,(function(n){var t=getContentDocument(n);if(t!==e.document){return false}e._frameElement=n;return true}));return e._frameElement}function getFrameElement(e){var n=getWindow(e);if(!n.parent||n.parent===n){return null}try{return n.frameElement||findDocumentHostElement(n)}catch(e){return null}}var we=/^(area)$/;function computedStyle(e,n){return window.getComputedStyle(e,null).getPropertyValue(n)}function notDisplayed(e){return e.some((function(e){return computedStyle(e,"display")==="none"}))}function notVisible(e){var n=findIndex(e,(function(e){var n=computedStyle(e,"visibility");return n==="hidden"||n==="collapse"}));if(n===-1){return false}var t=findIndex(e,(function(e){return computedStyle(e,"visibility")==="visible"}));if(t===-1){return true}if(n<t){return true}return false}function collapsedParent(e){var n=1;if(e[0].nodeName.toLowerCase()==="summary"){n=2}return e.slice(n).some((function(e){return e.nodeName.toLowerCase()==="details"&&e.open===false}))}function isVisibleRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{notRendered:false,cssDisplay:false,cssVisibility:false,detailsElement:false,browsingContext:false}:t;var a=contextToElement({label:"is/visible",resolveDocument:true,context:n});var i=a.nodeName.toLowerCase();if(!r.notRendered&&we.test(i)){return true}var o=getParents({context:a});var l=i==="audio"&&!a.hasAttribute("controls");if(!r.cssDisplay&&notDisplayed(l?o.slice(1):o)){return false}if(!r.cssVisibility&&notVisible(o)){return false}if(!r.detailsElement&&collapsedParent(o)){return false}if(!r.browsingContext){var s=getFrameElement(a);var u=isVisibleRules.except(r);if(s&&!u(s)){return false}}return true}isVisibleRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isVisible(n){return isVisibleRules({context:n,except:e})};n.rules=isVisibleRules;return n};var _e=isVisibleRules.except({});function getMapByName(e,n){var t=n.querySelector('map[name="'+(0,i["default"])(e)+'"]');return t||null}function getImageOfArea(e){var n=e.parentElement;if(!n.name||n.nodeName.toLowerCase()!=="map"){return null}var t=getDocument(e);return t.querySelector('img[usemap="#'+(0,i["default"])(n.name)+'"]')||null}var Ee=void 0;function isValidArea(e){if(!Ee){Ee=_supports()}var n=contextToElement({label:"is/valid-area",context:e});var t=n.nodeName.toLowerCase();if(t!=="area"){return false}var r=n.hasAttribute("tabindex");if(!Ee.focusAreaTabindex&&r){return false}var a=getImageOfArea(n);if(!a||!_e(a)){return false}if(!Ee.focusBrokenImageMap&&(!a.complete||!a.naturalHeight||a.offsetWidth<=0||a.offsetHeight<=0)){return false}if(!Ee.focusAreaWithoutHref&&!n.href){return Ee.focusAreaTabindex&&r||Ee.focusAreaImgTabindex&&a.hasAttribute("tabindex")}var i=getParents({context:a}).slice(1).some((function(e){var n=e.nodeName.toLowerCase();return n==="button"||n==="a"}));if(i){return false}return true}var Oe=void 0;var je=void 0;var Se={input:true,select:true,textarea:true,button:true,fieldset:true,form:true};function isNativeDisabledSupported(e){if(!Oe){Oe=_supports();if(Oe.focusFieldsetDisabled){delete Se.fieldset}if(Oe.focusFormDisabled){delete Se.form}je=new RegExp("^("+Object.keys(Se).join("|")+")$")}var n=contextToElement({label:"is/native-disabled-supported",context:e});var t=n.nodeName.toLowerCase();return Boolean(je.test(t))}var ke=void 0;function isDisabledFieldset(e){var n=e.nodeName.toLowerCase();return n==="fieldset"&&e.disabled}function isDisabledForm(e){var n=e.nodeName.toLowerCase();return n==="form"&&e.disabled}function isDisabled(e){if(!ke){ke=_supports()}var n=contextToElement({label:"is/disabled",context:e});if(n.hasAttribute("data-ally-disabled")){return true}if(!isNativeDisabledSupported(n)){return false}if(n.disabled){return true}var t=getParents({context:n});if(t.some(isDisabledFieldset)){return true}if(!ke.focusFormDisabled&&t.some(isDisabledForm)){return true}return false}function isOnlyTabbableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{onlyFocusableBrowsingContext:false,visible:false}:t;var a=contextToElement({label:"is/only-tabbable",resolveDocument:true,context:n});if(!r.visible&&!_e(a)){return false}if(!r.onlyFocusableBrowsingContext&&(s.is.GECKO||s.is.TRIDENT||s.is.EDGE)){var i=getFrameElement(a);if(i){if(tabindexValue(i)<0){return false}}}var o=a.nodeName.toLowerCase();var l=tabindexValue(a);if(o==="label"&&s.is.GECKO){return l!==null&&l>=0}if(s.is.GECKO&&a.ownerSVGElement&&!a.focus){if(o==="a"&&a.hasAttribute("xlink:href")){if(s.is.GECKO){return true}}}return false}isOnlyTabbableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isOnlyTabbable(n){return isOnlyTabbableRules({context:n,except:e})};n.rules=isOnlyTabbableRules;return n};var Te=isOnlyTabbableRules.except({});var Ce=void 0;function isOnlyFocusRelevant(e){var n=e.nodeName.toLowerCase();if(n==="embed"||n==="keygen"){return true}var t=tabindexValue(e);if(e.shadowRoot&&t===null){return true}if(n==="label"){return!Ce.focusLabelTabindex||t===null}if(n==="legend"){return t===null}if(Ce.focusSvgFocusableAttribute&&(e.ownerSVGElement||n==="svg")){var r=e.getAttribute("focusable");return r&&r==="false"}if(n==="img"&&e.hasAttribute("usemap")){return t===null||!Ce.focusImgUsemapTabindex}if(n==="area"){return!isValidArea(e)}return false}function isFocusableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{disabled:false,visible:false,onlyTabbable:false}:t;if(!Ce){Ce=_supports()}var a=Te.rules.except({onlyFocusableBrowsingContext:true,visible:r.visible});var i=contextToElement({label:"is/focusable",resolveDocument:true,context:n});var o=ge.rules({context:i,except:r});if(!o||isOnlyFocusRelevant(i)){return false}if(!r.disabled&&isDisabled(i)){return false}if(!r.onlyTabbable&&a(i)){return false}if(!r.visible){var l={context:i,except:{}};if(Ce.focusInHiddenIframe){l.except.browsingContext=true}if(Ce.focusObjectSvgHidden){var s=i.nodeName.toLowerCase();if(s==="object"){l.except.cssVisibility=true}}if(!_e.rules(l)){return false}}var u=getFrameElement(i);if(u){var c=u.nodeName.toLowerCase();if(c==="object"&&!Ce.focusInZeroDimensionObject){if(!u.offsetWidth||!u.offsetHeight){return false}}}var f=i.nodeName.toLowerCase();if(f==="svg"&&Ce.focusSvgInIframe&&!u&&i.getAttribute("tabindex")===null){return false}return true}isFocusableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isFocusable(n){return isFocusableRules({context:n,except:e})};n.rules=isFocusableRules;return n};var Pe=isFocusableRules.except({});function createFilter(e){var n=function filter(n){if(n.shadowRoot){return NodeFilter.FILTER_ACCEPT}if(e(n)){return NodeFilter.FILTER_ACCEPT}return NodeFilter.FILTER_SKIP};n.acceptNode=n;return n}var Me=createFilter(ge);function queryFocusableStrict(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;if(!n){n=document.documentElement}var i=Pe.rules.except({onlyTabbable:r});var o=getDocument(n);var l=o.createTreeWalker(n,NodeFilter.SHOW_ELEMENT,a==="all"?Me:createFilter(i),false);var s=[];while(l.nextNode()){if(l.currentNode.shadowRoot){if(i(l.currentNode)){s.push(l.currentNode)}s=s.concat(queryFocusableStrict({context:l.currentNode.shadowRoot,includeOnlyTabbable:r,strategy:a}))}else{s.push(l.currentNode)}}if(t){if(a==="all"){if(ge(n)){s.unshift(n)}}else if(i(n)){s.unshift(n)}}return s}var Ae=void 0;var Re=void 0;function selector$2(){if(!Ae){Ae=_supports()}if(typeof Re==="string"){return Re}Re=""+(Ae.focusTable?"table, td,":"")+(Ae.focusFieldset?"fieldset,":"")+"svg a,"+"a[href],"+"area[href],"+"input, select, textarea, button,"+"iframe, object, embed,"+"keygen,"+(Ae.focusAudioWithoutControls?"audio,":"audio[controls],")+(Ae.focusVideoWithoutControls?"video,":"video[controls],")+(Ae.focusSummary?"summary,":"")+"[tabindex],"+"[contenteditable]";Re=selectInShadows(Re);return Re}function queryFocusableQuick(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable;var a=selector$2();var i=n.querySelectorAll(a);var o=Pe.rules.except({onlyTabbable:r});var l=[].filter.call(i,o);if(t&&o(n)){l.unshift(n)}return l}function queryFocusable(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy,i=a===undefined?"quick":a;var o=contextToElement({label:"query/focusable",resolveDocument:true,defaultToDocument:true,context:n});var l={context:o,includeContext:t,includeOnlyTabbable:r,strategy:i};if(i==="quick"){return queryFocusableQuick(l)}else if(i==="strict"||i==="all"){return queryFocusableStrict(l)}throw new TypeError('query/focusable requires option.strategy to be one of ["quick", "strict", "all"]')}var Ne=void 0;var De=/^(fieldset|table|td|body)$/;function isTabbableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{flexbox:false,scrollable:false,shadow:false,visible:false,onlyTabbable:false}:t;if(!Ne){Ne=_supports()}var a=contextToElement({label:"is/tabbable",resolveDocument:true,context:n});if(s.is.BLINK&&s.is.ANDROID&&s.majorVersion>42){return false}var i=getFrameElement(a);if(i){if(s.is.WEBKIT&&s.is.IOS){return false}if(tabindexValue(i)<0){return false}if(!r.visible&&(s.is.BLINK||s.is.WEBKIT)&&!_e(i)){return false}var o=i.nodeName.toLowerCase();if(o==="object"){var l=s.name==="Chrome"&&s.majorVersion>=54||s.name==="Opera"&&s.majorVersion>=41;if(s.is.WEBKIT||s.is.BLINK&&!l){return false}}}var u=a.nodeName.toLowerCase();var c=tabindexValue(a);var f=c===null?null:c>=0;if(s.is.EDGE&&s.majorVersion>=14&&i&&a.ownerSVGElement&&c<0){return true}var d=f!==false;var p=c!==null&&c>=0;if(a.hasAttribute("contenteditable")){return d}if(De.test(u)&&f!==true){return false}if(s.is.WEBKIT&&s.is.IOS){var v=u==="input"&&a.type==="text"||a.type==="password"||u==="select"||u==="textarea"||a.hasAttribute("contenteditable");if(!v){var b=window.getComputedStyle(a,null);v=isUserModifyWritable(b)}if(!v){return false}}if(u==="use"&&c!==null){if(s.is.BLINK||s.is.WEBKIT&&s.majorVersion===9){return true}}if(elementMatches(a,"svg a")&&a.hasAttribute("xlink:href")){if(d){return true}if(a.focus&&!Ne.focusSvgNegativeTabindexAttribute){return true}}if(u==="svg"&&Ne.focusSvgInIframe&&d){return true}if(s.is.TRIDENT||s.is.EDGE){if(u==="svg"){if(Ne.focusSvg){return true}return a.hasAttribute("focusable")||p}if(a.ownerSVGElement){if(Ne.focusSvgTabindexAttribute&&p){return true}return a.hasAttribute("focusable")}}if(a.tabIndex===undefined){return Boolean(r.onlyTabbable)}if(u==="audio"){if(!a.hasAttribute("controls")){return false}else if(s.is.BLINK){return true}}if(u==="video"){if(!a.hasAttribute("controls")){if(s.is.TRIDENT||s.is.EDGE){return false}}else if(s.is.BLINK||s.is.GECKO){return true}}if(u==="object"){if(s.is.BLINK||s.is.WEBKIT){return false}}if(u==="iframe"){return false}if(!r.scrollable&&s.is.GECKO){var m=window.getComputedStyle(a,null);if(hasCssOverflowScroll(m)){return d}}if(s.is.TRIDENT||s.is.EDGE){if(u==="area"){var h=getImageOfArea(a);if(h&&tabindexValue(h)<0){return false}}var g=window.getComputedStyle(a,null);if(isUserModifyWritable(g)){return a.tabIndex>=0}if(!r.flexbox&&hasCssDisplayFlex(g)){if(c!==null){return p}return ze(a)&&Fe(a)}if(isScrollableContainer(a,u)){return false}var y=a.parentElement;if(y){var x=y.nodeName.toLowerCase();var w=window.getComputedStyle(y,null);if(isScrollableContainer(y,u,x,w)){return false}if(hasCssDisplayFlex(w)){return p}}}return a.tabIndex>=0}isTabbableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isTabbable(n){return isTabbableRules({context:n,except:e})};n.rules=isTabbableRules;return n};var ze=ge.rules.except({flexbox:true});var Fe=isTabbableRules.except({flexbox:true});var Ie=isTabbableRules.except({});function queryTabbable(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;var i=Ie.rules.except({onlyTabbable:r});return queryFocusable({context:n,includeContext:t,includeOnlyTabbable:r,strategy:a}).filter(i)}function compareDomPosition(e,n){return e.compareDocumentPosition(n)&Node.DOCUMENT_POSITION_FOLLOWING?-1:1}function sortDomOrder(e){return e.sort(compareDomPosition)}function getFirstSuccessorOffset(e,n){return findIndex(e,(function(e){return n.compareDocumentPosition(e)&Node.DOCUMENT_POSITION_FOLLOWING}))}function findInsertionOffsets(e,n,t){var r=[];n.forEach((function(n){var a=true;var i=e.indexOf(n);if(i===-1){i=getFirstSuccessorOffset(e,n);a=false}if(i===-1){i=e.length}var o=nodeArray(t?t(n):n);if(!o.length){return}r.push({offset:i,replace:a,elements:o})}));return r}function insertElementsAtOffsets(e,n){var t=0;n.sort((function(e,n){return e.offset-n.offset}));n.forEach((function(n){var r=n.replace?1:0;var a=[n.offset+t,r].concat(n.elements);e.splice.apply(e,a);t+=n.elements.length-r}))}function mergeInDomOrder(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.list,t=e.elements,r=e.resolveElement;var a=n.slice(0);var i=nodeArray(t).slice(0);sortDomOrder(i);var o=findInsertionOffsets(a,i,r);insertElementsAtOffsets(a,o);return a}var Le=function(){function defineProperties(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||false;r.configurable=true;if("value"in r)r.writable=true;Object.defineProperty(e,r.key,r)}}return function(e,n,t){if(n)defineProperties(e.prototype,n);if(t)defineProperties(e,t);return e}}();function _classCallCheck(e,n){if(!(e instanceof n)){throw new TypeError("Cannot call a class as a function")}}var Be=function(){function Maps(e){_classCallCheck(this,Maps);this._document=getDocument(e);this.maps={}}Le(Maps,[{key:"getAreasFor",value:function getAreasFor(e){if(!this.maps[e]){this.addMapByName(e)}return this.maps[e]}},{key:"addMapByName",value:function addMapByName(e){var n=getMapByName(e,this._document);if(!n){return}this.maps[n.name]=queryTabbable({context:n})}},{key:"extractAreasFromList",value:function extractAreasFromList(e){return e.filter((function(e){var n=e.nodeName.toLowerCase();if(n!=="area"){return true}var t=e.parentNode;if(!this.maps[t.name]){this.maps[t.name]=[]}this.maps[t.name].push(e);return false}),this)}}]);return Maps}();function sortArea(e,n){var t=n.querySelectorAll("img[usemap]");var r=new Be(n);var a=r.extractAreasFromList(e);if(!t.length){return a}return mergeInDomOrder({list:a,elements:t,resolveElement:function resolveElement(e){var n=e.getAttribute("usemap").slice(1);return r.getAreasFor(n)}})}var He=function(){function defineProperties(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||false;r.configurable=true;if("value"in r)r.writable=true;Object.defineProperty(e,r.key,r)}}return function(e,n,t){if(n)defineProperties(e.prototype,n);if(t)defineProperties(e,t);return e}}();function _classCallCheck$1(e,n){if(!(e instanceof n)){throw new TypeError("Cannot call a class as a function")}}var We=function(){function Shadows(e,n){_classCallCheck$1(this,Shadows);this.context=e;this.sortElements=n;this.hostCounter=1;this.inHost={};this.inDocument=[];this.hosts={};this.elements={}}He(Shadows,[{key:"_registerHost",value:function _registerHost(e){if(e._sortingId){return}e._sortingId="shadow-"+this.hostCounter++;this.hosts[e._sortingId]=e;var n=getShadowHost({context:e});if(n){this._registerHost(n);this._registerHostParent(e,n)}else{this.inDocument.push(e)}}},{key:"_registerHostParent",value:function _registerHostParent(e,n){if(!this.inHost[n._sortingId]){this.inHost[n._sortingId]=[]}this.inHost[n._sortingId].push(e)}},{key:"_registerElement",value:function _registerElement(e,n){if(!this.elements[n._sortingId]){this.elements[n._sortingId]=[]}this.elements[n._sortingId].push(e)}},{key:"extractElements",value:function extractElements(e){return e.filter((function(e){var n=getShadowHost({context:e});if(!n){return true}this._registerHost(n);this._registerElement(e,n);return false}),this)}},{key:"sort",value:function sort(e){var n=this._injectHosts(e);n=this._replaceHosts(n);this._cleanup();return n}},{key:"_injectHosts",value:function _injectHosts(e){Object.keys(this.hosts).forEach((function(e){var n=this.elements[e];var t=this.inHost[e];var r=this.hosts[e].shadowRoot;this.elements[e]=this._merge(n,t,r)}),this);return this._merge(e,this.inDocument,this.context)}},{key:"_merge",value:function _merge(e,n,t){var r=mergeInDomOrder({list:e,elements:n});return this.sortElements(r,t)}},{key:"_replaceHosts",value:function _replaceHosts(e){return mergeInDomOrder({list:e,elements:this.inDocument,resolveElement:this._resolveHostElement.bind(this)})}},{key:"_resolveHostElement",value:function _resolveHostElement(e){var n=mergeInDomOrder({list:this.elements[e._sortingId],elements:this.inHost[e._sortingId],resolveElement:this._resolveHostElement.bind(this)});var t=tabindexValue(e);if(t!==null&&t>-1){return[e].concat(n)}return n}},{key:"_cleanup",value:function _cleanup(){Object.keys(this.hosts).forEach((function(e){delete this.hosts[e]._sortingId}),this)}}]);return Shadows}();function sortShadowed(e,n,t){var r=new We(n,t);var a=r.extractElements(e);if(a.length===e.length){return t(e)}return r.sort(a)}function sortTabindex(e){var n={};var t=[];var r=e.filter((function(e){var r=e.tabIndex;if(r===undefined){r=tabindexValue(e)}if(r<=0||r===null||r===undefined){return true}if(!n[r]){n[r]=[];t.push(r)}n[r].push(e);return false}));var a=t.sort().map((function(e){return n[e]})).reduceRight((function(e,n){return n.concat(e)}),r);return a}var qe=void 0;function moveContextToBeginning(e,n){var t=e.indexOf(n);if(t>0){var r=e.splice(t,1);return r.concat(e)}return e}function sortElements(e,n){if(qe.tabsequenceAreaAtImgPosition){e=sortArea(e,n)}e=sortTabindex(e);return e}function queryTabsequence(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;if(!qe){qe=_supports()}var i=nodeArray(n)[0]||document.documentElement;var o=queryTabbable({context:i,includeContext:t,includeOnlyTabbable:r,strategy:a});if(document.body.createShadowRoot&&s.is.BLINK){o=sortShadowed(o,i,sortElements)}else{o=sortElements(o,i)}if(t){o=moveContextToBeginning(o,i)}return o}var Ue={tab:9,left:37,up:38,right:39,down:40,pageUp:33,"page-up":33,pageDown:34,"page-down":34,end:35,home:36,enter:13,escape:27,space:32,shift:16,capsLock:20,"caps-lock":20,ctrl:17,alt:18,meta:91,pause:19,insert:45,delete:46,backspace:8,_alias:{91:[92,93,224]}};for(var Ge=1;Ge<26;Ge++){Ue["f"+Ge]=Ge+111}for(var Ve=0;Ve<10;Ve++){var Ke=Ve+48;var Ze=Ve+96;Ue[Ve]=Ke;Ue["num-"+Ve]=Ze;Ue._alias[Ke]=[Ze]}for(var $e=0;$e<26;$e++){var Ye=$e+65;var Xe=String.fromCharCode(Ye).toLowerCase();Ue[Xe]=Ye}var Je={alt:"altKey",ctrl:"ctrlKey",meta:"metaKey",shift:"shiftKey"};var Qe=Object.keys(Je).map((function(e){return Je[e]}));function createExpectedModifiers(e){var n=e?null:false;return{altKey:n,ctrlKey:n,metaKey:n,shiftKey:n}}function resolveModifiers(e){var n=e.indexOf("*")!==-1;var t=createExpectedModifiers(n);e.forEach((function(e){if(e==="*"){return}var n=true;var r=e.slice(0,1);if(r==="?"){n=null}else if(r==="!"){n=false}if(n!==true){e=e.slice(1)}var a=Je[e];if(!a){throw new TypeError('Unknown modifier "'+e+'"')}t[a]=n}));return t}function resolveKey(e){var n=Ue[e]||parseInt(e,10);if(!n||typeof n!=="number"||isNaN(n)){throw new TypeError('Unknown key "'+e+'"')}return[n].concat(Ue._alias[n]||[])}function matchModifiers(e,n){return!Qe.some((function(t){return typeof e[t]==="boolean"&&Boolean(n[t])!==e[t]}))}function keyBinding(e){return e.split(/\s+/).map((function(e){var n=e.split("+");var t=resolveModifiers(n.slice(0,-1));var r=resolveKey(n.slice(-1));return{keyCodes:r,modifiers:t,matchModifiers:matchModifiers.bind(null,t)}}))}function getParentComparator(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.parent,t=e.element,r=e.includeSelf;if(n){return function isChildOf(e){return Boolean(r&&e===n||n.compareDocumentPosition(e)&Node.DOCUMENT_POSITION_CONTAINED_BY)}}else if(t){return function isParentOf(e){return Boolean(r&&t===e||e.compareDocumentPosition(t)&Node.DOCUMENT_POSITION_CONTAINED_BY)}}throw new TypeError("util/compare-position#getParentComparator required either options.parent or options.element")}function whenKey(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n={};var t=nodeArray(e.context)[0]||document.documentElement;delete e.context;var r=nodeArray(e.filter);delete e.filter;var a=Object.keys(e);if(!a.length){throw new TypeError("when/key requires at least one option key")}var i=function registerBinding(e){e.keyCodes.forEach((function(t){if(!n[t]){n[t]=[]}n[t].push(e)}))};a.forEach((function(n){if(typeof e[n]!=="function"){throw new TypeError('when/key requires option["'+n+'"] to be a function')}var t=function addCallback(t){t.callback=e[n];return t};keyBinding(n).map(t).forEach(i)}));var o=function handleKeyDown(e){if(e.defaultPrevented){return}if(r.length){var a=getParentComparator({element:e.target,includeSelf:true});if(r.some(a)){return}}var i=e.keyCode||e.which;if(!n[i]){return}n[i].forEach((function(n){if(!n.matchModifiers(e)){return}n.callback.call(t,e,l)}))};t.addEventListener("keydown",o,false);var l=function disengage(){t.removeEventListener("keydown",o,false)};return{disengage:l}}function default_1(e){var n=e===void 0?{}:e,t=n.context;if(!t){t=document.documentElement}queryTabsequence();return whenKey({"?alt+?shift+tab":function altShiftTab(e){e.preventDefault();var n=queryTabsequence({context:t});var r=e.shiftKey;var a=n[0];var i=n[n.length-1];var o=r?a:i;var l=r?i:a;if(isActiveElement(o)){l.focus();return}var s=void 0;var u=n.some((function(e,n){if(!isActiveElement(e)){return false}s=n;return true}));if(!u){a.focus();return}var c=r?-1:1;n[s+c].focus()}})}n["default"]=default_1},173:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog-overlay] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    overflow: auto;\n    z-index: 9000;\n\n    display: flex;\n    align-content: center;\n    align-items: center;\n    flex-direction: column;\n    padding: 10vh 15px 0;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      padding: 15px 15px 0;\n    }\n  }\n\n  [data-nextjs-dialog-backdrop] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    background-color: rgba(17, 17, 17, 0.2);\n    pointer-events: all;\n    z-index: -1;\n  }\n\n  [data-nextjs-dialog-backdrop-fixed] {\n    cursor: not-allowed;\n    -webkit-backdrop-filter: blur(8px);\n    backdrop-filter: blur(8px);\n  }\n"],["\n  [data-nextjs-dialog-overlay] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    overflow: auto;\n    z-index: 9000;\n\n    display: flex;\n    align-content: center;\n    align-items: center;\n    flex-direction: column;\n    padding: 10vh 15px 0;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      padding: 15px 15px 0;\n    }\n  }\n\n  [data-nextjs-dialog-backdrop] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    background-color: rgba(17, 17, 17, 0.2);\n    pointer-events: all;\n    z-index: -1;\n  }\n\n  [data-nextjs-dialog-backdrop-fixed] {\n    cursor: not-allowed;\n    -webkit-backdrop-filter: blur(8px);\n    backdrop-filter: blur(8px);\n  }\n"])));n.styles=i;var o},395:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.ShadowPortal=void 0;var l=i(t(522));var s=t(255);var u=function Portal(e){var n=e.children;var t=l.useRef(null);var r=l.useRef(null);var a=l.useRef(null);var i=o(l.useState(),2),u=i[1];l.useLayoutEffect((function(){var e=t.current.ownerDocument;r.current=e.createElement("nextjs-portal");a.current=r.current.attachShadow({mode:"open"});e.body.appendChild(r.current);u({});return function(){if(r.current&&r.current.ownerDocument){r.current.ownerDocument.body.removeChild(r.current)}}}),[]);return a.current?(0,s.createPortal)(n,a.current):l.createElement("span",{ref:t})};n.ShadowPortal=u},261:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.Terminal=void 0;var s=l(t(194));var u=o(t(522));var c=function Terminal(e){var n=e.content;var t=u.useMemo((function(){return s["default"].ansiToJson(n,{json:true,use_classes:true,remove_empty:true})}),[n]);return u.createElement("div",{"data-nextjs-terminal":true},u.createElement("pre",null,t.map((function(e,n){return u.createElement("span",{key:"terminal-entry-"+n,style:r({color:e.fg?"var(--color-"+e.fg+")":undefined},e.decoration==="bold"?{fontWeight:800}:e.decoration==="italic"?{fontStyle:"italic"}:undefined)},e.content)}))))};n.Terminal=c},504:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Terminal=void 0;var a=t(261);r(n,a,"Terminal")},492:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-terminal] {\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-terminal]::selection,\n  [data-nextjs-terminal] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-terminal] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n  [data-nextjs-terminal] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n\n  [data-nextjs-terminal] pre {\n    white-space: pre-wrap;\n    word-break: break-word;\n  }\n"],["\n  [data-nextjs-terminal] {\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-terminal]::selection,\n  [data-nextjs-terminal] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-terminal] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n  [data-nextjs-terminal] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n\n  [data-nextjs-terminal] pre {\n    white-space: pre-wrap;\n    word-break: break-word;\n  }\n"])));n.styles=i;var o},314:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.Toast=void 0;var o=i(t(522));var l=function Toast(e){var n=e.onClick,t=e.children,r=e.className;return o.createElement("div",{"data-nextjs-toast":true,onClick:n,className:r},o.createElement("div",{"data-nextjs-toast-wrapper":true},t))};n.Toast=l},626:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Toast=n.styles=void 0;var a=t(565);r(n,a,"styles");var i=t(314);r(n,i,"Toast")},565:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-toast] {\n    position: fixed;\n    bottom: var(--size-gap-double);\n    left: var(--size-gap-double);\n    max-width: 420px;\n    z-index: 9000;\n  }\n\n  @media (max-width: 440px) {\n    [data-nextjs-toast] {\n      max-width: 90vw;\n      left: 5vw;\n    }\n  }\n\n  [data-nextjs-toast-wrapper] {\n    padding: 16px;\n    border-radius: var(--size-gap-half);\n    font-weight: 500;\n    color: var(--color-ansi-bright-white);\n    background-color: var(--color-ansi-red);\n    box-shadow: 0px var(--size-gap-double) var(--size-gap-quad)\n      rgba(0, 0, 0, 0.25);\n  }\n"],["\n  [data-nextjs-toast] {\n    position: fixed;\n    bottom: var(--size-gap-double);\n    left: var(--size-gap-double);\n    max-width: 420px;\n    z-index: 9000;\n  }\n\n  @media (max-width: 440px) {\n    [data-nextjs-toast] {\n      max-width: 90vw;\n      left: 5vw;\n    }\n  }\n\n  [data-nextjs-toast-wrapper] {\n    padding: 16px;\n    border-radius: var(--size-gap-half);\n    font-weight: 500;\n    color: var(--color-ansi-bright-white);\n    background-color: var(--color-ansi-red);\n    box-shadow: 0px var(--size-gap-double) var(--size-gap-quad)\n      rgba(0, 0, 0, 0.25);\n  }\n"])));n.styles=i;var o},71:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.styles=n.BuildError=void 0;var l=o(t(522));var s=t(824);var u=t(382);var c=t(504);var f=t(373);var d=function BuildError(e){var n=e.message;var t=l.useCallback((function(){}),[]);return l.createElement(u.Overlay,{fixed:true},l.createElement(s.Dialog,{type:"error","aria-labelledby":"nextjs__container_build_error_label","aria-describedby":"nextjs__container_build_error_desc",onClose:t},l.createElement(s.DialogContent,null,l.createElement(s.DialogHeader,{className:"nextjs-container-build-error-header"},l.createElement("h4",{id:"nextjs__container_build_error_label"},"Failed to compile")),l.createElement(s.DialogBody,{className:"nextjs-container-build-error-body"},l.createElement(c.Terminal,{content:n}),l.createElement("footer",null,l.createElement("p",{id:"nextjs__container_build_error_desc"},l.createElement("small",null,"This error occurred during the build process and can only be dismissed by fixing the error.")))))))};n.BuildError=d;n.styles=(0,f.noop)(p||(p=r(["\n  .nextjs-container-build-error-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-build-error-body footer {\n    margin-top: var(--size-gap);\n  }\n  .nextjs-container-build-error-body footer p {\n    margin: 0;\n  }\n\n  .nextjs-container-build-error-body small {\n    color: #757575;\n  }\n"],["\n  .nextjs-container-build-error-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-build-error-body footer {\n    margin-top: var(--size-gap);\n  }\n  .nextjs-container-build-error-body footer p {\n    margin: 0;\n  }\n\n  .nextjs-container-build-error-body small {\n    color: #757575;\n  }\n"])));var p},610:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__assign||function(){a=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return a.apply(this,arguments)};var i=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var o=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var l=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))i(n,e,t);o(n,e);return n};var s=this&&this.__awaiter||function(e,n,t,r){function adopt(e){return e instanceof t?e:new t((function(n){n(e)}))}return new(t||(t=Promise))((function(t,a){function fulfilled(e){try{step(r.next(e))}catch(e){a(e)}}function rejected(e){try{step(r["throw"](e))}catch(e){a(e)}}function step(e){e.done?t(e.value):adopt(e.value).then(fulfilled,rejected)}step((r=r.apply(e,n||[])).next())}))};var u=this&&this.__generator||function(e,n){var t={label:0,sent:function(){if(i[0]&1)throw i[1];return i[1]},trys:[],ops:[]},r,a,i,o;return o={next:verb(0),throw:verb(1),return:verb(2)},typeof Symbol==="function"&&(o[Symbol.iterator]=function(){return this}),o;function verb(e){return function(n){return step([e,n])}}function step(o){if(r)throw new TypeError("Generator is already executing.");while(t)try{if(r=1,a&&(i=o[0]&2?a["return"]:o[0]?a["throw"]||((i=a["return"])&&i.call(a),0):a.next)&&!(i=i.call(a,o[1])).done)return i;if(a=0,i)o=[o[0]&2,i.value];switch(o[0]){case 0:case 1:i=o;break;case 4:t.label++;return{value:o[1],done:false};case 5:t.label++;a=o[1];o=[0];continue;case 7:o=t.ops.pop();t.trys.pop();continue;default:if(!(i=t.trys,i=i.length>0&&i[i.length-1])&&(o[0]===6||o[0]===2)){t=0;continue}if(o[0]===3&&(!i||o[1]>i[0]&&o[1]<i[3])){t.label=o[1];break}if(o[0]===6&&t.label<i[1]){t.label=i[1];i=o;break}if(i&&t.label<i[2]){t.label=i[2];t.ops.push(o);break}if(i[2])t.ops.pop();t.trys.pop();continue}o=n.call(e,t)}catch(e){o=[6,e];a=0}finally{r=i=0}if(o[0]&5)throw o[1];return{value:o[0]?o[1]:void 0,done:true}}};var c=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.styles=n.Errors=void 0;var f=l(t(522));var d=t(591);var p=t(824);var v=t(830);var b=t(382);var m=t(626);var h=t(915);var g=t(373);var y=t(392);var x=t(157);function getErrorSignature(e){var n=e.event;switch(n.type){case d.TYPE_UNHANDLED_ERROR:case d.TYPE_UNHANDLED_REJECTION:{return n.reason.name+"::"+n.reason.message+"::"+n.reason.stack}default:{}}var t=n;return""}var w=function HotlinkedText(e){var n=e.text;var t=/https?:\/\/[^\s/$.?#].[^\s"]*/i;return f.createElement(f.Fragment,null,t.test(n)?n.split(" ").map((function(e,n,r){if(t.test(e)){return f.createElement(f.Fragment,{key:"link-"+n},f.createElement("a",{href:e},e),n===r.length-1?"":" ")}return n===r.length-1?f.createElement(f.Fragment,{key:"text-"+n},e):f.createElement(f.Fragment,{key:"text-"+n},e," ")})):n)};function getErrorByType(e){return s(this,void 0,void 0,(function(){var n,t,r,a;var i;return u(this,(function(o){switch(o.label){case 0:n=e.id,t=e.event;r=t.type;switch(r){case d.TYPE_UNHANDLED_ERROR:return[3,1];case d.TYPE_UNHANDLED_REJECTION:return[3,1]}return[3,3];case 1:i={id:n,runtime:true,error:t.reason};return[4,(0,y.getOriginalStackFrames)((0,h.isNodeError)(t.reason),t.frames)];case 2:return[2,(i.frames=o.sent(),i)];case 3:{return[3,4]}o.label=4;case 4:a=t;throw new Error("type system invariant violation")}}))}))}var _=function Errors(e){var n=e.errors;var t=c(f.useState({}),2),r=t[0],i=t[1];var o=c(f.useMemo((function(){var e=[];var t=null;for(var a=0;a<n.length;++a){var i=n[a];var o=i.id;if(o in r){e.push(r[o]);continue}if(a>0){var l=n[a-1];if(getErrorSignature(l)===getErrorSignature(i)){continue}}t=i;break}return[e,t]}),[n,r]),2),l=o[0],s=o[1];var u=f.useMemo((function(){return l.length<1&&Boolean(n.length)}),[n.length,l.length]);f.useEffect((function(){if(s==null){return}var e=true;getErrorByType(s).then((function(n){if(e){i((function(e){var t;return a(a({},e),(t={},t[n.id]=n,t))}))}}),(function(){}));return function(){e=false}}),[s]);var d=c(f.useState(false),2),g=d[0],y=d[1];var _=c(f.useState(0),2),E=_[0],O=_[1];var j=f.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();O((function(e){return Math.max(0,e-1)}))}),[]);var S=f.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();O((function(e){return Math.max(0,Math.min(l.length-1,e+1))}))}),[l.length]);var k=f.useMemo((function(){var e;return(e=l[E])!==null&&e!==void 0?e:null}),[E,l]);f.useEffect((function(){if(n.length<1){i({});y(false);O(0)}}),[n.length]);var T=f.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();y(true)}),[]);var C=f.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();y(false)}),[]);if(n.length<1||k==null){return null}if(u){return f.createElement(b.Overlay,null)}if(g){return f.createElement(m.Toast,{className:"nextjs-toast-errors-parent",onClick:C},f.createElement("div",{className:"nextjs-toast-errors"},f.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:"24",height:"24",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},f.createElement("circle",{cx:"12",cy:"12",r:"10"}),f.createElement("line",{x1:"12",y1:"8",x2:"12",y2:"12"}),f.createElement("line",{x1:"12",y1:"16",x2:"12.01",y2:"16"})),f.createElement("span",null,l.length," error",l.length>1?"s":"")))}var P=(0,h.isNodeError)(k.error);return f.createElement(b.Overlay,null,f.createElement(p.Dialog,{type:"error","aria-labelledby":"nextjs__container_errors_label","aria-describedby":"nextjs__container_errors_desc",onClose:P?undefined:T},f.createElement(p.DialogContent,null,f.createElement(p.DialogHeader,{className:"nextjs-container-errors-header"},f.createElement(v.LeftRightDialogHeader,{previous:E>0?j:null,next:E<l.length-1?S:null,close:P?undefined:T},f.createElement("small",null,f.createElement("span",null,E+1)," of"," ",f.createElement("span",null,l.length)," unhandled error",l.length<2?"":"s")),f.createElement("h1",{id:"nextjs__container_errors_label"},P?"Server Error":"Unhandled Runtime Error"),f.createElement("p",{id:"nextjs__container_errors_desc"},k.error.name,":"," ",f.createElement(w,{text:k.error.message})),P?f.createElement("div",null,f.createElement("small",null,"This error happened while generating the page. Any console logs will be displayed in the terminal window.")):undefined),f.createElement(p.DialogBody,{className:"nextjs-container-errors-body"},f.createElement(x.RuntimeError,{key:k.id.toString(),error:k})))))};n.Errors=_;n.styles=(0,g.noop)(E||(E=r(["\n  .nextjs-container-errors-header > h1 {\n    font-size: var(--size-font-big);\n    line-height: var(--size-font-bigger);\n    font-weight: bold;\n    margin: 0;\n    margin-top: calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  .nextjs-container-errors-header small {\n    font-size: var(--size-font-small);\n    color: var(--color-accents-1);\n    margin-left: var(--size-gap-double);\n  }\n  .nextjs-container-errors-header small > span {\n    font-family: var(--font-stack-monospace);\n  }\n  .nextjs-container-errors-header > p {\n    font-family: var(--font-stack-monospace);\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-big);\n    font-weight: bold;\n    margin: 0;\n    margin-top: var(--size-gap-half);\n    color: var(--color-ansi-red);\n    white-space: pre-wrap;\n  }\n  .nextjs-container-errors-header > div > small {\n    margin: 0;\n    margin-top: var(--size-gap-half);\n  }\n  .nextjs-container-errors-header > p > a {\n    color: var(--color-ansi-red);\n  }\n\n  .nextjs-container-errors-body > h5:not(:first-child) {\n    margin-top: calc(var(--size-gap-double) + var(--size-gap));\n  }\n  .nextjs-container-errors-body > h5 {\n    margin-bottom: var(--size-gap);\n  }\n\n  .nextjs-toast-errors-parent {\n    cursor: pointer;\n    transition: transform 0.2s ease;\n  }\n  .nextjs-toast-errors-parent:hover {\n    transform: scale(1.1);\n  }\n  .nextjs-toast-errors {\n    display: flex;\n    align-items: center;\n    justify-content: flex-start;\n  }\n  .nextjs-toast-errors > svg {\n    margin-right: var(--size-gap);\n  }\n"],["\n  .nextjs-container-errors-header > h1 {\n    font-size: var(--size-font-big);\n    line-height: var(--size-font-bigger);\n    font-weight: bold;\n    margin: 0;\n    margin-top: calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  .nextjs-container-errors-header small {\n    font-size: var(--size-font-small);\n    color: var(--color-accents-1);\n    margin-left: var(--size-gap-double);\n  }\n  .nextjs-container-errors-header small > span {\n    font-family: var(--font-stack-monospace);\n  }\n  .nextjs-container-errors-header > p {\n    font-family: var(--font-stack-monospace);\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-big);\n    font-weight: bold;\n    margin: 0;\n    margin-top: var(--size-gap-half);\n    color: var(--color-ansi-red);\n    white-space: pre-wrap;\n  }\n  .nextjs-container-errors-header > div > small {\n    margin: 0;\n    margin-top: var(--size-gap-half);\n  }\n  .nextjs-container-errors-header > p > a {\n    color: var(--color-ansi-red);\n  }\n\n  .nextjs-container-errors-body > h5:not(:first-child) {\n    margin-top: calc(var(--size-gap-double) + var(--size-gap));\n  }\n  .nextjs-container-errors-body > h5 {\n    margin-bottom: var(--size-gap);\n  }\n\n  .nextjs-toast-errors-parent {\n    cursor: pointer;\n    transition: transform 0.2s ease;\n  }\n  .nextjs-toast-errors-parent:hover {\n    transform: scale(1.1);\n  }\n  .nextjs-toast-errors {\n    display: flex;\n    align-items: center;\n    justify-content: flex-start;\n  }\n  .nextjs-toast-errors > svg {\n    margin-right: var(--size-gap);\n  }\n"])));var E},710:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.styles=n.FullRefreshWarning=void 0;var l=o(t(522));var s=t(824);var u=t(382);var c=t(504);var f=t(373);var d=function FullRefreshWarning(e){var n=e.reason;var t=l.useCallback((function(){window.location.reload()}),[]);return l.createElement(u.Overlay,{fixed:true},l.createElement(s.Dialog,{type:"warning","aria-labelledby":"nextjs__container_refresh_warning_label","aria-describedby":"nextjs__container_refresh_warning_desc",onClose:t},l.createElement(s.DialogContent,null,l.createElement(s.DialogHeader,{className:"nextjs-container-refresh-warning-header"},l.createElement("h4",{id:"nextjs__container_refresh_warning_label"},"About to perform a full refresh")),l.createElement(s.DialogBody,{className:"nextjs-container-refresh-warning-body"},l.createElement(FullRefreshWarningReason,{reason:n}),l.createElement("footer",null,l.createElement("p",null,"You can read more about Fast Refresh in"," ",l.createElement("a",{href:"https://nextjs.org/docs/basic-features/fast-refresh#how-it-works"},"our documentation"),"."),l.createElement("button",{onClick:t},"Reload"))))))};n.FullRefreshWarning=d;n.styles=(0,f.noop)(p||(p=r(["\n  .nextjs-container-refresh-warning-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-refresh-warning-body footer {\n    margin-top: var(--size-gap-double);\n  }\n\n  .nextjs-container-build-error-body p {\n    color: #757575;\n  }\n\n  .nextjs-container-refresh-warning-body button {\n    background-color: var(--color-ansi-yellow);\n    border: 0;\n    border-radius: var(--size-gap-half);\n    color: var(--color-ansi-black);\n    cursor: pointer;\n    display: block;\n    margin-left: auto;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n    transition: background-color 0.25s ease;\n  }\n\n  .nextjs-container-refresh-warning-body button:hover {\n    background-color: var(--color-ansi-bright-yellow);\n  }\n"],["\n  .nextjs-container-refresh-warning-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-refresh-warning-body footer {\n    margin-top: var(--size-gap-double);\n  }\n\n  .nextjs-container-build-error-body p {\n    color: #757575;\n  }\n\n  .nextjs-container-refresh-warning-body button {\n    background-color: var(--color-ansi-yellow);\n    border: 0;\n    border-radius: var(--size-gap-half);\n    color: var(--color-ansi-black);\n    cursor: pointer;\n    display: block;\n    margin-left: auto;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n    transition: background-color 0.25s ease;\n  }\n\n  .nextjs-container-refresh-warning-body button:hover {\n    background-color: var(--color-ansi-bright-yellow);\n  }\n"])));var FullRefreshWarningReason=function(e){var n=e.reason;if(n===null){return l.createElement("p",null,"Fast Refresh will perform a full reload because your application had an unrecoverable error.")}return l.createElement(l.Fragment,null,l.createElement("p",null,"Fast Refresh will perform a full reload when you edit a file that is imported by modules outside of the React rendering tree. It is also possible the parent component of the component you edited is a class component, which disables Fast Refresh. Fast Refresh requires at least one parent function component in your React tree."),l.createElement("p",null,"You can find more information in the related error below:"),l.createElement(c.Terminal,{content:n}))};var p},157:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.RuntimeError=n.styles=void 0;var s=o(t(522));var u=t(983);var c=t(373);var f=t(392);var d=function CallStackFrame(e){var n;var t=e.frame;var r=(n=t.originalStackFrame)!==null&&n!==void 0?n:t.sourceStackFrame;var a=Boolean(t.originalCodeFrame);var i=s.useCallback((function(){var e;if(!a)return;var n=new URLSearchParams;for(var t in r){n.append(t,((e=r[t])!==null&&e!==void 0?e:"").toString())}self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_launch-editor?"+n.toString()).then((function(){}),(function(){console.error("There was an issue opening this code in your editor.")}))}),[a,r]);return s.createElement("div",{"data-nextjs-call-stack-frame":true},s.createElement("h6",{"data-nextjs-frame-expanded":Boolean(t.expanded)},r.methodName),s.createElement("div",{"data-has-source":a?"true":undefined,tabIndex:a?10:undefined,role:a?"link":undefined,onClick:i,title:a?"Click to open in your editor":undefined},s.createElement("span",null,(0,f.getFrameSource)(r)),s.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},s.createElement("path",{d:"M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"}),s.createElement("polyline",{points:"15 3 21 3 21 9"}),s.createElement("line",{x1:"10",y1:"14",x2:"21",y2:"3"}))))};var p=function RuntimeError(e){var n=e.error;var t=s.useMemo((function(){return n.frames.findIndex((function(e){return e.expanded&&Boolean(e.originalCodeFrame)&&Boolean(e.originalStackFrame)}))}),[n.frames]);var r=s.useMemo((function(){var e;return(e=n.frames[t])!==null&&e!==void 0?e:null}),[n.frames,t]);var a=s.useMemo((function(){return t<0?[]:n.frames.slice(0,t)}),[n.frames,t]);var i=l(s.useState(r==null),2),o=i[0],c=i[1];var f=s.useCallback((function(){c((function(e){return!e}))}),[]);var p=s.useMemo((function(){return a.filter((function(e){return e.expanded||o}))}),[o,a]);var v=s.useMemo((function(){return n.frames.slice(t+1)}),[n.frames,t]);var b=s.useMemo((function(){return v.filter((function(e){return e.expanded||o}))}),[o,v]);var m=s.useMemo((function(){return v.length!==b.length||o&&r!=null}),[o,v.length,r,b.length]);return s.createElement(s.Fragment,null,r?s.createElement(s.Fragment,null,s.createElement("h5",null,"Source"),p.map((function(e,n){return s.createElement(d,{key:"leading-frame-"+n+"-"+o,frame:e})})),s.createElement(u.CodeFrame,{stackFrame:r.originalStackFrame,codeFrame:r.originalCodeFrame})):undefined,b.length?s.createElement(s.Fragment,null,s.createElement("h5",null,"Call Stack"),b.map((function(e,n){return s.createElement(d,{key:"call-stack-"+n+"-"+o,frame:e})}))):undefined,m?s.createElement(s.Fragment,null,s.createElement("button",{tabIndex:10,"data-nextjs-data-runtime-error-collapsed-action":true,type:"button",onClick:f},o?"Hide":"Show"," collapsed frames")):undefined)};n.RuntimeError=p;n.styles=(0,c.noop)(v||(v=r(["\n  button[data-nextjs-data-runtime-error-collapsed-action] {\n    background: none;\n    border: none;\n    padding: 0;\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-bigger);\n    color: var(--color-accents-3);\n  }\n\n  [data-nextjs-call-stack-frame]:not(:last-child) {\n    margin-bottom: var(--size-gap-double);\n  }\n\n  [data-nextjs-call-stack-frame] > h6 {\n    margin-top: 0;\n    margin-bottom: var(--size-gap);\n    font-family: var(--font-stack-monospace);\n    color: #222;\n  }\n  [data-nextjs-call-stack-frame] > h6[data-nextjs-frame-expanded='false'] {\n    color: #666;\n  }\n  [data-nextjs-call-stack-frame] > div {\n    display: flex;\n    align-items: center;\n    padding-left: calc(var(--size-gap) + var(--size-gap-half));\n    font-size: var(--size-font-small);\n    color: #999;\n  }\n  [data-nextjs-call-stack-frame] > div > svg {\n    width: auto;\n    height: var(--size-font-small);\n    margin-left: var(--size-gap);\n\n    display: none;\n  }\n\n  [data-nextjs-call-stack-frame] > div[data-has-source] {\n    cursor: pointer;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source]:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source] > svg {\n    display: unset;\n  }\n"],["\n  button[data-nextjs-data-runtime-error-collapsed-action] {\n    background: none;\n    border: none;\n    padding: 0;\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-bigger);\n    color: var(--color-accents-3);\n  }\n\n  [data-nextjs-call-stack-frame]:not(:last-child) {\n    margin-bottom: var(--size-gap-double);\n  }\n\n  [data-nextjs-call-stack-frame] > h6 {\n    margin-top: 0;\n    margin-bottom: var(--size-gap);\n    font-family: var(--font-stack-monospace);\n    color: #222;\n  }\n  [data-nextjs-call-stack-frame] > h6[data-nextjs-frame-expanded='false'] {\n    color: #666;\n  }\n  [data-nextjs-call-stack-frame] > div {\n    display: flex;\n    align-items: center;\n    padding-left: calc(var(--size-gap) + var(--size-gap-half));\n    font-size: var(--size-font-small);\n    color: #999;\n  }\n  [data-nextjs-call-stack-frame] > div > svg {\n    width: auto;\n    height: var(--size-font-small);\n    margin-left: var(--size-gap);\n\n    display: none;\n  }\n\n  [data-nextjs-call-stack-frame] > div[data-has-source] {\n    cursor: pointer;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source]:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source] > svg {\n    display: unset;\n  }\n"])));var v},915:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};n.__esModule=true;n.getNodeError=n.isNodeError=n.getFilesystemFrame=void 0;var a=t(922);function getFilesystemFrame(e){var n=r({},e);if(typeof n.file==="string"){if(n.file.startsWith("/")||/^[a-z]:\\/i.test(n.file)||n.file.startsWith("\\\\")){n.file="file://"+n.file}}return n}n.getFilesystemFrame=getFilesystemFrame;var i=Symbol("NextjsNodeError");function isNodeError(e){return i in e}n.isNodeError=isNodeError;function getNodeError(e){var n;try{throw new Error(e.message)}catch(e){n=e}n.name=e.name;try{n.stack=(0,a.parse)(e.stack).map(getFilesystemFrame).map((function(e){var n="    at "+e.methodName;if(e.file){var t=e.file;if(e.lineNumber){t+=":"+e.lineNumber;if(e.column){t+=":"+e.column}}n+=" ("+t+")"}return n})).join("\n")}catch(t){n.stack=e.stack}Object.defineProperty(n,i,{writable:false,enumerable:false,configurable:false});return n}n.getNodeError=getNodeError},373:function(e,n){"use strict";n.__esModule=true;n.noop=void 0;function noop(e){var n=[];for(var t=1;t<arguments.length;t++){n[t-1]=arguments[t]}var r=e.length-1;return e.slice(0,r).reduce((function(e,t,r){return e+t+n[r]}),"")+e[r]}n.noop=noop},525:function(e,n,t){"use strict";n.__esModule=true;n.parseStack=void 0;var r=t(922);var a=/\/_next(\/static\/.+)/g;function parseStack(e){var n=(0,r.parse)(e);return n.map((function(e){var n,t;try{var r=new URL(e.file);var i=a.exec(r.pathname);if(i){var o=(t=(n=process.env.__NEXT_DIST_DIR)===null||n===void 0?void 0:n.replace(/\\/g,"/"))===null||t===void 0?void 0:t.replace(/\/$/,"");if(o){e.file="file://"+o.concat(i.pop())}}}catch(e){}return e}))}n.parseStack=parseStack},392:function(e,n){"use strict";var t=this&&this.__awaiter||function(e,n,t,r){function adopt(e){return e instanceof t?e:new t((function(n){n(e)}))}return new(t||(t=Promise))((function(t,a){function fulfilled(e){try{step(r.next(e))}catch(e){a(e)}}function rejected(e){try{step(r["throw"](e))}catch(e){a(e)}}function step(e){e.done?t(e.value):adopt(e.value).then(fulfilled,rejected)}step((r=r.apply(e,n||[])).next())}))};var r=this&&this.__generator||function(e,n){var t={label:0,sent:function(){if(i[0]&1)throw i[1];return i[1]},trys:[],ops:[]},r,a,i,o;return o={next:verb(0),throw:verb(1),return:verb(2)},typeof Symbol==="function"&&(o[Symbol.iterator]=function(){return this}),o;function verb(e){return function(n){return step([e,n])}}function step(o){if(r)throw new TypeError("Generator is already executing.");while(t)try{if(r=1,a&&(i=o[0]&2?a["return"]:o[0]?a["throw"]||((i=a["return"])&&i.call(a),0):a.next)&&!(i=i.call(a,o[1])).done)return i;if(a=0,i)o=[o[0]&2,i.value];switch(o[0]){case 0:case 1:i=o;break;case 4:t.label++;return{value:o[1],done:false};case 5:t.label++;a=o[1];o=[0];continue;case 7:o=t.ops.pop();t.trys.pop();continue;default:if(!(i=t.trys,i=i.length>0&&i[i.length-1])&&(o[0]===6||o[0]===2)){t=0;continue}if(o[0]===3&&(!i||o[1]>i[0]&&o[1]<i[3])){t.label=o[1];break}if(o[0]===6&&t.label<i[1]){t.label=i[1];i=o;break}if(i&&t.label<i[2]){t.label=i[2];t.ops.push(o);break}if(i[2])t.ops.pop();t.trys.pop();continue}o=n.call(e,t)}catch(e){o=[6,e];a=0}finally{r=i=0}if(o[0]&5)throw o[1];return{value:o[0]?o[1]:void 0,done:true}}};n.__esModule=true;n.getFrameSource=n.getOriginalStackFrame=n.getOriginalStackFrames=void 0;function getOriginalStackFrames(e,n){return Promise.all(n.map((function(n){return getOriginalStackFrame(e,n)})))}n.getOriginalStackFrames=getOriginalStackFrames;function getOriginalStackFrame(e,n){var a,i;function _getOriginalStackFrame(){var a,i,o,l,s;return t(this,void 0,void 0,(function(){var t,u,c,f,d,p,v,b,m;return r(this,(function(r){switch(r.label){case 0:t=new URLSearchParams;t.append("isServerSide",String(e));for(u in n){t.append(u,((a=n[u])!==null&&a!==void 0?a:"").toString())}c=new AbortController;f=setTimeout((function(){return c.abort()}),3e3);return[4,self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_original-stack-frame?"+t.toString(),{signal:c.signal})["finally"]((function(){clearTimeout(f)}))];case 1:d=r.sent();if(!(!d.ok||d.status===204))return[3,3];v=(p=Promise).reject;b=Error.bind;return[4,d.text()];case 2:return[2,v.apply(p,[new(b.apply(Error,[void 0,r.sent()]))])];case 3:return[4,d.json()];case 4:m=r.sent();return[2,{error:false,reason:null,external:false,expanded:!Boolean((s=((i=n.file)===null||i===void 0?void 0:i.includes("node_modules"))||((l=(o=m.originalStackFrame)===null||o===void 0?void 0:o.file)===null||l===void 0?void 0:l.includes("node_modules")))!==null&&s!==void 0?s:true),sourceStackFrame:n,originalStackFrame:m.originalStackFrame,originalCodeFrame:m.originalCodeFrame||null}]}}))}))}if(!(((a=n.file)===null||a===void 0?void 0:a.startsWith("webpack-internal:"))||((i=n.file)===null||i===void 0?void 0:i.startsWith("file:")))){return Promise.resolve({error:false,reason:null,external:true,expanded:false,sourceStackFrame:n,originalStackFrame:null,originalCodeFrame:null})}return _getOriginalStackFrame()["catch"]((function(e){var t,r;return{error:true,reason:(r=(t=e===null||e===void 0?void 0:e.message)!==null&&t!==void 0?t:e===null||e===void 0?void 0:e.toString())!==null&&r!==void 0?r:"Unknown Error",external:false,expanded:false,sourceStackFrame:n,originalStackFrame:null,originalCodeFrame:null}}))}n.getOriginalStackFrame=getOriginalStackFrame;function getFrameSource(e){var n;var t="";try{var r=new URL(e.file);if(typeof globalThis!=="undefined"&&((n=globalThis.location)===null||n===void 0?void 0:n.origin)!==r.origin){if(r.origin==="null"){t+=r.protocol}else{t+=r.origin}}t+=r.pathname;t+=" "}catch(n){t+=(e.file||"(unknown)")+" "}if(e.lineNumber!=null){if(e.column!=null){t+="("+e.lineNumber+":"+e.column+") "}else{t+="("+e.lineNumber+") "}}return t.slice(0,-1)}n.getFrameSource=getFrameSource},814:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.useOnClickOutside=void 0;var o=i(t(522));function useOnClickOutside(e,n){o.useEffect((function(){if(e==null||n==null){return}var listener=function(t){if(!e||e.contains(t.target)){return}n(t)};var t=e.getRootNode();t.addEventListener("mousedown",listener);t.addEventListener("touchstart",listener);return function(){t.removeEventListener("mousedown",listener);t.removeEventListener("touchstart",listener)}}),[n,e])}n.useOnClickOutside=useOnClickOutside},796:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.Base=void 0;var l=o(t(522));var s=t(373);function Base(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,s.noop)(u||(u=r(["\n          :host {\n            --size-gap-half: 4px;\n            --size-gap: 8px;\n            --size-gap-double: 16px;\n            --size-gap-quad: 32px;\n\n            --size-font-small: 14px;\n            --size-font: 16px;\n            --size-font-big: 20px;\n            --size-font-bigger: 24px;\n\n            --color-accents-1: #808080;\n            --color-accents-2: #222222;\n            --color-accents-3: #404040;\n\n            --font-stack-monospace: 'SFMono-Regular', Consolas,\n              'Liberation Mono', Menlo, Courier, monospace;\n\n            --color-ansi-selection: rgba(95, 126, 151, 0.48);\n            --color-ansi-bg: #111111;\n            --color-ansi-fg: #cccccc;\n\n            --color-ansi-white: #777777;\n            --color-ansi-black: #141414;\n            --color-ansi-blue: #00aaff;\n            --color-ansi-cyan: #88ddff;\n            --color-ansi-green: #98ec65;\n            --color-ansi-magenta: #aa88ff;\n            --color-ansi-red: #ff5555;\n            --color-ansi-yellow: #ffcc33;\n            --color-ansi-bright-white: #ffffff;\n            --color-ansi-bright-black: #777777;\n            --color-ansi-bright-blue: #33bbff;\n            --color-ansi-bright-cyan: #bbecff;\n            --color-ansi-bright-green: #b6f292;\n            --color-ansi-bright-magenta: #cebbff;\n            --color-ansi-bright-red: #ff8888;\n            --color-ansi-bright-yellow: #ffd966;\n          }\n\n          .mono {\n            font-family: var(--font-stack-monospace);\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-bottom: var(--size-gap);\n            font-weight: 500;\n            line-height: 1.5;\n          }\n\n          h1 {\n            font-size: 40px;\n          }\n          h2 {\n            font-size: 32px;\n          }\n          h3 {\n            font-size: 28px;\n          }\n          h4 {\n            font-size: 24px;\n          }\n          h5 {\n            font-size: 20px;\n          }\n          h6 {\n            font-size: 16px;\n          }\n        "],["\n          :host {\n            --size-gap-half: 4px;\n            --size-gap: 8px;\n            --size-gap-double: 16px;\n            --size-gap-quad: 32px;\n\n            --size-font-small: 14px;\n            --size-font: 16px;\n            --size-font-big: 20px;\n            --size-font-bigger: 24px;\n\n            --color-accents-1: #808080;\n            --color-accents-2: #222222;\n            --color-accents-3: #404040;\n\n            --font-stack-monospace: 'SFMono-Regular', Consolas,\n              'Liberation Mono', Menlo, Courier, monospace;\n\n            --color-ansi-selection: rgba(95, 126, 151, 0.48);\n            --color-ansi-bg: #111111;\n            --color-ansi-fg: #cccccc;\n\n            --color-ansi-white: #777777;\n            --color-ansi-black: #141414;\n            --color-ansi-blue: #00aaff;\n            --color-ansi-cyan: #88ddff;\n            --color-ansi-green: #98ec65;\n            --color-ansi-magenta: #aa88ff;\n            --color-ansi-red: #ff5555;\n            --color-ansi-yellow: #ffcc33;\n            --color-ansi-bright-white: #ffffff;\n            --color-ansi-bright-black: #777777;\n            --color-ansi-bright-blue: #33bbff;\n            --color-ansi-bright-cyan: #bbecff;\n            --color-ansi-bright-green: #b6f292;\n            --color-ansi-bright-magenta: #cebbff;\n            --color-ansi-bright-red: #ff8888;\n            --color-ansi-bright-yellow: #ffd966;\n          }\n\n          .mono {\n            font-family: var(--font-stack-monospace);\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-bottom: var(--size-gap);\n            font-weight: 500;\n            line-height: 1.5;\n          }\n\n          h1 {\n            font-size: 40px;\n          }\n          h2 {\n            font-size: 32px;\n          }\n          h3 {\n            font-size: 28px;\n          }\n          h4 {\n            font-size: 24px;\n          }\n          h5 {\n            font-size: 20px;\n          }\n          h6 {\n            font-size: 16px;\n          }\n        "])))}})}n.Base=Base;var u},174:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.ComponentStyles=void 0;var l=o(t(522));var s=t(761);var u=t(824);var c=t(448);var f=t(173);var d=t(492);var p=t(626);var v=t(71);var b=t(610);var m=t(710);var h=t(157);var g=t(373);function ComponentStyles(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,g.noop)(y||(y=r(["\n          ","\n          ","\n          ","\n          ","\n          ","\n          ","\n\n          ","\n          ","\n          ","\n          ","\n        "],["\n          ","\n          ","\n          ","\n          ","\n          ","\n          ","\n\n          ","\n          ","\n          ","\n          ","\n        "])),f.styles,p.styles,u.styles,c.styles,s.styles,d.styles,v.styles,b.styles,h.styles,m.styles)}})}n.ComponentStyles=ComponentStyles;var y},803:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.CssReset=void 0;var l=o(t(522));var s=t(373);function CssReset(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,s.noop)(u||(u=r(["\n          :host {\n            all: initial;\n\n            /* the direction property is not reset by 'all' */\n            direction: ltr;\n          }\n\n          /*!\n           * Bootstrap Reboot v4.4.1 (https://getbootstrap.com/)\n           * Copyright 2011-2019 The Bootstrap Authors\n           * Copyright 2011-2019 Twitter, Inc.\n           * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)\n           * Forked from Normalize.css, licensed MIT (https://github.com/necolas/normalize.css/blob/master/LICENSE.md)\n           */\n          *,\n          *::before,\n          *::after {\n            box-sizing: border-box;\n          }\n\n          :host {\n            font-family: sans-serif;\n            line-height: 1.15;\n            -webkit-text-size-adjust: 100%;\n            -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n          }\n\n          article,\n          aside,\n          figcaption,\n          figure,\n          footer,\n          header,\n          hgroup,\n          main,\n          nav,\n          section {\n            display: block;\n          }\n\n          :host {\n            margin: 0;\n            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,\n              'Helvetica Neue', Arial, 'Noto Sans', sans-serif,\n              'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',\n              'Noto Color Emoji';\n            font-size: 16px;\n            font-weight: 400;\n            line-height: 1.5;\n            color: #212529;\n            text-align: left;\n            background-color: #fff;\n          }\n\n          [tabindex='-1']:focus:not(:focus-visible) {\n            outline: 0 !important;\n          }\n\n          hr {\n            box-sizing: content-box;\n            height: 0;\n            overflow: visible;\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-top: 0;\n            margin-bottom: 8px;\n          }\n\n          p {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          abbr[title],\n          abbr[data-original-title] {\n            text-decoration: underline;\n            -webkit-text-decoration: underline dotted;\n            text-decoration: underline dotted;\n            cursor: help;\n            border-bottom: 0;\n            -webkit-text-decoration-skip-ink: none;\n            text-decoration-skip-ink: none;\n          }\n\n          address {\n            margin-bottom: 16px;\n            font-style: normal;\n            line-height: inherit;\n          }\n\n          ol,\n          ul,\n          dl {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          ol ol,\n          ul ul,\n          ol ul,\n          ul ol {\n            margin-bottom: 0;\n          }\n\n          dt {\n            font-weight: 700;\n          }\n\n          dd {\n            margin-bottom: 8px;\n            margin-left: 0;\n          }\n\n          blockquote {\n            margin: 0 0 16px;\n          }\n\n          b,\n          strong {\n            font-weight: bolder;\n          }\n\n          small {\n            font-size: 80%;\n          }\n\n          sub,\n          sup {\n            position: relative;\n            font-size: 75%;\n            line-height: 0;\n            vertical-align: baseline;\n          }\n\n          sub {\n            bottom: -0.25em;\n          }\n\n          sup {\n            top: -0.5em;\n          }\n\n          a {\n            color: #007bff;\n            text-decoration: none;\n            background-color: transparent;\n          }\n\n          a:hover {\n            color: #0056b3;\n            text-decoration: underline;\n          }\n\n          a:not([href]) {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          a:not([href]):hover {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          pre,\n          code,\n          kbd,\n          samp {\n            font-family: SFMono-Regular, Menlo, Monaco, Consolas,\n              'Liberation Mono', 'Courier New', monospace;\n            font-size: 1em;\n          }\n\n          pre {\n            margin-top: 0;\n            margin-bottom: 16px;\n            overflow: auto;\n          }\n\n          figure {\n            margin: 0 0 16px;\n          }\n\n          img {\n            vertical-align: middle;\n            border-style: none;\n          }\n\n          svg {\n            overflow: hidden;\n            vertical-align: middle;\n          }\n\n          table {\n            border-collapse: collapse;\n          }\n\n          caption {\n            padding-top: 12px;\n            padding-bottom: 12px;\n            color: #6c757d;\n            text-align: left;\n            caption-side: bottom;\n          }\n\n          th {\n            text-align: inherit;\n          }\n\n          label {\n            display: inline-block;\n            margin-bottom: 8px;\n          }\n\n          button {\n            border-radius: 0;\n          }\n\n          button:focus {\n            outline: 1px dotted;\n            outline: 5px auto -webkit-focus-ring-color;\n          }\n\n          input,\n          button,\n          select,\n          optgroup,\n          textarea {\n            margin: 0;\n            font-family: inherit;\n            font-size: inherit;\n            line-height: inherit;\n          }\n\n          button,\n          input {\n            overflow: visible;\n          }\n\n          button,\n          select {\n            text-transform: none;\n          }\n\n          select {\n            word-wrap: normal;\n          }\n\n          button,\n          [type='button'],\n          [type='reset'],\n          [type='submit'] {\n            -webkit-appearance: button;\n          }\n\n          button:not(:disabled),\n          [type='button']:not(:disabled),\n          [type='reset']:not(:disabled),\n          [type='submit']:not(:disabled) {\n            cursor: pointer;\n          }\n\n          button::-moz-focus-inner,\n          [type='button']::-moz-focus-inner,\n          [type='reset']::-moz-focus-inner,\n          [type='submit']::-moz-focus-inner {\n            padding: 0;\n            border-style: none;\n          }\n\n          input[type='radio'],\n          input[type='checkbox'] {\n            box-sizing: border-box;\n            padding: 0;\n          }\n\n          input[type='date'],\n          input[type='time'],\n          input[type='datetime-local'],\n          input[type='month'] {\n            -webkit-appearance: listbox;\n          }\n\n          textarea {\n            overflow: auto;\n            resize: vertical;\n          }\n\n          fieldset {\n            min-width: 0;\n            padding: 0;\n            margin: 0;\n            border: 0;\n          }\n\n          legend {\n            display: block;\n            width: 100%;\n            max-width: 100%;\n            padding: 0;\n            margin-bottom: 8px;\n            font-size: 24px;\n            line-height: inherit;\n            color: inherit;\n            white-space: normal;\n          }\n\n          progress {\n            vertical-align: baseline;\n          }\n\n          [type='number']::-webkit-inner-spin-button,\n          [type='number']::-webkit-outer-spin-button {\n            height: auto;\n          }\n\n          [type='search'] {\n            outline-offset: -2px;\n            -webkit-appearance: none;\n          }\n\n          [type='search']::-webkit-search-decoration {\n            -webkit-appearance: none;\n          }\n\n          ::-webkit-file-upload-button {\n            font: inherit;\n            -webkit-appearance: button;\n          }\n\n          output {\n            display: inline-block;\n          }\n\n          summary {\n            display: list-item;\n            cursor: pointer;\n          }\n\n          template {\n            display: none;\n          }\n\n          [hidden] {\n            display: none !important;\n          }\n        "],["\n          :host {\n            all: initial;\n\n            /* the direction property is not reset by 'all' */\n            direction: ltr;\n          }\n\n          /*!\n           * Bootstrap Reboot v4.4.1 (https://getbootstrap.com/)\n           * Copyright 2011-2019 The Bootstrap Authors\n           * Copyright 2011-2019 Twitter, Inc.\n           * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)\n           * Forked from Normalize.css, licensed MIT (https://github.com/necolas/normalize.css/blob/master/LICENSE.md)\n           */\n          *,\n          *::before,\n          *::after {\n            box-sizing: border-box;\n          }\n\n          :host {\n            font-family: sans-serif;\n            line-height: 1.15;\n            -webkit-text-size-adjust: 100%;\n            -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n          }\n\n          article,\n          aside,\n          figcaption,\n          figure,\n          footer,\n          header,\n          hgroup,\n          main,\n          nav,\n          section {\n            display: block;\n          }\n\n          :host {\n            margin: 0;\n            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,\n              'Helvetica Neue', Arial, 'Noto Sans', sans-serif,\n              'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',\n              'Noto Color Emoji';\n            font-size: 16px;\n            font-weight: 400;\n            line-height: 1.5;\n            color: #212529;\n            text-align: left;\n            background-color: #fff;\n          }\n\n          [tabindex='-1']:focus:not(:focus-visible) {\n            outline: 0 !important;\n          }\n\n          hr {\n            box-sizing: content-box;\n            height: 0;\n            overflow: visible;\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-top: 0;\n            margin-bottom: 8px;\n          }\n\n          p {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          abbr[title],\n          abbr[data-original-title] {\n            text-decoration: underline;\n            -webkit-text-decoration: underline dotted;\n            text-decoration: underline dotted;\n            cursor: help;\n            border-bottom: 0;\n            -webkit-text-decoration-skip-ink: none;\n            text-decoration-skip-ink: none;\n          }\n\n          address {\n            margin-bottom: 16px;\n            font-style: normal;\n            line-height: inherit;\n          }\n\n          ol,\n          ul,\n          dl {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          ol ol,\n          ul ul,\n          ol ul,\n          ul ol {\n            margin-bottom: 0;\n          }\n\n          dt {\n            font-weight: 700;\n          }\n\n          dd {\n            margin-bottom: 8px;\n            margin-left: 0;\n          }\n\n          blockquote {\n            margin: 0 0 16px;\n          }\n\n          b,\n          strong {\n            font-weight: bolder;\n          }\n\n          small {\n            font-size: 80%;\n          }\n\n          sub,\n          sup {\n            position: relative;\n            font-size: 75%;\n            line-height: 0;\n            vertical-align: baseline;\n          }\n\n          sub {\n            bottom: -0.25em;\n          }\n\n          sup {\n            top: -0.5em;\n          }\n\n          a {\n            color: #007bff;\n            text-decoration: none;\n            background-color: transparent;\n          }\n\n          a:hover {\n            color: #0056b3;\n            text-decoration: underline;\n          }\n\n          a:not([href]) {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          a:not([href]):hover {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          pre,\n          code,\n          kbd,\n          samp {\n            font-family: SFMono-Regular, Menlo, Monaco, Consolas,\n              'Liberation Mono', 'Courier New', monospace;\n            font-size: 1em;\n          }\n\n          pre {\n            margin-top: 0;\n            margin-bottom: 16px;\n            overflow: auto;\n          }\n\n          figure {\n            margin: 0 0 16px;\n          }\n\n          img {\n            vertical-align: middle;\n            border-style: none;\n          }\n\n          svg {\n            overflow: hidden;\n            vertical-align: middle;\n          }\n\n          table {\n            border-collapse: collapse;\n          }\n\n          caption {\n            padding-top: 12px;\n            padding-bottom: 12px;\n            color: #6c757d;\n            text-align: left;\n            caption-side: bottom;\n          }\n\n          th {\n            text-align: inherit;\n          }\n\n          label {\n            display: inline-block;\n            margin-bottom: 8px;\n          }\n\n          button {\n            border-radius: 0;\n          }\n\n          button:focus {\n            outline: 1px dotted;\n            outline: 5px auto -webkit-focus-ring-color;\n          }\n\n          input,\n          button,\n          select,\n          optgroup,\n          textarea {\n            margin: 0;\n            font-family: inherit;\n            font-size: inherit;\n            line-height: inherit;\n          }\n\n          button,\n          input {\n            overflow: visible;\n          }\n\n          button,\n          select {\n            text-transform: none;\n          }\n\n          select {\n            word-wrap: normal;\n          }\n\n          button,\n          [type='button'],\n          [type='reset'],\n          [type='submit'] {\n            -webkit-appearance: button;\n          }\n\n          button:not(:disabled),\n          [type='button']:not(:disabled),\n          [type='reset']:not(:disabled),\n          [type='submit']:not(:disabled) {\n            cursor: pointer;\n          }\n\n          button::-moz-focus-inner,\n          [type='button']::-moz-focus-inner,\n          [type='reset']::-moz-focus-inner,\n          [type='submit']::-moz-focus-inner {\n            padding: 0;\n            border-style: none;\n          }\n\n          input[type='radio'],\n          input[type='checkbox'] {\n            box-sizing: border-box;\n            padding: 0;\n          }\n\n          input[type='date'],\n          input[type='time'],\n          input[type='datetime-local'],\n          input[type='month'] {\n            -webkit-appearance: listbox;\n          }\n\n          textarea {\n            overflow: auto;\n            resize: vertical;\n          }\n\n          fieldset {\n            min-width: 0;\n            padding: 0;\n            margin: 0;\n            border: 0;\n          }\n\n          legend {\n            display: block;\n            width: 100%;\n            max-width: 100%;\n            padding: 0;\n            margin-bottom: 8px;\n            font-size: 24px;\n            line-height: inherit;\n            color: inherit;\n            white-space: normal;\n          }\n\n          progress {\n            vertical-align: baseline;\n          }\n\n          [type='number']::-webkit-inner-spin-button,\n          [type='number']::-webkit-outer-spin-button {\n            height: auto;\n          }\n\n          [type='search'] {\n            outline-offset: -2px;\n            -webkit-appearance: none;\n          }\n\n          [type='search']::-webkit-search-decoration {\n            -webkit-appearance: none;\n          }\n\n          ::-webkit-file-upload-button {\n            font: inherit;\n            -webkit-appearance: button;\n          }\n\n          output {\n            display: inline-block;\n          }\n\n          summary {\n            display: list-item;\n            cursor: pointer;\n          }\n\n          template {\n            display: none;\n          }\n\n          [hidden] {\n            display: none !important;\n          }\n        "])))}})}n.CssReset=CssReset;var u},518:function(e){"use strict";e.exports=require("next/dist/compiled/strip-ansi")},522:function(e){"use strict";e.exports=require("react")},255:function(e){"use strict";e.exports=require("react-dom")}};var n={};function __nccwpck_require__(t){var r=n[t];if(r!==undefined){return r.exports}var a=n[t]={id:t,loaded:false,exports:{}};var i=true;try{e[t].call(a.exports,a,a.exports,__nccwpck_require__);i=false}finally{if(i)delete n[t]}a.loaded=true;return a.exports}!function(){__nccwpck_require__.nmd=function(e){e.paths=[];if(!e.children)e.children=[];return e}}();if(typeof __nccwpck_require__!=="undefined")__nccwpck_require__.ab=__dirname+"/";var t=__nccwpck_require__(977);module.exports=t})();
+ */(function(){"use strict";var t={function:true,object:true};var r=t[typeof window]&&window||this;var a=r;var i=t[typeof n]&&n;var o=t["object"]&&e&&!e.nodeType&&e;var l=i&&o&&typeof global=="object"&&global;if(l&&(l.global===l||l.window===l||l.self===l)){r=l}var s=Math.pow(2,53)-1;var u=/\bOpera/;var c=this;var f=Object.prototype;var d=f.hasOwnProperty;var p=f.toString;function capitalize(e){e=String(e);return e.charAt(0).toUpperCase()+e.slice(1)}function cleanupOS(e,n,t){var r={"10.0":"10",6.4:"10 Technical Preview",6.3:"8.1",6.2:"8",6.1:"Server 2008 R2 / 7","6.0":"Server 2008 / Vista",5.2:"Server 2003 / XP 64-bit",5.1:"XP",5.01:"2000 SP1","5.0":"2000","4.0":"NT","4.90":"ME"};if(n&&t&&/^Win/i.test(e)&&!/^Windows Phone /i.test(e)&&(r=r[/[\d.]+$/.exec(e)])){e="Windows "+r}e=String(e);if(n&&t){e=e.replace(RegExp(n,"i"),t)}e=format(e.replace(/ ce$/i," CE").replace(/\bhpw/i,"web").replace(/\bMacintosh\b/,"Mac OS").replace(/_PowerPC\b/i," OS").replace(/\b(OS X) [^ \d]+/i,"$1").replace(/\bMac (OS X)\b/,"$1").replace(/\/(\d)/," $1").replace(/_/g,".").replace(/(?: BePC|[ .]*fc[ \d.]+)$/i,"").replace(/\bx86\.64\b/gi,"x86_64").replace(/\b(Windows Phone) OS\b/,"$1").replace(/\b(Chrome OS \w+) [\d.]+\b/,"$1").split(" on ")[0]);return e}function each(e,n){var t=-1,r=e?e.length:0;if(typeof r=="number"&&r>-1&&r<=s){while(++t<r){n(e[t],t,e)}}else{forOwn(e,n)}}function format(e){e=trim(e);return/^(?:webOS|i(?:OS|P))/.test(e)?e:capitalize(e)}function forOwn(e,n){for(var t in e){if(d.call(e,t)){n(e[t],t,e)}}}function getClassOf(e){return e==null?capitalize(e):p.call(e).slice(8,-1)}function isHostType(e,n){var t=e!=null?typeof e[n]:"number";return!/^(?:boolean|number|string|undefined)$/.test(t)&&(t=="object"?!!e[n]:true)}function qualify(e){return String(e).replace(/([ -])(?!$)/g,"$1?")}function reduce(e,n){var t=null;each(e,(function(r,a){t=n(t,r,a,e)}));return t}function trim(e){return String(e).replace(/^ +| +$/g,"")}function parse(e){var n=r;var t=e&&typeof e=="object"&&getClassOf(e)!="String";if(t){n=e;e=null}var i=n.navigator||{};var o=i.userAgent||"";e||(e=o);var l=t||c==a;var s=t?!!i.likeChrome:/\bChrome\b/.test(e)&&!/internal|\n/i.test(p.toString());var f="Object",d=t?f:"ScriptBridgingProxyObject",v=t?f:"Environment",b=t&&n.java?"JavaPackage":getClassOf(n.java),m=t?f:"RuntimeObject";var h=/\bJava/.test(b)&&n.java;var g=h&&getClassOf(n.environment)==v;var y=h?"a":"α";var x=h?"b":"β";var w=n.document||{};var _=n.operamini||n.opera;var E=u.test(E=t&&_?_["[[Class]]"]:getClassOf(_))?E:_=null;var O;var j=e;var S=[];var k=null;var T=e==o;var C=T&&_&&typeof _.version=="function"&&_.version();var P;var M=getLayout([{label:"EdgeHTML",pattern:"Edge"},"Trident",{label:"WebKit",pattern:"AppleWebKit"},"iCab","Presto","NetFront","Tasman","KHTML","Gecko"]);var A=getName(["Adobe AIR","Arora","Avant Browser","Breach","Camino","Electron","Epiphany","Fennec","Flock","Galeon","GreenBrowser","iCab","Iceweasel","K-Meleon","Konqueror","Lunascape","Maxthon",{label:"Microsoft Edge",pattern:"(?:Edge|Edg|EdgA|EdgiOS)"},"Midori","Nook Browser","PaleMoon","PhantomJS","Raven","Rekonq","RockMelt",{label:"Samsung Internet",pattern:"SamsungBrowser"},"SeaMonkey",{label:"Silk",pattern:"(?:Cloud9|Silk-Accelerated)"},"Sleipnir","SlimBrowser",{label:"SRWare Iron",pattern:"Iron"},"Sunrise","Swiftfox","Vivaldi","Waterfox","WebPositive",{label:"Yandex Browser",pattern:"YaBrowser"},{label:"UC Browser",pattern:"UCBrowser"},"Opera Mini",{label:"Opera Mini",pattern:"OPiOS"},"Opera",{label:"Opera",pattern:"OPR"},"Chromium","Chrome",{label:"Chrome",pattern:"(?:HeadlessChrome)"},{label:"Chrome Mobile",pattern:"(?:CriOS|CrMo)"},{label:"Firefox",pattern:"(?:Firefox|Minefield)"},{label:"Firefox for iOS",pattern:"FxiOS"},{label:"IE",pattern:"IEMobile"},{label:"IE",pattern:"MSIE"},"Safari"]);var R=getProduct([{label:"BlackBerry",pattern:"BB10"},"BlackBerry",{label:"Galaxy S",pattern:"GT-I9000"},{label:"Galaxy S2",pattern:"GT-I9100"},{label:"Galaxy S3",pattern:"GT-I9300"},{label:"Galaxy S4",pattern:"GT-I9500"},{label:"Galaxy S5",pattern:"SM-G900"},{label:"Galaxy S6",pattern:"SM-G920"},{label:"Galaxy S6 Edge",pattern:"SM-G925"},{label:"Galaxy S7",pattern:"SM-G930"},{label:"Galaxy S7 Edge",pattern:"SM-G935"},"Google TV","Lumia","iPad","iPod","iPhone","Kindle",{label:"Kindle Fire",pattern:"(?:Cloud9|Silk-Accelerated)"},"Nexus","Nook","PlayBook","PlayStation Vita","PlayStation","TouchPad","Transformer",{label:"Wii U",pattern:"WiiU"},"Wii","Xbox One",{label:"Xbox 360",pattern:"Xbox"},"Xoom"]);var N=getManufacturer({Apple:{iPad:1,iPhone:1,iPod:1},Alcatel:{},Archos:{},Amazon:{Kindle:1,"Kindle Fire":1},Asus:{Transformer:1},"Barnes & Noble":{Nook:1},BlackBerry:{PlayBook:1},Google:{"Google TV":1,Nexus:1},HP:{TouchPad:1},HTC:{},Huawei:{},Lenovo:{},LG:{},Microsoft:{Xbox:1,"Xbox One":1},Motorola:{Xoom:1},Nintendo:{"Wii U":1,Wii:1},Nokia:{Lumia:1},Oppo:{},Samsung:{"Galaxy S":1,"Galaxy S2":1,"Galaxy S3":1,"Galaxy S4":1},Sony:{PlayStation:1,"PlayStation Vita":1},Xiaomi:{Mi:1,Redmi:1}});var D=getOS(["Windows Phone","KaiOS","Android","CentOS",{label:"Chrome OS",pattern:"CrOS"},"Debian",{label:"DragonFly BSD",pattern:"DragonFly"},"Fedora","FreeBSD","Gentoo","Haiku","Kubuntu","Linux Mint","OpenBSD","Red Hat","SuSE","Ubuntu","Xubuntu","Cygwin","Symbian OS","hpwOS","webOS ","webOS","Tablet OS","Tizen","Linux","Mac OS X","Macintosh","Mac","Windows 98;","Windows "]);function getLayout(n){return reduce(n,(function(n,t){return n||RegExp("\\b"+(t.pattern||qualify(t))+"\\b","i").exec(e)&&(t.label||t)}))}function getManufacturer(n){return reduce(n,(function(n,t,r){return n||(t[R]||t[/^[a-z]+(?: +[a-z]+\b)*/i.exec(R)]||RegExp("\\b"+qualify(r)+"(?:\\b|\\w*\\d)","i").exec(e))&&r}))}function getName(n){return reduce(n,(function(n,t){return n||RegExp("\\b"+(t.pattern||qualify(t))+"\\b","i").exec(e)&&(t.label||t)}))}function getOS(n){return reduce(n,(function(n,t){var r=t.pattern||qualify(t);if(!n&&(n=RegExp("\\b"+r+"(?:/[\\d.]+|[ \\w.]*)","i").exec(e))){n=cleanupOS(n,r,t.label||t)}return n}))}function getProduct(n){return reduce(n,(function(n,t){var r=t.pattern||qualify(t);if(!n&&(n=RegExp("\\b"+r+" *\\d+[.\\w_]*","i").exec(e)||RegExp("\\b"+r+" *\\w+-[\\w]*","i").exec(e)||RegExp("\\b"+r+"(?:; *(?:[a-z]+[_-])?[a-z]+\\d+|[^ ();-]*)","i").exec(e))){if((n=String(t.label&&!RegExp(r,"i").test(t.label)?t.label:n).split("/"))[1]&&!/[\d.]+/.test(n[0])){n[0]+=" "+n[1]}t=t.label||t;n=format(n[0].replace(RegExp(r,"i"),t).replace(RegExp("; *(?:"+t+"[_-])?","i")," ").replace(RegExp("("+t+")[-_.]?(\\w)","i"),"$1 $2"))}return n}))}function getVersion(n){return reduce(n,(function(n,t){return n||(RegExp(t+"(?:-[\\d.]+/|(?: for [\\w-]+)?[ /-])([\\d.]+[^ ();/_-]*)","i").exec(e)||0)[1]||null}))}function toStringPlatform(){return this.description||""}M&&(M=[M]);if(/\bAndroid\b/.test(D)&&!R&&(O=/\bAndroid[^;]*;(.*?)(?:Build|\) AppleWebKit)\b/i.exec(e))){R=trim(O[1]).replace(/^[a-z]{2}-[a-z]{2};\s*/i,"")||null}if(N&&!R){R=getProduct([N])}else if(N&&R){R=R.replace(RegExp("^("+qualify(N)+")[-_.\\s]","i"),N+" ").replace(RegExp("^("+qualify(N)+")[-_.]?(\\w)","i"),N+" $2")}if(O=/\bGoogle TV\b/.exec(R)){R=O[0]}if(/\bSimulator\b/i.test(e)){R=(R?R+" ":"")+"Simulator"}if(A=="Opera Mini"&&/\bOPiOS\b/.test(e)){S.push("running in Turbo/Uncompressed mode")}if(A=="IE"&&/\blike iPhone OS\b/.test(e)){O=parse(e.replace(/like iPhone OS/,""));N=O.manufacturer;R=O.product}else if(/^iP/.test(R)){A||(A="Safari");D="iOS"+((O=/ OS ([\d_]+)/i.exec(e))?" "+O[1].replace(/_/g,"."):"")}else if(A=="Konqueror"&&/^Linux\b/i.test(D)){D="Kubuntu"}else if(N&&N!="Google"&&(/Chrome/.test(A)&&!/\bMobile Safari\b/i.test(e)||/\bVita\b/.test(R))||/\bAndroid\b/.test(D)&&/^Chrome/.test(A)&&/\bVersion\//i.test(e)){A="Android Browser";D=/\bAndroid\b/.test(D)?D:"Android"}else if(A=="Silk"){if(!/\bMobi/i.test(e)){D="Android";S.unshift("desktop mode")}if(/Accelerated *= *true/i.test(e)){S.unshift("accelerated")}}else if(A=="UC Browser"&&/\bUCWEB\b/.test(e)){S.push("speed mode")}else if(A=="PaleMoon"&&(O=/\bFirefox\/([\d.]+)\b/.exec(e))){S.push("identifying as Firefox "+O[1])}else if(A=="Firefox"&&(O=/\b(Mobile|Tablet|TV)\b/i.exec(e))){D||(D="Firefox OS");R||(R=O[1])}else if(!A||(O=!/\bMinefield\b/i.test(e)&&/\b(?:Firefox|Safari)\b/.exec(A))){if(A&&!R&&/[\/,]|^[^(]+?\)/.test(e.slice(e.indexOf(O+"/")+8))){A=null}if((O=R||N||D)&&(R||N||/\b(?:Android|Symbian OS|Tablet OS|webOS)\b/.test(D))){A=/[a-z]+(?: Hat)?/i.exec(/\bAndroid\b/.test(D)?D:O)+" Browser"}}else if(A=="Electron"&&(O=(/\bChrome\/([\d.]+)\b/.exec(e)||0)[1])){S.push("Chromium "+O)}if(!C){C=getVersion(["(?:Cloud9|CriOS|CrMo|Edge|Edg|EdgA|EdgiOS|FxiOS|HeadlessChrome|IEMobile|Iron|Opera ?Mini|OPiOS|OPR|Raven|SamsungBrowser|Silk(?!/[\\d.]+$)|UCBrowser|YaBrowser)","Version",qualify(A),"(?:Firefox|Minefield|NetFront)"])}if(O=M=="iCab"&&parseFloat(C)>3&&"WebKit"||/\bOpera\b/.test(A)&&(/\bOPR\b/.test(e)?"Blink":"Presto")||/\b(?:Midori|Nook|Safari)\b/i.test(e)&&!/^(?:Trident|EdgeHTML)$/.test(M)&&"WebKit"||!M&&/\bMSIE\b/i.test(e)&&(D=="Mac OS"?"Tasman":"Trident")||M=="WebKit"&&/\bPlayStation\b(?! Vita\b)/i.test(A)&&"NetFront"){M=[O]}if(A=="IE"&&(O=(/; *(?:XBLWP|ZuneWP)(\d+)/i.exec(e)||0)[1])){A+=" Mobile";D="Windows Phone "+(/\+$/.test(O)?O:O+".x");S.unshift("desktop mode")}else if(/\bWPDesktop\b/i.test(e)){A="IE Mobile";D="Windows Phone 8.x";S.unshift("desktop mode");C||(C=(/\brv:([\d.]+)/.exec(e)||0)[1])}else if(A!="IE"&&M=="Trident"&&(O=/\brv:([\d.]+)/.exec(e))){if(A){S.push("identifying as "+A+(C?" "+C:""))}A="IE";C=O[1]}if(T){if(isHostType(n,"global")){if(h){O=h.lang.System;j=O.getProperty("os.arch");D=D||O.getProperty("os.name")+" "+O.getProperty("os.version")}if(g){try{C=n.require("ringo/engine").version.join(".");A="RingoJS"}catch(e){if((O=n.system)&&O.global.system==n.system){A="Narwhal";D||(D=O[0].os||null)}}if(!A){A="Rhino"}}else if(typeof n.process=="object"&&!n.process.browser&&(O=n.process)){if(typeof O.versions=="object"){if(typeof O.versions.electron=="string"){S.push("Node "+O.versions.node);A="Electron";C=O.versions.electron}else if(typeof O.versions.nw=="string"){S.push("Chromium "+C,"Node "+O.versions.node);A="NW.js";C=O.versions.nw}}if(!A){A="Node.js";j=O.arch;D=O.platform;C=/[\d.]+/.exec(O.version);C=C?C[0]:null}}}else if(getClassOf(O=n.runtime)==d){A="Adobe AIR";D=O.flash.system.Capabilities.os}else if(getClassOf(O=n.phantom)==m){A="PhantomJS";C=(O=O.version||null)&&O.major+"."+O.minor+"."+O.patch}else if(typeof w.documentMode=="number"&&(O=/\bTrident\/(\d+)/i.exec(e))){C=[C,w.documentMode];if((O=+O[1]+4)!=C[1]){S.push("IE "+C[1]+" mode");M&&(M[1]="");C[1]=O}C=A=="IE"?String(C[1].toFixed(1)):C[0]}else if(typeof w.documentMode=="number"&&/^(?:Chrome|Firefox)\b/.test(A)){S.push("masking as "+A+" "+C);A="IE";C="11.0";M=["Trident"];D="Windows"}D=D&&format(D)}if(C&&(O=/(?:[ab]|dp|pre|[ab]\d+pre)(?:\d+\+?)?$/i.exec(C)||/(?:alpha|beta)(?: ?\d)?/i.exec(e+";"+(T&&i.appMinorVersion))||/\bMinefield\b/i.test(e)&&"a")){k=/b/i.test(O)?"beta":"alpha";C=C.replace(RegExp(O+"\\+?$"),"")+(k=="beta"?x:y)+(/\d+\+?/.exec(O)||"")}if(A=="Fennec"||A=="Firefox"&&/\b(?:Android|Firefox OS|KaiOS)\b/.test(D)){A="Firefox Mobile"}else if(A=="Maxthon"&&C){C=C.replace(/\.[\d.]+/,".x")}else if(/\bXbox\b/i.test(R)){if(R=="Xbox 360"){D=null}if(R=="Xbox 360"&&/\bIEMobile\b/.test(e)){S.unshift("mobile mode")}}else if((/^(?:Chrome|IE|Opera)$/.test(A)||A&&!R&&!/Browser|Mobi/.test(A))&&(D=="Windows CE"||/Mobi/i.test(e))){A+=" Mobile"}else if(A=="IE"&&T){try{if(n.external===null){S.unshift("platform preview")}}catch(e){S.unshift("embedded")}}else if((/\bBlackBerry\b/.test(R)||/\bBB10\b/.test(e))&&(O=(RegExp(R.replace(/ +/g," *")+"/([.\\d]+)","i").exec(e)||0)[1]||C)){O=[O,/BB10/.test(e)];D=(O[1]?(R=null,N="BlackBerry"):"Device Software")+" "+O[0];C=null}else if(this!=forOwn&&R!="Wii"&&(T&&_||/Opera/.test(A)&&/\b(?:MSIE|Firefox)\b/i.test(e)||A=="Firefox"&&/\bOS X (?:\d+\.){2,}/.test(D)||A=="IE"&&(D&&!/^Win/.test(D)&&C>5.5||/\bWindows XP\b/.test(D)&&C>8||C==8&&!/\bTrident\b/.test(e)))&&!u.test(O=parse.call(forOwn,e.replace(u,"")+";"))&&O.name){O="ing as "+O.name+((O=O.version)?" "+O:"");if(u.test(A)){if(/\bIE\b/.test(O)&&D=="Mac OS"){D=null}O="identify"+O}else{O="mask"+O;if(E){A=format(E.replace(/([a-z])([A-Z])/g,"$1 $2"))}else{A="Opera"}if(/\bIE\b/.test(O)){D=null}if(!T){C=null}}M=["Presto"];S.push(O)}if(O=(/\bAppleWebKit\/([\d.]+\+?)/i.exec(e)||0)[1]){O=[parseFloat(O.replace(/\.(\d)$/,".0$1")),O];if(A=="Safari"&&O[1].slice(-1)=="+"){A="WebKit Nightly";k="alpha";C=O[1].slice(0,-1)}else if(C==O[1]||C==(O[2]=(/\bSafari\/([\d.]+\+?)/i.exec(e)||0)[1])){C=null}O[1]=(/\b(?:Headless)?Chrome\/([\d.]+)/i.exec(e)||0)[1];if(O[0]==537.36&&O[2]==537.36&&parseFloat(O[1])>=28&&M=="WebKit"){M=["Blink"]}if(!T||!s&&!O[1]){M&&(M[1]="like Safari");O=(O=O[0],O<400?1:O<500?2:O<526?3:O<533?4:O<534?"4+":O<535?5:O<537?6:O<538?7:O<601?8:O<602?9:O<604?10:O<606?11:O<608?12:"12")}else{M&&(M[1]="like Chrome");O=O[1]||(O=O[0],O<530?1:O<532?2:O<532.05?3:O<533?4:O<534.03?5:O<534.07?6:O<534.1?7:O<534.13?8:O<534.16?9:O<534.24?10:O<534.3?11:O<535.01?12:O<535.02?"13+":O<535.07?15:O<535.11?16:O<535.19?17:O<536.05?18:O<536.1?19:O<537.01?20:O<537.11?"21+":O<537.13?23:O<537.18?24:O<537.24?25:O<537.36?26:M!="Blink"?"27":"28")}M&&(M[1]+=" "+(O+=typeof O=="number"?".x":/[.+]/.test(O)?"":"+"));if(A=="Safari"&&(!C||parseInt(C)>45)){C=O}else if(A=="Chrome"&&/\bHeadlessChrome/i.test(e)){S.unshift("headless")}}if(A=="Opera"&&(O=/\bzbov|zvav$/.exec(D))){A+=" ";S.unshift("desktop mode");if(O=="zvav"){A+="Mini";C=null}else{A+="Mobile"}D=D.replace(RegExp(" *"+O+"$"),"")}else if(A=="Safari"&&/\bChrome\b/.exec(M&&M[1])){S.unshift("desktop mode");A="Chrome Mobile";C=null;if(/\bOS X\b/.test(D)){N="Apple";D="iOS 4.3+"}else{D=null}}else if(/\bSRWare Iron\b/.test(A)&&!C){C=getVersion("Chrome")}if(C&&C.indexOf(O=/[\d.]+$/.exec(D))==0&&e.indexOf("/"+O+"-")>-1){D=trim(D.replace(O,""))}if(D&&D.indexOf(A)!=-1&&!RegExp(A+" OS").test(D)){D=D.replace(RegExp(" *"+qualify(A)+" *"),"")}if(M&&!/\b(?:Avant|Nook)\b/.test(A)&&(/Browser|Lunascape|Maxthon/.test(A)||A!="Safari"&&/^iOS/.test(D)&&/\bSafari\b/.test(M[1])||/^(?:Adobe|Arora|Breach|Midori|Opera|Phantom|Rekonq|Rock|Samsung Internet|Sleipnir|SRWare Iron|Vivaldi|Web)/.test(A)&&M[1])){(O=M[M.length-1])&&S.push(O)}if(S.length){S=["("+S.join("; ")+")"]}if(N&&R&&R.indexOf(N)<0){S.push("on "+N)}if(R){S.push((/^on /.test(S[S.length-1])?"":"on ")+R)}if(D){O=/ ([\d.+]+)$/.exec(D);P=O&&D.charAt(D.length-O[0].length-1)=="/";D={architecture:32,family:O&&!P?D.replace(O[0],""):D,version:O?O[1]:null,toString:function(){var e=this.version;return this.family+(e&&!P?" "+e:"")+(this.architecture==64?" 64-bit":"")}}}if((O=/\b(?:AMD|IA|Win|WOW|x86_|x)64\b/i.exec(j))&&!/\bi686\b/i.test(j)){if(D){D.architecture=64;D.family=D.family.replace(RegExp(" *"+O),"")}if(A&&(/\bWOW64\b/i.test(e)||T&&/\w(?:86|32)$/.test(i.cpuClass||i.platform)&&!/\bWin64; x64\b/i.test(e))){S.unshift("32-bit")}}else if(D&&/^OS X/.test(D.family)&&A=="Chrome"&&parseFloat(C)>=39){D.architecture=64}e||(e=null);var z={};z.description=e;z.layout=M&&M[0];z.manufacturer=N;z.name=A;z.prerelease=k;z.product=R;z.ua=e;z.version=A&&C;z.os=D||{architecture:null,family:null,version:null,toString:function(){return"null"}};z.parse=parse;z.toString=toStringPlatform;if(z.version){S.unshift(C)}if(z.name){S.unshift(A)}if(D&&A&&!(D==String(D).split(" ")[0]&&(D==A.split(" ")[0]||R))){S.push(R?"("+D+")":"on "+D)}if(S.length){z.description=S.join(" ")}return z}var v=parse();if(i&&o){forOwn(v,(function(e,n){i[n]=e}))}else{r.platform=v}}).call(this)},922:function(e,n){"use strict";Object.defineProperty(n,"__esModule",{value:true});var t="<unknown>";function parse(e){var n=e.split("\n");return n.reduce((function(e,n){var t=parseChrome(n)||parseWinjs(n)||parseGecko(n)||parseNode(n)||parseJSC(n);if(t){e.push(t)}return e}),[])}var r=/^\s*at (.*?) ?\(((?:file|https?|blob|chrome-extension|native|eval|webpack|<anonymous>|\/|[a-z]:\\|\\\\).*?)(?::(\d+))?(?::(\d+))?\)?\s*$/i;var a=/\((\S*)(?::(\d+))(?::(\d+))\)/;function parseChrome(e){var n=r.exec(e);if(!n){return null}var i=n[2]&&n[2].indexOf("native")===0;var o=n[2]&&n[2].indexOf("eval")===0;var l=a.exec(n[2]);if(o&&l!=null){n[2]=l[1];n[3]=l[2];n[4]=l[3]}return{file:!i?n[2]:null,methodName:n[1]||t,arguments:i?[n[2]]:[],lineNumber:n[3]?+n[3]:null,column:n[4]?+n[4]:null}}var i=/^\s*at (?:((?:\[object object\])?.+) )?\(?((?:file|ms-appx|https?|webpack|blob):.*?):(\d+)(?::(\d+))?\)?\s*$/i;function parseWinjs(e){var n=i.exec(e);if(!n){return null}return{file:n[2],methodName:n[1]||t,arguments:[],lineNumber:+n[3],column:n[4]?+n[4]:null}}var o=/^\s*(.*?)(?:\((.*?)\))?(?:^|@)((?:file|https?|blob|chrome|webpack|resource|\[native).*?|[^@]*bundle)(?::(\d+))?(?::(\d+))?\s*$/i;var l=/(\S+) line (\d+)(?: > eval line \d+)* > eval/i;function parseGecko(e){var n=o.exec(e);if(!n){return null}var r=n[3]&&n[3].indexOf(" > eval")>-1;var a=l.exec(n[3]);if(r&&a!=null){n[3]=a[1];n[4]=a[2];n[5]=null}return{file:n[3],methodName:n[1]||t,arguments:n[2]?n[2].split(","):[],lineNumber:n[4]?+n[4]:null,column:n[5]?+n[5]:null}}var s=/^\s*(?:([^@]*)(?:\((.*?)\))?@)?(\S.*?):(\d+)(?::(\d+))?\s*$/i;function parseJSC(e){var n=s.exec(e);if(!n){return null}return{file:n[3],methodName:n[1]||t,arguments:[],lineNumber:+n[4],column:n[5]?+n[5]:null}}var u=/^\s*at (?:((?:\[object object\])?[^\\/]+(?: \[as \S+\])?) )?\(?(.*?):(\d+)(?::(\d+))?\)?\s*$/i;function parseNode(e){var n=u.exec(e);if(!n){return null}return{file:n[2],methodName:n[1]||t,arguments:[],lineNumber:+n[3],column:n[4]?+n[4]:null}}n.parse=parse},977:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.onRefresh=n.unregister=n.register=n.onFullRefreshNeeded=n.onBuildError=n.onBuildOk=n.ReactDevOverlay=n.getNodeError=n.getErrorByType=void 0;var o=i(t(591));var l=t(525);var s=false;var u=undefined;function onUnhandledError(e){var n=e===null||e===void 0?void 0:e.error;if(!n||!(n instanceof Error)||typeof n.stack!=="string"){return}var t=n;o.emit({type:o.TYPE_UNHANDLED_ERROR,reason:n,frames:(0,l.parseStack)(t.stack)})}function onUnhandledRejection(e){var n=e===null||e===void 0?void 0:e.reason;if(!n||!(n instanceof Error)||typeof n.stack!=="string"){return}var t=n;o.emit({type:o.TYPE_UNHANDLED_REJECTION,reason:n,frames:(0,l.parseStack)(t.stack)})}function register(){if(s){return}s=true;try{var e=Error.stackTraceLimit;Error.stackTraceLimit=50;u=e}catch(e){}window.addEventListener("error",onUnhandledError);window.addEventListener("unhandledrejection",onUnhandledRejection)}n.register=register;function unregister(){if(!s){return}s=false;if(u!==undefined){try{Error.stackTraceLimit=u}catch(e){}u=undefined}window.removeEventListener("error",onUnhandledError);window.removeEventListener("unhandledrejection",onUnhandledRejection)}n.unregister=unregister;function onBuildOk(){o.emit({type:o.TYPE_BUILD_OK})}n.onBuildOk=onBuildOk;function onBuildError(e){o.emit({type:o.TYPE_BUILD_ERROR,message:e})}n.onBuildError=onBuildError;function onFullRefreshNeeded(e){o.emit({type:o.TYPE_FULL_REFRESH_NEEDED,reason:e!==null&&e!==void 0?e:null})}n.onFullRefreshNeeded=onFullRefreshNeeded;function onRefresh(){o.emit({type:o.TYPE_REFRESH})}n.onRefresh=onRefresh;var c=t(758);r(n,c,"getErrorByType");var f=t(915);r(n,f,"getNodeError");var d=t(943);r(n,d,"default","ReactDevOverlay")},198:function(e,n,t){"use strict";var r=this&&this.__extends||function(){var extendStatics=function(e,n){extendStatics=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,n){e.__proto__=n}||function(e,n){for(var t in n)if(Object.prototype.hasOwnProperty.call(n,t))e[t]=n[t]};return extendStatics(e,n)};return function(e,n){if(typeof n!=="function"&&n!==null)throw new TypeError("Class extends value "+String(n)+" is not a constructor or null");extendStatics(e,n);function __(){this.constructor=e}e.prototype=n===null?Object.create(n):(__.prototype=n.prototype,new __)}}();var a=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.ErrorBoundary=void 0;var i=a(t(522));var o=function(e){r(ErrorBoundary,e);function ErrorBoundary(){var n=e!==null&&e.apply(this,arguments)||this;n.state={error:null};return n}ErrorBoundary.prototype.componentDidCatch=function(e,n){this.props.onError(e,(n===null||n===void 0?void 0:n.componentStack)||null);this.setState({error:e})};ErrorBoundary.prototype.render=function(){return this.state.error?null:this.props.children};return ErrorBoundary}(i["default"].PureComponent);n.ErrorBoundary=o},943:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};var s=this&&this.__spreadArray||function(e,n,t){if(t||arguments.length===2)for(var r=0,a=n.length,i;r<a;r++){if(i||!(r in n)){if(!i)i=Array.prototype.slice.call(n,0,r);i[r]=n[r]}}return e.concat(i||Array.prototype.slice.call(n))};n.__esModule=true;var u=o(t(522));var c=o(t(591));var f=t(395);var d=t(71);var p=t(610);var v=t(710);var b=t(198);var m=t(796);var h=t(174);var g=t(803);function reducer(e,n){switch(n.type){case c.TYPE_BUILD_OK:{return r(r({},e),{buildError:null})}case c.TYPE_BUILD_ERROR:{return r(r({},e),{buildError:n.message})}case c.TYPE_FULL_REFRESH_NEEDED:{var t=r(r({},e),{fullRefreshReason:null,isAboutToFullRefresh:true});if(n.reason===null){return t}return r(r({},t),{fullRefreshReason:n.reason})}case c.TYPE_REFRESH:{return r(r({},e),{buildError:null,errors:[]})}case c.TYPE_UNHANDLED_ERROR:case c.TYPE_UNHANDLED_REJECTION:{return r(r({},e),{nextId:e.nextId+1,errors:s(s([],l(e.errors),false),[{id:e.nextId,event:n}],false)})}default:{var a=n;return e}}}var y=function ReactDevOverlay(e){var n=e.children,t=e.preventDisplay;var r=l(u.useReducer(reducer,{nextId:1,buildError:null,errors:[],fullRefreshReason:null,isAboutToFullRefresh:false}),2),a=r[0],i=r[1];u.useEffect((function(){c.on(i);return function(){c.off(i)}}),[i]);var o=u.useCallback((function(e,n){}),[]);var s=a.buildError!=null;var y=Boolean(a.errors.length);var x=a.isAboutToFullRefresh;var w=s||y||x;return u.createElement(u.Fragment,null,u.createElement(b.ErrorBoundary,{onError:o},n!==null&&n!==void 0?n:null),w?u.createElement(f.ShadowPortal,null,u.createElement(g.CssReset,null),u.createElement(m.Base,null),u.createElement(h.ComponentStyles,null),shouldPreventDisplay(s?"build":y?"runtime":x?"full-refresh":null,t)?null:s?u.createElement(d.BuildError,{message:a.buildError}):y?u.createElement(p.Errors,{errors:a.errors}):x?u.createElement(v.FullRefreshWarning,{reason:a.fullRefreshReason}):undefined):undefined)};var shouldPreventDisplay=function(e,n){if(!n||!e){return false}return n.includes(e)};n["default"]=y},591:function(e,n){"use strict";var t=this&&this.__assign||function(){t=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return t.apply(this,arguments)};n.__esModule=true;n.off=n.on=n.emit=n.TYPE_UNHANDLED_REJECTION=n.TYPE_UNHANDLED_ERROR=n.TYPE_REFRESH=n.TYPE_FULL_REFRESH_NEEDED=n.TYPE_BUILD_ERROR=n.TYPE_BUILD_OK=void 0;n.TYPE_BUILD_OK="build-ok";n.TYPE_BUILD_ERROR="build-error";n.TYPE_FULL_REFRESH_NEEDED="full-refresh-needed";n.TYPE_REFRESH="fast-refresh";n.TYPE_UNHANDLED_ERROR="unhandled-error";n.TYPE_UNHANDLED_REJECTION="unhandled-rejection";var r=new Set;var a=[];function drain(){setTimeout((function(){var _loop_1=function(){var e=a.shift();r.forEach((function(n){return n(e)}))};while(Boolean(a.length)&&Boolean(r.size)){_loop_1()}}),1)}function emit(e){a.push(Object.freeze(t({},e)));drain()}n.emit=emit;function on(e){if(r.has(e)){return false}r.add(e);drain();return true}n.on=on;function off(e){if(r.has(e)){r["delete"](e);return true}return false}n.off=off},7:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.CodeFrame=void 0;var s=l(t(194));var u=o(t(522));var c=l(t(518));var f=t(392);var d=function CodeFrame(e){var n=e.stackFrame,t=e.codeFrame;var a=u.useMemo((function(){var e=t.split(/\r?\n/g);var n=e.map((function(e){return/^>? +\d+ +\| ( *)/.exec((0,c["default"])(e))})).filter(Boolean).map((function(e){return e.pop()})).reduce((function(e,n){return isNaN(e)?n.length:Math.min(e,n.length)}),NaN);if(n>1){var r=" ".repeat(n);return e.map((function(e,n){return~(n=e.indexOf("|"))?e.substring(0,n)+e.substring(n).replace(r,""):e})).join("\n")}return e.join("\n")}),[t]);var i=u.useMemo((function(){return s["default"].ansiToJson(a,{json:true,use_classes:true,remove_empty:true})}),[a]);var o=u.useCallback((function(){var e;var t=new URLSearchParams;for(var r in n){t.append(r,((e=n[r])!==null&&e!==void 0?e:"").toString())}self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_launch-editor?"+t.toString()).then((function(){}),(function(){console.error("There was an issue opening this code in your editor.")}))}),[n]);return u.createElement("div",{"data-nextjs-codeframe":true},u.createElement("div",null,u.createElement("p",{role:"link",onClick:o,tabIndex:1,title:"Click to open in your editor"},u.createElement("span",null,(0,f.getFrameSource)(n)," @ ",n.methodName),u.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},u.createElement("path",{d:"M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"}),u.createElement("polyline",{points:"15 3 21 3 21 9"}),u.createElement("line",{x1:"10",y1:"14",x2:"21",y2:"3"})))),u.createElement("pre",null,i.map((function(e,n){return u.createElement("span",{key:"frame-"+n,style:r({color:e.fg?"var(--color-"+e.fg+")":undefined},e.decoration==="bold"?{fontWeight:800}:e.decoration==="italic"?{fontStyle:"italic"}:undefined)},e.content)}))))};n.CodeFrame=d},983:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.CodeFrame=void 0;var a=t(7);r(n,a,"CodeFrame")},761:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-codeframe] {\n    overflow: auto;\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-codeframe]::selection,\n  [data-nextjs-codeframe] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-codeframe] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n\n  [data-nextjs-codeframe] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  [data-nextjs-codeframe] > div {\n    display: inline-block;\n    width: auto;\n    min-width: 100%;\n    border-bottom: 1px solid var(--color-ansi-bright-black);\n  }\n  [data-nextjs-codeframe] > div > p {\n    display: flex;\n    align-items: center;\n    justify-content: space-between;\n    cursor: pointer;\n    margin: 0;\n  }\n  [data-nextjs-codeframe] > div > p:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-codeframe] div > p > svg {\n    width: auto;\n    height: 1em;\n    margin-left: 8px;\n  }\n  [data-nextjs-codeframe] div > pre {\n    overflow: hidden;\n    display: inline-block;\n  }\n"],["\n  [data-nextjs-codeframe] {\n    overflow: auto;\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-codeframe]::selection,\n  [data-nextjs-codeframe] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-codeframe] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n\n  [data-nextjs-codeframe] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  [data-nextjs-codeframe] > div {\n    display: inline-block;\n    width: auto;\n    min-width: 100%;\n    border-bottom: 1px solid var(--color-ansi-bright-black);\n  }\n  [data-nextjs-codeframe] > div > p {\n    display: flex;\n    align-items: center;\n    justify-content: space-between;\n    cursor: pointer;\n    margin: 0;\n  }\n  [data-nextjs-codeframe] > div > p:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-codeframe] div > p > svg {\n    width: auto;\n    height: 1em;\n    margin-left: 8px;\n  }\n  [data-nextjs-codeframe] div > pre {\n    overflow: hidden;\n    display: inline-block;\n  }\n"])));n.styles=i;var o},870:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__rest||function(e,n){var t={};for(var r in e)if(Object.prototype.hasOwnProperty.call(e,r)&&n.indexOf(r)<0)t[r]=e[r];if(e!=null&&typeof Object.getOwnPropertySymbols==="function")for(var a=0,r=Object.getOwnPropertySymbols(e);a<r.length;a++){if(n.indexOf(r[a])<0&&Object.prototype.propertyIsEnumerable.call(e,r[a]))t[r[a]]=e[r[a]]}return t};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.Dialog=void 0;var s=i(t(522));var u=t(814);var c=function Dialog(e){var n=e.children,t=e.type,r=e.onClose,a=o(e,["children","type","onClose"]);var i=l(s.useState(null),2),c=i[0],f=i[1];var d=s.useCallback((function(e){f(e)}),[]);(0,u.useOnClickOutside)(c,r);s.useEffect((function(){if(c==null){return}var e=c.getRootNode();if(!(e instanceof ShadowRoot)){return}var n=e;function handler(e){var t=n.activeElement;if(e.key==="Enter"&&t instanceof HTMLElement&&t.getAttribute("role")==="link"){e.preventDefault();e.stopPropagation();t.click()}}n.addEventListener("keydown",handler);return function(){return n.removeEventListener("keydown",handler)}}),[c]);return s.createElement("div",{ref:d,"data-nextjs-dialog":true,tabIndex:-1,role:"dialog","aria-labelledby":a["aria-labelledby"],"aria-describedby":a["aria-describedby"],"aria-modal":"true"},s.createElement("div",{"data-nextjs-dialog-banner":true,className:"banner-"+t}),n)};n.Dialog=c},244:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogBody=void 0;var o=i(t(522));var l=function DialogBody(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-body":true,className:t},n)};n.DialogBody=l},519:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogContent=void 0;var o=i(t(522));var l=function DialogContent(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-content":true,className:t},n)};n.DialogContent=l},734:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.DialogHeader=void 0;var o=i(t(522));var l=function DialogHeader(e){var n=e.children,t=e.className;return o.createElement("div",{"data-nextjs-dialog-header":true,className:t},n)};n.DialogHeader=l},824:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.styles=n.DialogHeader=n.DialogContent=n.DialogBody=n.Dialog=void 0;var a=t(870);r(n,a,"Dialog");var i=t(244);r(n,i,"DialogBody");var o=t(519);r(n,o,"DialogContent");var l=t(734);r(n,l,"DialogHeader");var s=t(629);r(n,s,"styles")},629:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog] {\n    display: flex;\n    flex-direction: column;\n    width: 100%;\n    margin-right: auto;\n    margin-left: auto;\n    outline: none;\n    background: white;\n    border-radius: var(--size-gap);\n    box-shadow: 0 var(--size-gap-half) var(--size-gap-double)\n      rgba(0, 0, 0, 0.25);\n    max-height: calc(100% - 56px);\n    overflow-y: hidden;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      max-height: calc(100% - 15px);\n    }\n  }\n\n  @media (min-width: 576px) {\n    [data-nextjs-dialog] {\n      max-width: 540px;\n      box-shadow: 0 var(--size-gap) var(--size-gap-quad) rgba(0, 0, 0, 0.25);\n    }\n  }\n\n  @media (min-width: 768px) {\n    [data-nextjs-dialog] {\n      max-width: 720px;\n    }\n  }\n\n  @media (min-width: 992px) {\n    [data-nextjs-dialog] {\n      max-width: 960px;\n    }\n  }\n\n  [data-nextjs-dialog-banner] {\n    position: relative;\n  }\n  [data-nextjs-dialog-banner].banner-warning {\n    border-color: var(--color-ansi-yellow);\n  }\n  [data-nextjs-dialog-banner].banner-error {\n    border-color: var(--color-ansi-red);\n  }\n\n  [data-nextjs-dialog-banner]::after {\n    z-index: 2;\n    content: '';\n    position: absolute;\n    top: 0;\n    right: 0;\n    width: 100%;\n    /* banner width: */\n    border-top-width: var(--size-gap-half);\n    border-bottom-width: 0;\n    border-top-style: solid;\n    border-bottom-style: solid;\n    border-top-color: inherit;\n    border-bottom-color: transparent;\n  }\n\n  [data-nextjs-dialog-content] {\n    overflow-y: auto;\n    border: none;\n    margin: 0;\n    /* calc(padding + banner width offset) */\n    padding: calc(var(--size-gap-double) + var(--size-gap-half))\n      var(--size-gap-double);\n    height: 100%;\n    display: flex;\n    flex-direction: column;\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-header] {\n    flex-shrink: 0;\n    margin-bottom: var(--size-gap-double);\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-body] {\n    position: relative;\n    flex: 1 1 auto;\n  }\n"],["\n  [data-nextjs-dialog] {\n    display: flex;\n    flex-direction: column;\n    width: 100%;\n    margin-right: auto;\n    margin-left: auto;\n    outline: none;\n    background: white;\n    border-radius: var(--size-gap);\n    box-shadow: 0 var(--size-gap-half) var(--size-gap-double)\n      rgba(0, 0, 0, 0.25);\n    max-height: calc(100% - 56px);\n    overflow-y: hidden;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      max-height: calc(100% - 15px);\n    }\n  }\n\n  @media (min-width: 576px) {\n    [data-nextjs-dialog] {\n      max-width: 540px;\n      box-shadow: 0 var(--size-gap) var(--size-gap-quad) rgba(0, 0, 0, 0.25);\n    }\n  }\n\n  @media (min-width: 768px) {\n    [data-nextjs-dialog] {\n      max-width: 720px;\n    }\n  }\n\n  @media (min-width: 992px) {\n    [data-nextjs-dialog] {\n      max-width: 960px;\n    }\n  }\n\n  [data-nextjs-dialog-banner] {\n    position: relative;\n  }\n  [data-nextjs-dialog-banner].banner-warning {\n    border-color: var(--color-ansi-yellow);\n  }\n  [data-nextjs-dialog-banner].banner-error {\n    border-color: var(--color-ansi-red);\n  }\n\n  [data-nextjs-dialog-banner]::after {\n    z-index: 2;\n    content: '';\n    position: absolute;\n    top: 0;\n    right: 0;\n    width: 100%;\n    /* banner width: */\n    border-top-width: var(--size-gap-half);\n    border-bottom-width: 0;\n    border-top-style: solid;\n    border-bottom-style: solid;\n    border-top-color: inherit;\n    border-bottom-color: transparent;\n  }\n\n  [data-nextjs-dialog-content] {\n    overflow-y: auto;\n    border: none;\n    margin: 0;\n    /* calc(padding + banner width offset) */\n    padding: calc(var(--size-gap-double) + var(--size-gap-half))\n      var(--size-gap-double);\n    height: 100%;\n    display: flex;\n    flex-direction: column;\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-header] {\n    flex-shrink: 0;\n    margin-bottom: var(--size-gap-double);\n  }\n  [data-nextjs-dialog-content] > [data-nextjs-dialog-body] {\n    position: relative;\n    flex: 1 1 auto;\n  }\n"])));n.styles=i;var o},727:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.LeftRightDialogHeader=void 0;var l=i(t(522));var s=function LeftRightDialogHeader(e){var n=e.children,t=e.className,r=e.previous,a=e.next,i=e.close;var s=l.useRef(null);var u=l.useRef(null);var c=l.useRef(null);var f=o(l.useState(null),2),d=f[0],p=f[1];var v=l.useCallback((function(e){p(e)}),[]);l.useEffect((function(){if(d==null){return}var e=d.getRootNode();var n=self.document;function handler(n){if(n.key==="ArrowLeft"){n.stopPropagation();if(s.current){s.current.focus()}r&&r()}else if(n.key==="ArrowRight"){n.stopPropagation();if(u.current){u.current.focus()}a&&a()}else if(n.key==="Escape"){n.stopPropagation();if(e instanceof ShadowRoot){var t=e.activeElement;if(t&&t!==c.current&&t instanceof HTMLElement){t.blur();return}}if(i){i()}}}e.addEventListener("keydown",handler);if(e!==n){n.addEventListener("keydown",handler)}return function(){e.removeEventListener("keydown",handler);if(e!==n){n.removeEventListener("keydown",handler)}}}),[i,d,a,r]);l.useEffect((function(){if(d==null){return}var e=d.getRootNode();if(e instanceof ShadowRoot){var n=e.activeElement;if(r==null){if(s.current&&n===s.current){s.current.blur()}}else if(a==null){if(u.current&&n===u.current){u.current.blur()}}}}),[d,a,r]);return l.createElement("div",{"data-nextjs-dialog-left-right":true,className:t},l.createElement("nav",{ref:v},l.createElement("button",{ref:s,type:"button",disabled:r==null?true:undefined,"aria-disabled":r==null?true:undefined,onClick:r!==null&&r!==void 0?r:undefined},l.createElement("svg",{viewBox:"0 0 14 14",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M6.99996 1.16666L1.16663 6.99999L6.99996 12.8333M12.8333 6.99999H1.99996H12.8333Z",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"}))),l.createElement("button",{ref:u,type:"button",disabled:a==null?true:undefined,"aria-disabled":a==null?true:undefined,onClick:a!==null&&a!==void 0?a:undefined},l.createElement("svg",{viewBox:"0 0 14 14",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M6.99996 1.16666L12.8333 6.99999L6.99996 12.8333M1.16663 6.99999H12H1.16663Z",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"})))," ",n),i?l.createElement("button",{ref:c,type:"button",onClick:i,"aria-label":"Close"},l.createElement("span",{"aria-hidden":"true"},l.createElement("svg",{width:"24",height:"24",viewBox:"0 0 24 24",fill:"none",xmlns:"http://www.w3.org/2000/svg"},l.createElement("path",{d:"M18 6L6 18",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"}),l.createElement("path",{d:"M6 6L18 18",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"})))):null)};n.LeftRightDialogHeader=s},830:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.styles=n.LeftRightDialogHeader=void 0;var a=t(727);r(n,a,"LeftRightDialogHeader");var i=t(448);r(n,i,"styles")},448:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog-left-right] {\n    display: flex;\n    flex-direction: row;\n    align-content: center;\n    align-items: center;\n    justify-content: space-between;\n  }\n  [data-nextjs-dialog-left-right] > nav > button {\n    display: inline-flex;\n    align-items: center;\n    justify-content: center;\n\n    width: calc(var(--size-gap-double) + var(--size-gap));\n    height: calc(var(--size-gap-double) + var(--size-gap));\n    font-size: 0;\n    border: none;\n    background-color: rgba(255, 85, 85, 0.1);\n    color: var(--color-ansi-red);\n    cursor: pointer;\n    transition: background-color 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > nav > button > svg {\n    width: auto;\n    height: calc(var(--size-gap) + var(--size-gap-half));\n  }\n  [data-nextjs-dialog-left-right] > nav > button:hover {\n    background-color: rgba(255, 85, 85, 0.2);\n  }\n  [data-nextjs-dialog-left-right] > nav > button:disabled {\n    background-color: rgba(255, 85, 85, 0.1);\n    color: rgba(255, 85, 85, 0.4);\n    cursor: not-allowed;\n  }\n\n  [data-nextjs-dialog-left-right] > nav > button:first-of-type {\n    border-radius: var(--size-gap-half) 0 0 var(--size-gap-half);\n    margin-right: 1px;\n  }\n  [data-nextjs-dialog-left-right] > nav > button:last-of-type {\n    border-radius: 0 var(--size-gap-half) var(--size-gap-half) 0;\n  }\n\n  [data-nextjs-dialog-left-right] > button:last-of-type {\n    border: 0;\n    padding: 0;\n\n    background-color: transparent;\n    appearance: none;\n\n    opacity: 0.4;\n    transition: opacity 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > button:last-of-type:hover {\n    opacity: 0.7;\n  }\n"],["\n  [data-nextjs-dialog-left-right] {\n    display: flex;\n    flex-direction: row;\n    align-content: center;\n    align-items: center;\n    justify-content: space-between;\n  }\n  [data-nextjs-dialog-left-right] > nav > button {\n    display: inline-flex;\n    align-items: center;\n    justify-content: center;\n\n    width: calc(var(--size-gap-double) + var(--size-gap));\n    height: calc(var(--size-gap-double) + var(--size-gap));\n    font-size: 0;\n    border: none;\n    background-color: rgba(255, 85, 85, 0.1);\n    color: var(--color-ansi-red);\n    cursor: pointer;\n    transition: background-color 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > nav > button > svg {\n    width: auto;\n    height: calc(var(--size-gap) + var(--size-gap-half));\n  }\n  [data-nextjs-dialog-left-right] > nav > button:hover {\n    background-color: rgba(255, 85, 85, 0.2);\n  }\n  [data-nextjs-dialog-left-right] > nav > button:disabled {\n    background-color: rgba(255, 85, 85, 0.1);\n    color: rgba(255, 85, 85, 0.4);\n    cursor: not-allowed;\n  }\n\n  [data-nextjs-dialog-left-right] > nav > button:first-of-type {\n    border-radius: var(--size-gap-half) 0 0 var(--size-gap-half);\n    margin-right: 1px;\n  }\n  [data-nextjs-dialog-left-right] > nav > button:last-of-type {\n    border-radius: 0 var(--size-gap-half) var(--size-gap-half) 0;\n  }\n\n  [data-nextjs-dialog-left-right] > button:last-of-type {\n    border: 0;\n    padding: 0;\n\n    background-color: transparent;\n    appearance: none;\n\n    opacity: 0.4;\n    transition: opacity 0.25s ease;\n  }\n  [data-nextjs-dialog-left-right] > button:last-of-type:hover {\n    opacity: 0.7;\n  }\n"])));n.styles=i;var o},434:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.Overlay=void 0;var s=l(t(426));var u=i(t(522));var c=t(253);var f=function Overlay(e){var n=e.className,t=e.children,r=e.fixed;u.useEffect((function(){(0,c.lock)();return function(){(0,c.unlock)()}}),[]);var a=o(u.useState(null),2),i=a[0],l=a[1];var f=u.useCallback((function(e){l(e)}),[]);u.useEffect((function(){if(i==null){return}var e=(0,s["default"])({context:i});return function(){e.disengage()}}),[i]);return u.createElement("div",{"data-nextjs-dialog-overlay":true,className:n,ref:f},u.createElement("div",{"data-nextjs-dialog-backdrop":true,"data-nextjs-dialog-backdrop-fixed":r?true:undefined}),t)};n.Overlay=f},253:function(e,n){"use strict";n.__esModule=true;n.unlock=n.lock=void 0;var t;var r;var a=0;function lock(){setTimeout((function(){if(a++>0){return}var e=window.innerWidth-document.documentElement.clientWidth;if(e>0){t=document.body.style.paddingRight;document.body.style.paddingRight=e+"px"}r=document.body.style.overflow;document.body.style.overflow="hidden"}))}n.lock=lock;function unlock(){setTimeout((function(){if(a===0||--a!==0){return}if(t!==undefined){document.body.style.paddingRight=t;t=undefined}if(r!==undefined){document.body.style.overflow=r;r=undefined}}))}n.unlock=unlock},382:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Overlay=void 0;var a=t(434);r(n,a,"Overlay")},426:function(e,n,t){"use strict";var r=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;var a=r(t(36));var i=r(t(730));function nodeArray(e){if(!e){return[]}if(Array.isArray(e)){return e}if(e.nodeType!==undefined){return[e]}if(typeof e==="string"){e=document.querySelectorAll(e)}if(e.length!==undefined){return[].slice.call(e,0)}throw new TypeError("unexpected input "+String(e))}function contextToElement(e){var n=e.context,t=e.label,r=t===undefined?"context-to-element":t,a=e.resolveDocument,i=e.defaultToDocument;var o=nodeArray(n)[0];if(a&&o&&o.nodeType===Node.DOCUMENT_NODE){o=o.documentElement}if(!o&&i){return document.documentElement}if(!o){throw new TypeError(r+" requires valid options.context")}if(o.nodeType!==Node.ELEMENT_NODE&&o.nodeType!==Node.DOCUMENT_FRAGMENT_NODE){throw new TypeError(r+" requires options.context to be an Element")}return o}function getShadowHost(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context;var t=contextToElement({label:"get/shadow-host",context:n});var r=null;while(t){r=t;t=t.parentNode}if(r.nodeType===r.DOCUMENT_FRAGMENT_NODE&&r.host){return r.host}return null}function getDocument(e){if(!e){return document}if(e.nodeType===Node.DOCUMENT_NODE){return e}return e.ownerDocument||document}function isActiveElement(e){var n=contextToElement({label:"is/active-element",resolveDocument:true,context:e});var t=getDocument(n);if(t.activeElement===n){return true}var r=getShadowHost({context:n});if(r&&r.shadowRoot.activeElement===n){return true}return false}function getParents(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context;var t=[];var r=contextToElement({label:"get/parents",context:n});while(r){t.push(r);r=r.parentNode;if(r&&r.nodeType!==Node.ELEMENT_NODE){r=null}}return t}var o=["matches","webkitMatchesSelector","mozMatchesSelector","msMatchesSelector"];var l=null;function findMethodName(e){o.some((function(n){if(!e[n]){return false}l=n;return true}))}function elementMatches(e,n){if(!l){findMethodName(e)}return e[l](n)}var s=JSON.parse(JSON.stringify(a["default"]));var u=s.os.family||"";var c=u==="Android";var f=u.slice(0,7)==="Windows";var d=u==="OS X";var p=u==="iOS";var v=s.layout==="Blink";var b=s.layout==="Gecko";var m=s.layout==="Trident";var h=s.layout==="EdgeHTML";var g=s.layout==="WebKit";var y=parseFloat(s.version);var x=Math.floor(y);s.majorVersion=x;s.is={ANDROID:c,WINDOWS:f,OSX:d,IOS:p,BLINK:v,GECKO:b,TRIDENT:m,EDGE:h,WEBKIT:g,IE9:m&&x===9,IE10:m&&x===10,IE11:m&&x===11};function before(){var e={activeElement:document.activeElement,windowScrollTop:window.scrollTop,windowScrollLeft:window.scrollLeft,bodyScrollTop:document.body.scrollTop,bodyScrollLeft:document.body.scrollLeft};var n=document.createElement("iframe");n.setAttribute("style","position:absolute; position:fixed; top:0; left:-2px; width:1px; height:1px; overflow:hidden;");n.setAttribute("aria-live","off");n.setAttribute("aria-busy","true");n.setAttribute("aria-hidden","true");document.body.appendChild(n);var t=n.contentWindow;var r=t.document;r.open();r.close();var a=r.createElement("div");r.body.appendChild(a);e.iframe=n;e.wrapper=a;e.window=t;e.document=r;return e}function test(e,n){e.wrapper.innerHTML="";var t=typeof n.element==="string"?e.document.createElement(n.element):n.element(e.wrapper,e.document);var r=n.mutate&&n.mutate(t,e.wrapper,e.document);if(!r&&r!==false){r=t}!t.parentNode&&e.wrapper.appendChild(t);r&&r.focus&&r.focus();return n.validate?n.validate(t,r,e.document):e.document.activeElement===r}function after(e){if(e.activeElement===document.body){document.activeElement&&document.activeElement.blur&&document.activeElement.blur();if(s.is.IE10){document.body.focus()}}else{e.activeElement&&e.activeElement.focus&&e.activeElement.focus()}document.body.removeChild(e.iframe);window.scrollTop=e.windowScrollTop;window.scrollLeft=e.windowScrollLeft;document.body.scrollTop=e.bodyScrollTop;document.body.scrollLeft=e.bodyScrollLeft}function detectFocus(e){var n=before();var t={};Object.keys(e).map((function(r){t[r]=test(n,e[r])}));after(n);return t}var w="1.4.1";function readLocalStorage(e){var n=void 0;try{n=window.localStorage&&window.localStorage.getItem(e);n=n?JSON.parse(n):{}}catch(e){n={}}return n}function writeLocalStorage(e,n){if(!document.hasFocus()){try{window.localStorage&&window.localStorage.removeItem(e)}catch(e){}return}try{window.localStorage&&window.localStorage.setItem(e,JSON.stringify(n))}catch(e){}}var _=typeof window!=="undefined"&&window.navigator.userAgent||"";var E="ally-supports-cache";var O=readLocalStorage(E);if(O.userAgent!==_||O.version!==w){O={}}O.userAgent=_;O.version=w;var j={get:function get(){return O},set:function set(e){Object.keys(e).forEach((function(n){O[n]=e[n]}));O.time=(new Date).toISOString();writeLocalStorage(E,O)}};function cssShadowPiercingDeepCombinator(){var e=void 0;try{document.querySelector("html >>> :first-child");e=">>>"}catch(n){try{document.querySelector("html /deep/ :first-child");e="/deep/"}catch(n){e=""}}return e}var S="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";var k={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test">'+'<area shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" tabindex="-1" alt="" src="'+S+'">';return e.querySelector("area")}};var T={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test">'+'<area href="#void" tabindex="-1" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" alt="" src="'+S+'">';return false},validate:function validate(e,n,t){if(s.is.GECKO){return true}var r=e.querySelector("area");r.focus();return t.activeElement===r}};var C={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-area-href-test">'+'<area shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-area-href-test" alt="" src="'+S+'">';return e.querySelector("area")},validate:function validate(e,n,t){if(s.is.GECKO){return true}return t.activeElement===n}};var P={name:"can-focus-audio-without-controls",element:"audio",mutate:function mutate(e){try{e.setAttribute("src",S)}catch(e){}}};var M="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ";var A={element:"div",mutate:function mutate(e){e.innerHTML='<map name="broken-image-map-test"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#broken-image-map-test" alt="" src="'+M+'">';return e.querySelector("area")}};var R={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","-1");e.setAttribute("style","display: -webkit-flex; display: -ms-flexbox; display: flex;");e.innerHTML='<span style="display: block;">hello</span>';return e.querySelector("span")}};var N={element:"fieldset",mutate:function mutate(e){e.setAttribute("tabindex",0);e.setAttribute("disabled","disabled")}};var D={element:"fieldset",mutate:function mutate(e){e.innerHTML="<legend>legend</legend><p>content</p>"}};var z={element:"span",mutate:function mutate(e){e.setAttribute("style","display: -webkit-flex; display: -ms-flexbox; display: flex;");e.innerHTML='<span style="display: block;">hello</span>'}};var F={element:"form",mutate:function mutate(e){e.setAttribute("tabindex",0);e.setAttribute("disabled","disabled")}};var I={element:"a",mutate:function mutate(e){e.href="#void";e.innerHTML='<img ismap src="'+S+'" alt="">';return e.querySelector("img")}};var L={element:"div",mutate:function mutate(e){e.innerHTML='<map name="image-map-tabindex-test"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#image-map-tabindex-test" tabindex="-1" alt="" '+'src="'+S+'">';return e.querySelector("img")}};var B={element:function element(e,n){var t=n.createElement("iframe");e.appendChild(t);var r=t.contentWindow.document;r.open();r.close();return t},mutate:function mutate(e){e.style.visibility="hidden";var n=e.contentWindow.document;var t=n.createElement("input");n.body.appendChild(t);return t},validate:function validate(e){var n=e.contentWindow.document;var t=n.querySelector("input");return n.activeElement===t}};var H=!s.is.WEBKIT;function focusInZeroDimensionObject(){return H}var W={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","invalid-value")}};var q={element:"label",mutate:function mutate(e){e.setAttribute("tabindex","-1")},validate:function validate(e,n,t){var r=e.offsetHeight;e.focus();return t.activeElement===e}};var U="data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHhtb"+"G5zOnhsaW5rPSJodHRwOi8vd3d3LnczLm9yZy8xOTk5L3hsaW5rIiBpZD0ic3ZnIj48dGV4dCB4PSIxMCIgeT0iMjAiIGlkPSJ"+"zdmctbGluay10ZXh0Ij50ZXh0PC90ZXh0Pjwvc3ZnPg==";var G={element:"object",mutate:function mutate(e){e.setAttribute("type","image/svg+xml");e.setAttribute("data",U);e.setAttribute("width","200");e.setAttribute("height","50");e.style.visibility="hidden"}};var V={name:"can-focus-object-svg",element:"object",mutate:function mutate(e){e.setAttribute("type","image/svg+xml");e.setAttribute("data",U);e.setAttribute("width","200");e.setAttribute("height","50")},validate:function validate(e,n,t){if(s.is.GECKO){return true}return t.activeElement===e}};var K=!s.is.IE9;function focusObjectSwf(){return K}var Z={element:"div",mutate:function mutate(e){e.innerHTML='<map name="focus-redirect-img-usemap"><area href="#void" shape="rect" coords="63,19,144,45"></map>'+'<img usemap="#focus-redirect-img-usemap" alt="" '+'src="'+S+'">';return e.querySelector("img")},validate:function validate(e,n,t){var r=e.querySelector("area");return t.activeElement===r}};var $={element:"fieldset",mutate:function mutate(e){e.innerHTML='<legend>legend</legend><input tabindex="-1"><input tabindex="0">';return false},validate:function validate(e,n,t){var r=e.querySelector('input[tabindex="-1"]');var a=e.querySelector('input[tabindex="0"]');e.focus();e.querySelector("legend").focus();return t.activeElement===r&&"focusable"||t.activeElement===a&&"tabbable"||""}};var Y={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px; overflow: auto;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>';return e.querySelector("div")}};var X={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>'}};var J={element:"div",mutate:function mutate(e){e.setAttribute("style","width: 100px; height: 50px; overflow: auto;");e.innerHTML='<div style="width: 500px; height: 40px;">scrollable content</div>'}};var Q={element:"details",mutate:function mutate(e){e.innerHTML="<summary>foo</summary><p>content</p>";return e.firstElementChild}};function makeFocusableForeignObject(){var e=document.createElement("div");e.innerHTML='<svg><foreignObject width="30" height="30">\n      <input type="text"/>\n  </foreignObject></svg>';return e.firstChild.firstChild}function focusSvgForeignObjectHack(e){var n=e.ownerSVGElement||e.nodeName.toLowerCase()==="svg";if(!n){return false}var t=makeFocusableForeignObject();e.appendChild(t);var r=t.querySelector("input");r.focus();r.disabled=true;e.removeChild(t);return true}function generate(e){return'<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">'+e+"</svg>"}function focus(e){if(e.focus){return}try{HTMLElement.prototype.focus.call(e)}catch(n){focusSvgForeignObjectHack(e)}}function validate(e,n,t){focus(n);return t.activeElement===n}var ee={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text focusable="true">a</text>');return e.querySelector("text")},validate:validate};var ne={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text tabindex="0">a</text>');return e.querySelector("text")},validate:validate};var te={element:"div",mutate:function mutate(e){e.innerHTML=generate('<text tabindex="-1">a</text>');return e.querySelector("text")},validate:validate};var re={element:"div",mutate:function mutate(e){e.innerHTML=generate(['<g id="ally-test-target"><a xlink:href="#void"><text>link</text></a></g>','<use xlink:href="#ally-test-target" x="0" y="0" tabindex="-1" />'].join(""));return e.querySelector("use")},validate:validate};var ae={element:"div",mutate:function mutate(e){e.innerHTML=generate('<foreignObject tabindex="-1"><input type="text" /></foreignObject>');return e.querySelector("foreignObject")||e.getElementsByTagName("foreignObject")[0]},validate:validate};var ie=Boolean(s.is.GECKO&&typeof SVGElement!=="undefined"&&SVGElement.prototype.focus);function focusSvgInIframe(){return ie}var oe={element:"div",mutate:function mutate(e){e.innerHTML=generate("");return e.firstChild},validate:validate};var le={element:"div",mutate:function mutate(e){e.setAttribute("tabindex","3x")}};var se={element:"table",mutate:function mutate(e,n,t){var r=t.createDocumentFragment();r.innerHTML="<tr><td>cell</td></tr>";e.appendChild(r)}};var ue={element:"video",mutate:function mutate(e){try{e.setAttribute("src",S)}catch(e){}}};var ce=s.is.GECKO||s.is.TRIDENT||s.is.EDGE;function tabsequenceAreaAtImgPosition(){return ce}var fe={cssShadowPiercingDeepCombinator:cssShadowPiercingDeepCombinator,focusInZeroDimensionObject:focusInZeroDimensionObject,focusObjectSwf:focusObjectSwf,focusSvgInIframe:focusSvgInIframe,tabsequenceAreaAtImgPosition:tabsequenceAreaAtImgPosition};var de={focusAreaImgTabindex:k,focusAreaTabindex:T,focusAreaWithoutHref:C,focusAudioWithoutControls:P,focusBrokenImageMap:A,focusChildrenOfFocusableFlexbox:R,focusFieldsetDisabled:N,focusFieldset:D,focusFlexboxContainer:z,focusFormDisabled:F,focusImgIsmap:I,focusImgUsemapTabindex:L,focusInHiddenIframe:B,focusInvalidTabindex:W,focusLabelTabindex:q,focusObjectSvg:V,focusObjectSvgHidden:G,focusRedirectImgUsemap:Z,focusRedirectLegend:$,focusScrollBody:Y,focusScrollContainerWithoutOverflow:X,focusScrollContainer:J,focusSummary:Q,focusSvgFocusableAttribute:ee,focusSvgTabindexAttribute:ne,focusSvgNegativeTabindexAttribute:te,focusSvgUseTabindex:re,focusSvgForeignobjectTabindex:ae,focusSvg:oe,focusTabindexTrailingCharacters:le,focusTable:se,focusVideoWithoutControls:ue};function executeTests(){var e=detectFocus(de);Object.keys(fe).forEach((function(n){e[n]=fe[n]()}));return e}var pe=null;function _supports(){if(pe){return pe}pe=j.get();if(!pe.time){j.set(executeTests());pe=j.get()}return pe}var ve=void 0;var be=/^\s*(-|\+)?[0-9]+\s*$/;var me=/^\s*(-|\+)?[0-9]+.*$/;function isValidTabindex(e){if(!ve){ve=_supports()}var n=ve.focusTabindexTrailingCharacters?me:be;var t=contextToElement({label:"is/valid-tabindex",resolveDocument:true,context:e});var r=t.hasAttribute("tabindex");var a=t.hasAttribute("tabIndex");if(!r&&!a){return false}var i=t.ownerSVGElement||t.nodeName.toLowerCase()==="svg";if(i&&!ve.focusSvgTabindexAttribute){return false}if(ve.focusInvalidTabindex){return true}var o=t.getAttribute(r?"tabindex":"tabIndex");if(o==="-32768"){return false}return Boolean(o&&n.test(o))}function tabindexValue(e){if(!isValidTabindex(e)){return null}var n=e.hasAttribute("tabindex");var t=n?"tabindex":"tabIndex";var r=parseInt(e.getAttribute(t),10);return isNaN(r)?-1:r}function isUserModifyWritable(e){var n=e.webkitUserModify||"";return Boolean(n&&n.indexOf("write")!==-1)}function hasCssOverflowScroll(e){return[e.getPropertyValue("overflow"),e.getPropertyValue("overflow-x"),e.getPropertyValue("overflow-y")].some((function(e){return e==="auto"||e==="scroll"}))}function hasCssDisplayFlex(e){return e.display.indexOf("flex")>-1}function isScrollableContainer(e,n,t,r){if(n!=="div"&&n!=="span"){return false}if(t&&t!=="div"&&t!=="span"&&!hasCssOverflowScroll(r)){return false}return e.offsetHeight<e.scrollHeight||e.offsetWidth<e.scrollWidth}var he=void 0;function isFocusRelevantRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{flexbox:false,scrollable:false,shadow:false}:t;if(!he){he=_supports()}var a=contextToElement({label:"is/focus-relevant",resolveDocument:true,context:n});if(!r.shadow&&a.shadowRoot){return true}var i=a.nodeName.toLowerCase();if(i==="input"&&a.type==="hidden"){return false}if(i==="input"||i==="select"||i==="button"||i==="textarea"){return true}if(i==="legend"&&he.focusRedirectLegend){return true}if(i==="label"){return true}if(i==="area"){return true}if(i==="a"&&a.hasAttribute("href")){return true}if(i==="object"&&a.hasAttribute("usemap")){return false}if(i==="object"){var o=a.getAttribute("type");if(!he.focusObjectSvg&&o==="image/svg+xml"){return false}else if(!he.focusObjectSwf&&o==="application/x-shockwave-flash"){return false}}if(i==="iframe"||i==="object"){return true}if(i==="embed"||i==="keygen"){return true}if(a.hasAttribute("contenteditable")){return true}if(i==="audio"&&(he.focusAudioWithoutControls||a.hasAttribute("controls"))){return true}if(i==="video"&&(he.focusVideoWithoutControls||a.hasAttribute("controls"))){return true}if(he.focusSummary&&i==="summary"){return true}var l=isValidTabindex(a);if(i==="img"&&a.hasAttribute("usemap")){return l&&he.focusImgUsemapTabindex||he.focusRedirectImgUsemap}if(he.focusTable&&(i==="table"||i==="td")){return true}if(he.focusFieldset&&i==="fieldset"){return true}var s=i==="svg";var u=a.ownerSVGElement;var c=a.getAttribute("focusable");var f=tabindexValue(a);if(i==="use"&&f!==null&&!he.focusSvgUseTabindex){return false}if(i==="foreignobject"){return f!==null&&he.focusSvgForeignobjectTabindex}if(elementMatches(a,"svg a")&&a.hasAttribute("xlink:href")){return true}if((s||u)&&a.focus&&!he.focusSvgNegativeTabindexAttribute&&f<0){return false}if(s){return l||he.focusSvg||he.focusSvgInIframe||Boolean(he.focusSvgFocusableAttribute&&c&&c==="true")}if(u){if(he.focusSvgTabindexAttribute&&l){return true}if(he.focusSvgFocusableAttribute){return c==="true"}}if(l){return true}var d=window.getComputedStyle(a,null);if(isUserModifyWritable(d)){return true}if(he.focusImgIsmap&&i==="img"&&a.hasAttribute("ismap")){var p=getParents({context:a}).some((function(e){return e.nodeName.toLowerCase()==="a"&&e.hasAttribute("href")}));if(p){return true}}if(!r.scrollable&&he.focusScrollContainer){if(he.focusScrollContainerWithoutOverflow){if(isScrollableContainer(a,i)){return true}}else if(hasCssOverflowScroll(d)){return true}}if(!r.flexbox&&he.focusFlexboxContainer&&hasCssDisplayFlex(d)){return true}var v=a.parentElement;if(!r.scrollable&&v){var b=v.nodeName.toLowerCase();var m=window.getComputedStyle(v,null);if(he.focusScrollBody&&isScrollableContainer(v,i,b,m)){return true}if(he.focusChildrenOfFocusableFlexbox){if(hasCssDisplayFlex(m)){return true}}}return false}isFocusRelevantRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isFocusRelevant(n){return isFocusRelevantRules({context:n,except:e})};n.rules=isFocusRelevantRules;return n};var ge=isFocusRelevantRules.except({});function findIndex(e,n){if(e.findIndex){return e.findIndex(n)}var t=e.length;if(t===0){return-1}for(var r=0;r<t;r++){if(n(e[r],r,e)){return r}}return-1}function getContentDocument(e){try{return e.contentDocument||e.contentWindow&&e.contentWindow.document||e.getSVGDocument&&e.getSVGDocument()||null}catch(e){return null}}function getWindow(e){var n=getDocument(e);return n.defaultView||window}var ye=void 0;function selectInShadows(e){if(typeof ye!=="string"){var n=cssShadowPiercingDeepCombinator();if(n){ye=", html "+n+" "}}if(!ye){return e}return e+ye+e.replace(/\s*,\s*/g,",").split(",").join(ye)}var xe=void 0;function findDocumentHostElement(e){if(!xe){xe=selectInShadows("object, iframe")}if(e._frameElement!==undefined){return e._frameElement}e._frameElement=null;var n=e.parent.document.querySelectorAll(xe);[].some.call(n,(function(n){var t=getContentDocument(n);if(t!==e.document){return false}e._frameElement=n;return true}));return e._frameElement}function getFrameElement(e){var n=getWindow(e);if(!n.parent||n.parent===n){return null}try{return n.frameElement||findDocumentHostElement(n)}catch(e){return null}}var we=/^(area)$/;function computedStyle(e,n){return window.getComputedStyle(e,null).getPropertyValue(n)}function notDisplayed(e){return e.some((function(e){return computedStyle(e,"display")==="none"}))}function notVisible(e){var n=findIndex(e,(function(e){var n=computedStyle(e,"visibility");return n==="hidden"||n==="collapse"}));if(n===-1){return false}var t=findIndex(e,(function(e){return computedStyle(e,"visibility")==="visible"}));if(t===-1){return true}if(n<t){return true}return false}function collapsedParent(e){var n=1;if(e[0].nodeName.toLowerCase()==="summary"){n=2}return e.slice(n).some((function(e){return e.nodeName.toLowerCase()==="details"&&e.open===false}))}function isVisibleRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{notRendered:false,cssDisplay:false,cssVisibility:false,detailsElement:false,browsingContext:false}:t;var a=contextToElement({label:"is/visible",resolveDocument:true,context:n});var i=a.nodeName.toLowerCase();if(!r.notRendered&&we.test(i)){return true}var o=getParents({context:a});var l=i==="audio"&&!a.hasAttribute("controls");if(!r.cssDisplay&&notDisplayed(l?o.slice(1):o)){return false}if(!r.cssVisibility&&notVisible(o)){return false}if(!r.detailsElement&&collapsedParent(o)){return false}if(!r.browsingContext){var s=getFrameElement(a);var u=isVisibleRules.except(r);if(s&&!u(s)){return false}}return true}isVisibleRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isVisible(n){return isVisibleRules({context:n,except:e})};n.rules=isVisibleRules;return n};var _e=isVisibleRules.except({});function getMapByName(e,n){var t=n.querySelector('map[name="'+(0,i["default"])(e)+'"]');return t||null}function getImageOfArea(e){var n=e.parentElement;if(!n.name||n.nodeName.toLowerCase()!=="map"){return null}var t=getDocument(e);return t.querySelector('img[usemap="#'+(0,i["default"])(n.name)+'"]')||null}var Ee=void 0;function isValidArea(e){if(!Ee){Ee=_supports()}var n=contextToElement({label:"is/valid-area",context:e});var t=n.nodeName.toLowerCase();if(t!=="area"){return false}var r=n.hasAttribute("tabindex");if(!Ee.focusAreaTabindex&&r){return false}var a=getImageOfArea(n);if(!a||!_e(a)){return false}if(!Ee.focusBrokenImageMap&&(!a.complete||!a.naturalHeight||a.offsetWidth<=0||a.offsetHeight<=0)){return false}if(!Ee.focusAreaWithoutHref&&!n.href){return Ee.focusAreaTabindex&&r||Ee.focusAreaImgTabindex&&a.hasAttribute("tabindex")}var i=getParents({context:a}).slice(1).some((function(e){var n=e.nodeName.toLowerCase();return n==="button"||n==="a"}));if(i){return false}return true}var Oe=void 0;var je=void 0;var Se={input:true,select:true,textarea:true,button:true,fieldset:true,form:true};function isNativeDisabledSupported(e){if(!Oe){Oe=_supports();if(Oe.focusFieldsetDisabled){delete Se.fieldset}if(Oe.focusFormDisabled){delete Se.form}je=new RegExp("^("+Object.keys(Se).join("|")+")$")}var n=contextToElement({label:"is/native-disabled-supported",context:e});var t=n.nodeName.toLowerCase();return Boolean(je.test(t))}var ke=void 0;function isDisabledFieldset(e){var n=e.nodeName.toLowerCase();return n==="fieldset"&&e.disabled}function isDisabledForm(e){var n=e.nodeName.toLowerCase();return n==="form"&&e.disabled}function isDisabled(e){if(!ke){ke=_supports()}var n=contextToElement({label:"is/disabled",context:e});if(n.hasAttribute("data-ally-disabled")){return true}if(!isNativeDisabledSupported(n)){return false}if(n.disabled){return true}var t=getParents({context:n});if(t.some(isDisabledFieldset)){return true}if(!ke.focusFormDisabled&&t.some(isDisabledForm)){return true}return false}function isOnlyTabbableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{onlyFocusableBrowsingContext:false,visible:false}:t;var a=contextToElement({label:"is/only-tabbable",resolveDocument:true,context:n});if(!r.visible&&!_e(a)){return false}if(!r.onlyFocusableBrowsingContext&&(s.is.GECKO||s.is.TRIDENT||s.is.EDGE)){var i=getFrameElement(a);if(i){if(tabindexValue(i)<0){return false}}}var o=a.nodeName.toLowerCase();var l=tabindexValue(a);if(o==="label"&&s.is.GECKO){return l!==null&&l>=0}if(s.is.GECKO&&a.ownerSVGElement&&!a.focus){if(o==="a"&&a.hasAttribute("xlink:href")){if(s.is.GECKO){return true}}}return false}isOnlyTabbableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isOnlyTabbable(n){return isOnlyTabbableRules({context:n,except:e})};n.rules=isOnlyTabbableRules;return n};var Te=isOnlyTabbableRules.except({});var Ce=void 0;function isOnlyFocusRelevant(e){var n=e.nodeName.toLowerCase();if(n==="embed"||n==="keygen"){return true}var t=tabindexValue(e);if(e.shadowRoot&&t===null){return true}if(n==="label"){return!Ce.focusLabelTabindex||t===null}if(n==="legend"){return t===null}if(Ce.focusSvgFocusableAttribute&&(e.ownerSVGElement||n==="svg")){var r=e.getAttribute("focusable");return r&&r==="false"}if(n==="img"&&e.hasAttribute("usemap")){return t===null||!Ce.focusImgUsemapTabindex}if(n==="area"){return!isValidArea(e)}return false}function isFocusableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{disabled:false,visible:false,onlyTabbable:false}:t;if(!Ce){Ce=_supports()}var a=Te.rules.except({onlyFocusableBrowsingContext:true,visible:r.visible});var i=contextToElement({label:"is/focusable",resolveDocument:true,context:n});var o=ge.rules({context:i,except:r});if(!o||isOnlyFocusRelevant(i)){return false}if(!r.disabled&&isDisabled(i)){return false}if(!r.onlyTabbable&&a(i)){return false}if(!r.visible){var l={context:i,except:{}};if(Ce.focusInHiddenIframe){l.except.browsingContext=true}if(Ce.focusObjectSvgHidden){var s=i.nodeName.toLowerCase();if(s==="object"){l.except.cssVisibility=true}}if(!_e.rules(l)){return false}}var u=getFrameElement(i);if(u){var c=u.nodeName.toLowerCase();if(c==="object"&&!Ce.focusInZeroDimensionObject){if(!u.offsetWidth||!u.offsetHeight){return false}}}var f=i.nodeName.toLowerCase();if(f==="svg"&&Ce.focusSvgInIframe&&!u&&i.getAttribute("tabindex")===null){return false}return true}isFocusableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isFocusable(n){return isFocusableRules({context:n,except:e})};n.rules=isFocusableRules;return n};var Pe=isFocusableRules.except({});function createFilter(e){var n=function filter(n){if(n.shadowRoot){return NodeFilter.FILTER_ACCEPT}if(e(n)){return NodeFilter.FILTER_ACCEPT}return NodeFilter.FILTER_SKIP};n.acceptNode=n;return n}var Me=createFilter(ge);function queryFocusableStrict(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;if(!n){n=document.documentElement}var i=Pe.rules.except({onlyTabbable:r});var o=getDocument(n);var l=o.createTreeWalker(n,NodeFilter.SHOW_ELEMENT,a==="all"?Me:createFilter(i),false);var s=[];while(l.nextNode()){if(l.currentNode.shadowRoot){if(i(l.currentNode)){s.push(l.currentNode)}s=s.concat(queryFocusableStrict({context:l.currentNode.shadowRoot,includeOnlyTabbable:r,strategy:a}))}else{s.push(l.currentNode)}}if(t){if(a==="all"){if(ge(n)){s.unshift(n)}}else if(i(n)){s.unshift(n)}}return s}var Ae=void 0;var Re=void 0;function selector$2(){if(!Ae){Ae=_supports()}if(typeof Re==="string"){return Re}Re=""+(Ae.focusTable?"table, td,":"")+(Ae.focusFieldset?"fieldset,":"")+"svg a,"+"a[href],"+"area[href],"+"input, select, textarea, button,"+"iframe, object, embed,"+"keygen,"+(Ae.focusAudioWithoutControls?"audio,":"audio[controls],")+(Ae.focusVideoWithoutControls?"video,":"video[controls],")+(Ae.focusSummary?"summary,":"")+"[tabindex],"+"[contenteditable]";Re=selectInShadows(Re);return Re}function queryFocusableQuick(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable;var a=selector$2();var i=n.querySelectorAll(a);var o=Pe.rules.except({onlyTabbable:r});var l=[].filter.call(i,o);if(t&&o(n)){l.unshift(n)}return l}function queryFocusable(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy,i=a===undefined?"quick":a;var o=contextToElement({label:"query/focusable",resolveDocument:true,defaultToDocument:true,context:n});var l={context:o,includeContext:t,includeOnlyTabbable:r,strategy:i};if(i==="quick"){return queryFocusableQuick(l)}else if(i==="strict"||i==="all"){return queryFocusableStrict(l)}throw new TypeError('query/focusable requires option.strategy to be one of ["quick", "strict", "all"]')}var Ne=void 0;var De=/^(fieldset|table|td|body)$/;function isTabbableRules(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.except,r=t===undefined?{flexbox:false,scrollable:false,shadow:false,visible:false,onlyTabbable:false}:t;if(!Ne){Ne=_supports()}var a=contextToElement({label:"is/tabbable",resolveDocument:true,context:n});if(s.is.BLINK&&s.is.ANDROID&&s.majorVersion>42){return false}var i=getFrameElement(a);if(i){if(s.is.WEBKIT&&s.is.IOS){return false}if(tabindexValue(i)<0){return false}if(!r.visible&&(s.is.BLINK||s.is.WEBKIT)&&!_e(i)){return false}var o=i.nodeName.toLowerCase();if(o==="object"){var l=s.name==="Chrome"&&s.majorVersion>=54||s.name==="Opera"&&s.majorVersion>=41;if(s.is.WEBKIT||s.is.BLINK&&!l){return false}}}var u=a.nodeName.toLowerCase();var c=tabindexValue(a);var f=c===null?null:c>=0;if(s.is.EDGE&&s.majorVersion>=14&&i&&a.ownerSVGElement&&c<0){return true}var d=f!==false;var p=c!==null&&c>=0;if(a.hasAttribute("contenteditable")){return d}if(De.test(u)&&f!==true){return false}if(s.is.WEBKIT&&s.is.IOS){var v=u==="input"&&a.type==="text"||a.type==="password"||u==="select"||u==="textarea"||a.hasAttribute("contenteditable");if(!v){var b=window.getComputedStyle(a,null);v=isUserModifyWritable(b)}if(!v){return false}}if(u==="use"&&c!==null){if(s.is.BLINK||s.is.WEBKIT&&s.majorVersion===9){return true}}if(elementMatches(a,"svg a")&&a.hasAttribute("xlink:href")){if(d){return true}if(a.focus&&!Ne.focusSvgNegativeTabindexAttribute){return true}}if(u==="svg"&&Ne.focusSvgInIframe&&d){return true}if(s.is.TRIDENT||s.is.EDGE){if(u==="svg"){if(Ne.focusSvg){return true}return a.hasAttribute("focusable")||p}if(a.ownerSVGElement){if(Ne.focusSvgTabindexAttribute&&p){return true}return a.hasAttribute("focusable")}}if(a.tabIndex===undefined){return Boolean(r.onlyTabbable)}if(u==="audio"){if(!a.hasAttribute("controls")){return false}else if(s.is.BLINK){return true}}if(u==="video"){if(!a.hasAttribute("controls")){if(s.is.TRIDENT||s.is.EDGE){return false}}else if(s.is.BLINK||s.is.GECKO){return true}}if(u==="object"){if(s.is.BLINK||s.is.WEBKIT){return false}}if(u==="iframe"){return false}if(!r.scrollable&&s.is.GECKO){var m=window.getComputedStyle(a,null);if(hasCssOverflowScroll(m)){return d}}if(s.is.TRIDENT||s.is.EDGE){if(u==="area"){var h=getImageOfArea(a);if(h&&tabindexValue(h)<0){return false}}var g=window.getComputedStyle(a,null);if(isUserModifyWritable(g)){return a.tabIndex>=0}if(!r.flexbox&&hasCssDisplayFlex(g)){if(c!==null){return p}return ze(a)&&Fe(a)}if(isScrollableContainer(a,u)){return false}var y=a.parentElement;if(y){var x=y.nodeName.toLowerCase();var w=window.getComputedStyle(y,null);if(isScrollableContainer(y,u,x,w)){return false}if(hasCssDisplayFlex(w)){return p}}}return a.tabIndex>=0}isTabbableRules.except=function(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n=function isTabbable(n){return isTabbableRules({context:n,except:e})};n.rules=isTabbableRules;return n};var ze=ge.rules.except({flexbox:true});var Fe=isTabbableRules.except({flexbox:true});var Ie=isTabbableRules.except({});function queryTabbable(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;var i=Ie.rules.except({onlyTabbable:r});return queryFocusable({context:n,includeContext:t,includeOnlyTabbable:r,strategy:a}).filter(i)}function compareDomPosition(e,n){return e.compareDocumentPosition(n)&Node.DOCUMENT_POSITION_FOLLOWING?-1:1}function sortDomOrder(e){return e.sort(compareDomPosition)}function getFirstSuccessorOffset(e,n){return findIndex(e,(function(e){return n.compareDocumentPosition(e)&Node.DOCUMENT_POSITION_FOLLOWING}))}function findInsertionOffsets(e,n,t){var r=[];n.forEach((function(n){var a=true;var i=e.indexOf(n);if(i===-1){i=getFirstSuccessorOffset(e,n);a=false}if(i===-1){i=e.length}var o=nodeArray(t?t(n):n);if(!o.length){return}r.push({offset:i,replace:a,elements:o})}));return r}function insertElementsAtOffsets(e,n){var t=0;n.sort((function(e,n){return e.offset-n.offset}));n.forEach((function(n){var r=n.replace?1:0;var a=[n.offset+t,r].concat(n.elements);e.splice.apply(e,a);t+=n.elements.length-r}))}function mergeInDomOrder(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.list,t=e.elements,r=e.resolveElement;var a=n.slice(0);var i=nodeArray(t).slice(0);sortDomOrder(i);var o=findInsertionOffsets(a,i,r);insertElementsAtOffsets(a,o);return a}var Le=function(){function defineProperties(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||false;r.configurable=true;if("value"in r)r.writable=true;Object.defineProperty(e,r.key,r)}}return function(e,n,t){if(n)defineProperties(e.prototype,n);if(t)defineProperties(e,t);return e}}();function _classCallCheck(e,n){if(!(e instanceof n)){throw new TypeError("Cannot call a class as a function")}}var Be=function(){function Maps(e){_classCallCheck(this,Maps);this._document=getDocument(e);this.maps={}}Le(Maps,[{key:"getAreasFor",value:function getAreasFor(e){if(!this.maps[e]){this.addMapByName(e)}return this.maps[e]}},{key:"addMapByName",value:function addMapByName(e){var n=getMapByName(e,this._document);if(!n){return}this.maps[n.name]=queryTabbable({context:n})}},{key:"extractAreasFromList",value:function extractAreasFromList(e){return e.filter((function(e){var n=e.nodeName.toLowerCase();if(n!=="area"){return true}var t=e.parentNode;if(!this.maps[t.name]){this.maps[t.name]=[]}this.maps[t.name].push(e);return false}),this)}}]);return Maps}();function sortArea(e,n){var t=n.querySelectorAll("img[usemap]");var r=new Be(n);var a=r.extractAreasFromList(e);if(!t.length){return a}return mergeInDomOrder({list:a,elements:t,resolveElement:function resolveElement(e){var n=e.getAttribute("usemap").slice(1);return r.getAreasFor(n)}})}var He=function(){function defineProperties(e,n){for(var t=0;t<n.length;t++){var r=n[t];r.enumerable=r.enumerable||false;r.configurable=true;if("value"in r)r.writable=true;Object.defineProperty(e,r.key,r)}}return function(e,n,t){if(n)defineProperties(e.prototype,n);if(t)defineProperties(e,t);return e}}();function _classCallCheck$1(e,n){if(!(e instanceof n)){throw new TypeError("Cannot call a class as a function")}}var We=function(){function Shadows(e,n){_classCallCheck$1(this,Shadows);this.context=e;this.sortElements=n;this.hostCounter=1;this.inHost={};this.inDocument=[];this.hosts={};this.elements={}}He(Shadows,[{key:"_registerHost",value:function _registerHost(e){if(e._sortingId){return}e._sortingId="shadow-"+this.hostCounter++;this.hosts[e._sortingId]=e;var n=getShadowHost({context:e});if(n){this._registerHost(n);this._registerHostParent(e,n)}else{this.inDocument.push(e)}}},{key:"_registerHostParent",value:function _registerHostParent(e,n){if(!this.inHost[n._sortingId]){this.inHost[n._sortingId]=[]}this.inHost[n._sortingId].push(e)}},{key:"_registerElement",value:function _registerElement(e,n){if(!this.elements[n._sortingId]){this.elements[n._sortingId]=[]}this.elements[n._sortingId].push(e)}},{key:"extractElements",value:function extractElements(e){return e.filter((function(e){var n=getShadowHost({context:e});if(!n){return true}this._registerHost(n);this._registerElement(e,n);return false}),this)}},{key:"sort",value:function sort(e){var n=this._injectHosts(e);n=this._replaceHosts(n);this._cleanup();return n}},{key:"_injectHosts",value:function _injectHosts(e){Object.keys(this.hosts).forEach((function(e){var n=this.elements[e];var t=this.inHost[e];var r=this.hosts[e].shadowRoot;this.elements[e]=this._merge(n,t,r)}),this);return this._merge(e,this.inDocument,this.context)}},{key:"_merge",value:function _merge(e,n,t){var r=mergeInDomOrder({list:e,elements:n});return this.sortElements(r,t)}},{key:"_replaceHosts",value:function _replaceHosts(e){return mergeInDomOrder({list:e,elements:this.inDocument,resolveElement:this._resolveHostElement.bind(this)})}},{key:"_resolveHostElement",value:function _resolveHostElement(e){var n=mergeInDomOrder({list:this.elements[e._sortingId],elements:this.inHost[e._sortingId],resolveElement:this._resolveHostElement.bind(this)});var t=tabindexValue(e);if(t!==null&&t>-1){return[e].concat(n)}return n}},{key:"_cleanup",value:function _cleanup(){Object.keys(this.hosts).forEach((function(e){delete this.hosts[e]._sortingId}),this)}}]);return Shadows}();function sortShadowed(e,n,t){var r=new We(n,t);var a=r.extractElements(e);if(a.length===e.length){return t(e)}return r.sort(a)}function sortTabindex(e){var n={};var t=[];var r=e.filter((function(e){var r=e.tabIndex;if(r===undefined){r=tabindexValue(e)}if(r<=0||r===null||r===undefined){return true}if(!n[r]){n[r]=[];t.push(r)}n[r].push(e);return false}));var a=t.sort().map((function(e){return n[e]})).reduceRight((function(e,n){return n.concat(e)}),r);return a}var qe=void 0;function moveContextToBeginning(e,n){var t=e.indexOf(n);if(t>0){var r=e.splice(t,1);return r.concat(e)}return e}function sortElements(e,n){if(qe.tabsequenceAreaAtImgPosition){e=sortArea(e,n)}e=sortTabindex(e);return e}function queryTabsequence(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.context,t=e.includeContext,r=e.includeOnlyTabbable,a=e.strategy;if(!qe){qe=_supports()}var i=nodeArray(n)[0]||document.documentElement;var o=queryTabbable({context:i,includeContext:t,includeOnlyTabbable:r,strategy:a});if(document.body.createShadowRoot&&s.is.BLINK){o=sortShadowed(o,i,sortElements)}else{o=sortElements(o,i)}if(t){o=moveContextToBeginning(o,i)}return o}var Ue={tab:9,left:37,up:38,right:39,down:40,pageUp:33,"page-up":33,pageDown:34,"page-down":34,end:35,home:36,enter:13,escape:27,space:32,shift:16,capsLock:20,"caps-lock":20,ctrl:17,alt:18,meta:91,pause:19,insert:45,delete:46,backspace:8,_alias:{91:[92,93,224]}};for(var Ge=1;Ge<26;Ge++){Ue["f"+Ge]=Ge+111}for(var Ve=0;Ve<10;Ve++){var Ke=Ve+48;var Ze=Ve+96;Ue[Ve]=Ke;Ue["num-"+Ve]=Ze;Ue._alias[Ke]=[Ze]}for(var $e=0;$e<26;$e++){var Ye=$e+65;var Xe=String.fromCharCode(Ye).toLowerCase();Ue[Xe]=Ye}var Je={alt:"altKey",ctrl:"ctrlKey",meta:"metaKey",shift:"shiftKey"};var Qe=Object.keys(Je).map((function(e){return Je[e]}));function createExpectedModifiers(e){var n=e?null:false;return{altKey:n,ctrlKey:n,metaKey:n,shiftKey:n}}function resolveModifiers(e){var n=e.indexOf("*")!==-1;var t=createExpectedModifiers(n);e.forEach((function(e){if(e==="*"){return}var n=true;var r=e.slice(0,1);if(r==="?"){n=null}else if(r==="!"){n=false}if(n!==true){e=e.slice(1)}var a=Je[e];if(!a){throw new TypeError('Unknown modifier "'+e+'"')}t[a]=n}));return t}function resolveKey(e){var n=Ue[e]||parseInt(e,10);if(!n||typeof n!=="number"||isNaN(n)){throw new TypeError('Unknown key "'+e+'"')}return[n].concat(Ue._alias[n]||[])}function matchModifiers(e,n){return!Qe.some((function(t){return typeof e[t]==="boolean"&&Boolean(n[t])!==e[t]}))}function keyBinding(e){return e.split(/\s+/).map((function(e){var n=e.split("+");var t=resolveModifiers(n.slice(0,-1));var r=resolveKey(n.slice(-1));return{keyCodes:r,modifiers:t,matchModifiers:matchModifiers.bind(null,t)}}))}function getParentComparator(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{},n=e.parent,t=e.element,r=e.includeSelf;if(n){return function isChildOf(e){return Boolean(r&&e===n||n.compareDocumentPosition(e)&Node.DOCUMENT_POSITION_CONTAINED_BY)}}else if(t){return function isParentOf(e){return Boolean(r&&t===e||e.compareDocumentPosition(t)&Node.DOCUMENT_POSITION_CONTAINED_BY)}}throw new TypeError("util/compare-position#getParentComparator required either options.parent or options.element")}function whenKey(){var e=arguments.length>0&&arguments[0]!==undefined?arguments[0]:{};var n={};var t=nodeArray(e.context)[0]||document.documentElement;delete e.context;var r=nodeArray(e.filter);delete e.filter;var a=Object.keys(e);if(!a.length){throw new TypeError("when/key requires at least one option key")}var i=function registerBinding(e){e.keyCodes.forEach((function(t){if(!n[t]){n[t]=[]}n[t].push(e)}))};a.forEach((function(n){if(typeof e[n]!=="function"){throw new TypeError('when/key requires option["'+n+'"] to be a function')}var t=function addCallback(t){t.callback=e[n];return t};keyBinding(n).map(t).forEach(i)}));var o=function handleKeyDown(e){if(e.defaultPrevented){return}if(r.length){var a=getParentComparator({element:e.target,includeSelf:true});if(r.some(a)){return}}var i=e.keyCode||e.which;if(!n[i]){return}n[i].forEach((function(n){if(!n.matchModifiers(e)){return}n.callback.call(t,e,l)}))};t.addEventListener("keydown",o,false);var l=function disengage(){t.removeEventListener("keydown",o,false)};return{disengage:l}}function default_1(e){var n=e===void 0?{}:e,t=n.context;if(!t){t=document.documentElement}queryTabsequence();return whenKey({"?alt+?shift+tab":function altShiftTab(e){e.preventDefault();var n=queryTabsequence({context:t});var r=e.shiftKey;var a=n[0];var i=n[n.length-1];var o=r?a:i;var l=r?i:a;if(isActiveElement(o)){l.focus();return}var s=void 0;var u=n.some((function(e,n){if(!isActiveElement(e)){return false}s=n;return true}));if(!u){a.focus();return}var c=r?-1:1;n[s+c].focus()}})}n["default"]=default_1},173:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-dialog-overlay] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    overflow: auto;\n    z-index: 9000;\n\n    display: flex;\n    align-content: center;\n    align-items: center;\n    flex-direction: column;\n    padding: 10vh 15px 0;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      padding: 15px 15px 0;\n    }\n  }\n\n  [data-nextjs-dialog-backdrop] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    background-color: rgba(17, 17, 17, 0.2);\n    pointer-events: all;\n    z-index: -1;\n  }\n\n  [data-nextjs-dialog-backdrop-fixed] {\n    cursor: not-allowed;\n    -webkit-backdrop-filter: blur(8px);\n    backdrop-filter: blur(8px);\n  }\n"],["\n  [data-nextjs-dialog-overlay] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    overflow: auto;\n    z-index: 9000;\n\n    display: flex;\n    align-content: center;\n    align-items: center;\n    flex-direction: column;\n    padding: 10vh 15px 0;\n  }\n\n  @media (max-height: 812px) {\n    [data-nextjs-dialog-overlay] {\n      padding: 15px 15px 0;\n    }\n  }\n\n  [data-nextjs-dialog-backdrop] {\n    position: fixed;\n    top: 0;\n    right: 0;\n    bottom: 0;\n    left: 0;\n    background-color: rgba(17, 17, 17, 0.2);\n    pointer-events: all;\n    z-index: -1;\n  }\n\n  [data-nextjs-dialog-backdrop-fixed] {\n    cursor: not-allowed;\n    -webkit-backdrop-filter: blur(8px);\n    backdrop-filter: blur(8px);\n  }\n"])));n.styles=i;var o},395:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};var o=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.ShadowPortal=void 0;var l=i(t(522));var s=t(255);var u=function Portal(e){var n=e.children;var t=l.useRef(null);var r=l.useRef(null);var a=l.useRef(null);var i=o(l.useState(),2),u=i[1];l.useLayoutEffect((function(){var e=t.current.ownerDocument;r.current=e.createElement("nextjs-portal");a.current=r.current.attachShadow({mode:"open"});e.body.appendChild(r.current);u({});return function(){if(r.current&&r.current.ownerDocument){r.current.ownerDocument.body.removeChild(r.current)}}}),[]);return a.current?(0,s.createPortal)(n,a.current):l.createElement("span",{ref:t})};n.ShadowPortal=u},261:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};n.__esModule=true;n.Terminal=void 0;var s=l(t(194));var u=o(t(522));var c=function Terminal(e){var n=e.content;var t=u.useMemo((function(){return s["default"].ansiToJson(n,{json:true,use_classes:true,remove_empty:true})}),[n]);return u.createElement("div",{"data-nextjs-terminal":true},u.createElement("pre",null,t.map((function(e,n){return u.createElement("span",{key:"terminal-entry-"+n,style:r({color:e.fg?"var(--color-"+e.fg+")":undefined},e.decoration==="bold"?{fontWeight:800}:e.decoration==="italic"?{fontStyle:"italic"}:undefined)},e.content)}))))};n.Terminal=c},504:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Terminal=void 0;var a=t(261);r(n,a,"Terminal")},492:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-terminal] {\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-terminal]::selection,\n  [data-nextjs-terminal] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-terminal] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n  [data-nextjs-terminal] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n\n  [data-nextjs-terminal] pre {\n    white-space: pre-wrap;\n    word-break: break-word;\n  }\n"],["\n  [data-nextjs-terminal] {\n    border-radius: var(--size-gap-half);\n    background-color: var(--color-ansi-bg);\n    color: var(--color-ansi-fg);\n  }\n  [data-nextjs-terminal]::selection,\n  [data-nextjs-terminal] *::selection {\n    background-color: var(--color-ansi-selection);\n  }\n  [data-nextjs-terminal] * {\n    color: inherit;\n    background-color: transparent;\n    font-family: var(--font-stack-monospace);\n  }\n  [data-nextjs-terminal] > * {\n    margin: 0;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n\n  [data-nextjs-terminal] pre {\n    white-space: pre-wrap;\n    word-break: break-word;\n  }\n"])));n.styles=i;var o},314:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.Toast=void 0;var o=i(t(522));var l=function Toast(e){var n=e.onClick,t=e.children,r=e.className;return o.createElement("div",{"data-nextjs-toast":true,onClick:n,className:r},o.createElement("div",{"data-nextjs-toast-wrapper":true},t))};n.Toast=l},626:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});n.__esModule=true;n.Toast=n.styles=void 0;var a=t(565);r(n,a,"styles");var i=t(314);r(n,i,"Toast")},565:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};n.__esModule=true;n.styles=void 0;var a=t(373);var i=(0,a.noop)(o||(o=r(["\n  [data-nextjs-toast] {\n    position: fixed;\n    bottom: var(--size-gap-double);\n    left: var(--size-gap-double);\n    max-width: 420px;\n    z-index: 9000;\n  }\n\n  @media (max-width: 440px) {\n    [data-nextjs-toast] {\n      max-width: 90vw;\n      left: 5vw;\n    }\n  }\n\n  [data-nextjs-toast-wrapper] {\n    padding: 16px;\n    border-radius: var(--size-gap-half);\n    font-weight: 500;\n    color: var(--color-ansi-bright-white);\n    background-color: var(--color-ansi-red);\n    box-shadow: 0px var(--size-gap-double) var(--size-gap-quad)\n      rgba(0, 0, 0, 0.25);\n  }\n"],["\n  [data-nextjs-toast] {\n    position: fixed;\n    bottom: var(--size-gap-double);\n    left: var(--size-gap-double);\n    max-width: 420px;\n    z-index: 9000;\n  }\n\n  @media (max-width: 440px) {\n    [data-nextjs-toast] {\n      max-width: 90vw;\n      left: 5vw;\n    }\n  }\n\n  [data-nextjs-toast-wrapper] {\n    padding: 16px;\n    border-radius: var(--size-gap-half);\n    font-weight: 500;\n    color: var(--color-ansi-bright-white);\n    background-color: var(--color-ansi-red);\n    box-shadow: 0px var(--size-gap-double) var(--size-gap-quad)\n      rgba(0, 0, 0, 0.25);\n  }\n"])));n.styles=i;var o},71:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.styles=n.BuildError=void 0;var l=o(t(522));var s=t(824);var u=t(382);var c=t(504);var f=t(373);var d=function BuildError(e){var n=e.message;var t=l.useCallback((function(){}),[]);return l.createElement(u.Overlay,{fixed:true},l.createElement(s.Dialog,{type:"error","aria-labelledby":"nextjs__container_build_error_label","aria-describedby":"nextjs__container_build_error_desc",onClose:t},l.createElement(s.DialogContent,null,l.createElement(s.DialogHeader,{className:"nextjs-container-build-error-header"},l.createElement("h4",{id:"nextjs__container_build_error_label"},"Failed to compile")),l.createElement(s.DialogBody,{className:"nextjs-container-build-error-body"},l.createElement(c.Terminal,{content:n}),l.createElement("footer",null,l.createElement("p",{id:"nextjs__container_build_error_desc"},l.createElement("small",null,"This error occurred during the build process and can only be dismissed by fixing the error.")))))))};n.BuildError=d;n.styles=(0,f.noop)(p||(p=r(["\n  .nextjs-container-build-error-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-build-error-body footer {\n    margin-top: var(--size-gap);\n  }\n  .nextjs-container-build-error-body footer p {\n    margin: 0;\n  }\n\n  .nextjs-container-build-error-body small {\n    color: #757575;\n  }\n"],["\n  .nextjs-container-build-error-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-build-error-body footer {\n    margin-top: var(--size-gap);\n  }\n  .nextjs-container-build-error-body footer p {\n    margin: 0;\n  }\n\n  .nextjs-container-build-error-body small {\n    color: #757575;\n  }\n"])));var p},610:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__assign||function(){a=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return a.apply(this,arguments)};var i=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var o=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var l=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))i(n,e,t);o(n,e);return n};var s=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.styles=n.Errors=void 0;var u=l(t(522));var c=t(591);var f=t(824);var d=t(830);var p=t(382);var v=t(626);var b=t(758);var m=t(915);var h=t(373);var g=t(157);function getErrorSignature(e){var n=e.event;switch(n.type){case c.TYPE_UNHANDLED_ERROR:case c.TYPE_UNHANDLED_REJECTION:{return n.reason.name+"::"+n.reason.message+"::"+n.reason.stack}default:{}}var t=n;return""}var y=function HotlinkedText(e){var n=e.text;var t=/https?:\/\/[^\s/$.?#].[^\s"]*/i;return u.createElement(u.Fragment,null,t.test(n)?n.split(" ").map((function(e,n,r){if(t.test(e)){return u.createElement(u.Fragment,{key:"link-"+n},u.createElement("a",{href:e},e),n===r.length-1?"":" ")}return n===r.length-1?u.createElement(u.Fragment,{key:"text-"+n},e):u.createElement(u.Fragment,{key:"text-"+n},e," ")})):n)};var x=function Errors(e){var n=e.errors;var t=s(u.useState({}),2),r=t[0],i=t[1];var o=s(u.useMemo((function(){var e=[];var t=null;for(var a=0;a<n.length;++a){var i=n[a];var o=i.id;if(o in r){e.push(r[o]);continue}if(a>0){var l=n[a-1];if(getErrorSignature(l)===getErrorSignature(i)){continue}}t=i;break}return[e,t]}),[n,r]),2),l=o[0],c=o[1];var h=u.useMemo((function(){return l.length<1&&Boolean(n.length)}),[n.length,l.length]);u.useEffect((function(){if(c==null){return}var e=true;(0,b.getErrorByType)(c).then((function(n){if(e){i((function(e){var t;return a(a({},e),(t={},t[n.id]=n,t))}))}}),(function(){}));return function(){e=false}}),[c]);var x=s(u.useState(false),2),w=x[0],_=x[1];var E=s(u.useState(0),2),O=E[0],j=E[1];var S=u.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();j((function(e){return Math.max(0,e-1)}))}),[]);var k=u.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();j((function(e){return Math.max(0,Math.min(l.length-1,e+1))}))}),[l.length]);var T=u.useMemo((function(){var e;return(e=l[O])!==null&&e!==void 0?e:null}),[O,l]);u.useEffect((function(){if(n.length<1){i({});_(false);j(0)}}),[n.length]);var C=u.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();_(true)}),[]);var P=u.useCallback((function(e){e===null||e===void 0?void 0:e.preventDefault();_(false)}),[]);if(n.length<1||T==null){return null}if(h){return u.createElement(p.Overlay,null)}if(w){return u.createElement(v.Toast,{className:"nextjs-toast-errors-parent",onClick:P},u.createElement("div",{className:"nextjs-toast-errors"},u.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:"24",height:"24",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},u.createElement("circle",{cx:"12",cy:"12",r:"10"}),u.createElement("line",{x1:"12",y1:"8",x2:"12",y2:"12"}),u.createElement("line",{x1:"12",y1:"16",x2:"12.01",y2:"16"})),u.createElement("span",null,l.length," error",l.length>1?"s":"")))}var M=(0,m.isNodeError)(T.error);return u.createElement(p.Overlay,null,u.createElement(f.Dialog,{type:"error","aria-labelledby":"nextjs__container_errors_label","aria-describedby":"nextjs__container_errors_desc",onClose:M?undefined:C},u.createElement(f.DialogContent,null,u.createElement(f.DialogHeader,{className:"nextjs-container-errors-header"},u.createElement(d.LeftRightDialogHeader,{previous:O>0?S:null,next:O<l.length-1?k:null,close:M?undefined:C},u.createElement("small",null,u.createElement("span",null,O+1)," of"," ",u.createElement("span",null,l.length)," unhandled error",l.length<2?"":"s")),u.createElement("h1",{id:"nextjs__container_errors_label"},M?"Server Error":"Unhandled Runtime Error"),u.createElement("p",{id:"nextjs__container_errors_desc"},T.error.name,":"," ",u.createElement(y,{text:T.error.message})),M?u.createElement("div",null,u.createElement("small",null,"This error happened while generating the page. Any console logs will be displayed in the terminal window.")):undefined),u.createElement(f.DialogBody,{className:"nextjs-container-errors-body"},u.createElement(g.RuntimeError,{key:T.id.toString(),error:T})))))};n.Errors=x;n.styles=(0,h.noop)(w||(w=r(["\n  .nextjs-container-errors-header > h1 {\n    font-size: var(--size-font-big);\n    line-height: var(--size-font-bigger);\n    font-weight: bold;\n    margin: 0;\n    margin-top: calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  .nextjs-container-errors-header small {\n    font-size: var(--size-font-small);\n    color: var(--color-accents-1);\n    margin-left: var(--size-gap-double);\n  }\n  .nextjs-container-errors-header small > span {\n    font-family: var(--font-stack-monospace);\n  }\n  .nextjs-container-errors-header > p {\n    font-family: var(--font-stack-monospace);\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-big);\n    font-weight: bold;\n    margin: 0;\n    margin-top: var(--size-gap-half);\n    color: var(--color-ansi-red);\n    white-space: pre-wrap;\n  }\n  .nextjs-container-errors-header > div > small {\n    margin: 0;\n    margin-top: var(--size-gap-half);\n  }\n  .nextjs-container-errors-header > p > a {\n    color: var(--color-ansi-red);\n  }\n\n  .nextjs-container-errors-body > h5:not(:first-child) {\n    margin-top: calc(var(--size-gap-double) + var(--size-gap));\n  }\n  .nextjs-container-errors-body > h5 {\n    margin-bottom: var(--size-gap);\n  }\n\n  .nextjs-toast-errors-parent {\n    cursor: pointer;\n    transition: transform 0.2s ease;\n  }\n  .nextjs-toast-errors-parent:hover {\n    transform: scale(1.1);\n  }\n  .nextjs-toast-errors {\n    display: flex;\n    align-items: center;\n    justify-content: flex-start;\n  }\n  .nextjs-toast-errors > svg {\n    margin-right: var(--size-gap);\n  }\n"],["\n  .nextjs-container-errors-header > h1 {\n    font-size: var(--size-font-big);\n    line-height: var(--size-font-bigger);\n    font-weight: bold;\n    margin: 0;\n    margin-top: calc(var(--size-gap-double) + var(--size-gap-half));\n  }\n  .nextjs-container-errors-header small {\n    font-size: var(--size-font-small);\n    color: var(--color-accents-1);\n    margin-left: var(--size-gap-double);\n  }\n  .nextjs-container-errors-header small > span {\n    font-family: var(--font-stack-monospace);\n  }\n  .nextjs-container-errors-header > p {\n    font-family: var(--font-stack-monospace);\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-big);\n    font-weight: bold;\n    margin: 0;\n    margin-top: var(--size-gap-half);\n    color: var(--color-ansi-red);\n    white-space: pre-wrap;\n  }\n  .nextjs-container-errors-header > div > small {\n    margin: 0;\n    margin-top: var(--size-gap-half);\n  }\n  .nextjs-container-errors-header > p > a {\n    color: var(--color-ansi-red);\n  }\n\n  .nextjs-container-errors-body > h5:not(:first-child) {\n    margin-top: calc(var(--size-gap-double) + var(--size-gap));\n  }\n  .nextjs-container-errors-body > h5 {\n    margin-bottom: var(--size-gap);\n  }\n\n  .nextjs-toast-errors-parent {\n    cursor: pointer;\n    transition: transform 0.2s ease;\n  }\n  .nextjs-toast-errors-parent:hover {\n    transform: scale(1.1);\n  }\n  .nextjs-toast-errors {\n    display: flex;\n    align-items: center;\n    justify-content: flex-start;\n  }\n  .nextjs-toast-errors > svg {\n    margin-right: var(--size-gap);\n  }\n"])));var w},710:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.styles=n.FullRefreshWarning=void 0;var l=o(t(522));var s=t(824);var u=t(382);var c=t(504);var f=t(373);var d=function FullRefreshWarning(e){var n=e.reason;var t=l.useCallback((function(){window.location.reload()}),[]);return l.createElement(u.Overlay,{fixed:true},l.createElement(s.Dialog,{type:"warning","aria-labelledby":"nextjs__container_refresh_warning_label","aria-describedby":"nextjs__container_refresh_warning_desc",onClose:t},l.createElement(s.DialogContent,null,l.createElement(s.DialogHeader,{className:"nextjs-container-refresh-warning-header"},l.createElement("h4",{id:"nextjs__container_refresh_warning_label"},"About to perform a full refresh")),l.createElement(s.DialogBody,{className:"nextjs-container-refresh-warning-body"},l.createElement(FullRefreshWarningReason,{reason:n}),l.createElement("footer",null,l.createElement("p",null,"You can read more about Fast Refresh in"," ",l.createElement("a",{href:"https://nextjs.org/docs/basic-features/fast-refresh#how-it-works"},"our documentation"),"."),l.createElement("button",{onClick:t},"Reload"))))))};n.FullRefreshWarning=d;n.styles=(0,f.noop)(p||(p=r(["\n  .nextjs-container-refresh-warning-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-refresh-warning-body footer {\n    margin-top: var(--size-gap-double);\n  }\n\n  .nextjs-container-build-error-body p {\n    color: #757575;\n  }\n\n  .nextjs-container-refresh-warning-body button {\n    background-color: var(--color-ansi-yellow);\n    border: 0;\n    border-radius: var(--size-gap-half);\n    color: var(--color-ansi-black);\n    cursor: pointer;\n    display: block;\n    margin-left: auto;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n    transition: background-color 0.25s ease;\n  }\n\n  .nextjs-container-refresh-warning-body button:hover {\n    background-color: var(--color-ansi-bright-yellow);\n  }\n"],["\n  .nextjs-container-refresh-warning-header > h4 {\n    line-height: 1.5;\n    margin: 0;\n    padding: 0;\n  }\n\n  .nextjs-container-refresh-warning-body footer {\n    margin-top: var(--size-gap-double);\n  }\n\n  .nextjs-container-build-error-body p {\n    color: #757575;\n  }\n\n  .nextjs-container-refresh-warning-body button {\n    background-color: var(--color-ansi-yellow);\n    border: 0;\n    border-radius: var(--size-gap-half);\n    color: var(--color-ansi-black);\n    cursor: pointer;\n    display: block;\n    margin-left: auto;\n    padding: calc(var(--size-gap) + var(--size-gap-half))\n      calc(var(--size-gap-double) + var(--size-gap-half));\n    transition: background-color 0.25s ease;\n  }\n\n  .nextjs-container-refresh-warning-body button:hover {\n    background-color: var(--color-ansi-bright-yellow);\n  }\n"])));var FullRefreshWarningReason=function(e){var n=e.reason;if(n===null){return l.createElement("p",null,"Fast Refresh will perform a full reload because your application had an unrecoverable error.")}return l.createElement(l.Fragment,null,l.createElement("p",null,"Fast Refresh will perform a full reload when you edit a file that is imported by modules outside of the React rendering tree. It is also possible the parent component of the component you edited is a class component, which disables Fast Refresh. Fast Refresh requires at least one parent function component in your React tree."),l.createElement("p",null,"You can find more information in the related error below:"),l.createElement(c.Terminal,{content:n}))};var p},157:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};var l=this&&this.__read||function(e,n){var t=typeof Symbol==="function"&&e[Symbol.iterator];if(!t)return e;var r=t.call(e),a,i=[],o;try{while((n===void 0||n-- >0)&&!(a=r.next()).done)i.push(a.value)}catch(e){o={error:e}}finally{try{if(a&&!a.done&&(t=r["return"]))t.call(r)}finally{if(o)throw o.error}}return i};n.__esModule=true;n.RuntimeError=n.styles=void 0;var s=o(t(522));var u=t(983);var c=t(373);var f=t(392);var d=function CallStackFrame(e){var n;var t=e.frame;var r=(n=t.originalStackFrame)!==null&&n!==void 0?n:t.sourceStackFrame;var a=Boolean(t.originalCodeFrame);var i=s.useCallback((function(){var e;if(!a)return;var n=new URLSearchParams;for(var t in r){n.append(t,((e=r[t])!==null&&e!==void 0?e:"").toString())}self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_launch-editor?"+n.toString()).then((function(){}),(function(){console.error("There was an issue opening this code in your editor.")}))}),[a,r]);return s.createElement("div",{"data-nextjs-call-stack-frame":true},s.createElement("h6",{"data-nextjs-frame-expanded":Boolean(t.expanded)},r.methodName),s.createElement("div",{"data-has-source":a?"true":undefined,tabIndex:a?10:undefined,role:a?"link":undefined,onClick:i,title:a?"Click to open in your editor":undefined},s.createElement("span",null,(0,f.getFrameSource)(r)),s.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"none",stroke:"currentColor",strokeWidth:"2",strokeLinecap:"round",strokeLinejoin:"round"},s.createElement("path",{d:"M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"}),s.createElement("polyline",{points:"15 3 21 3 21 9"}),s.createElement("line",{x1:"10",y1:"14",x2:"21",y2:"3"}))))};var p=function RuntimeError(e){var n=e.error;var t=s.useMemo((function(){return n.frames.findIndex((function(e){return e.expanded&&Boolean(e.originalCodeFrame)&&Boolean(e.originalStackFrame)}))}),[n.frames]);var r=s.useMemo((function(){var e;return(e=n.frames[t])!==null&&e!==void 0?e:null}),[n.frames,t]);var a=s.useMemo((function(){return t<0?[]:n.frames.slice(0,t)}),[n.frames,t]);var i=l(s.useState(r==null),2),o=i[0],c=i[1];var f=s.useCallback((function(){c((function(e){return!e}))}),[]);var p=s.useMemo((function(){return a.filter((function(e){return e.expanded||o}))}),[o,a]);var v=s.useMemo((function(){return n.frames.slice(t+1)}),[n.frames,t]);var b=s.useMemo((function(){return v.filter((function(e){return e.expanded||o}))}),[o,v]);var m=s.useMemo((function(){return v.length!==b.length||o&&r!=null}),[o,v.length,r,b.length]);return s.createElement(s.Fragment,null,r?s.createElement(s.Fragment,null,s.createElement("h5",null,"Source"),p.map((function(e,n){return s.createElement(d,{key:"leading-frame-"+n+"-"+o,frame:e})})),s.createElement(u.CodeFrame,{stackFrame:r.originalStackFrame,codeFrame:r.originalCodeFrame})):undefined,b.length?s.createElement(s.Fragment,null,s.createElement("h5",null,"Call Stack"),b.map((function(e,n){return s.createElement(d,{key:"call-stack-"+n+"-"+o,frame:e})}))):undefined,m?s.createElement(s.Fragment,null,s.createElement("button",{tabIndex:10,"data-nextjs-data-runtime-error-collapsed-action":true,type:"button",onClick:f},o?"Hide":"Show"," collapsed frames")):undefined)};n.RuntimeError=p;n.styles=(0,c.noop)(v||(v=r(["\n  button[data-nextjs-data-runtime-error-collapsed-action] {\n    background: none;\n    border: none;\n    padding: 0;\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-bigger);\n    color: var(--color-accents-3);\n  }\n\n  [data-nextjs-call-stack-frame]:not(:last-child) {\n    margin-bottom: var(--size-gap-double);\n  }\n\n  [data-nextjs-call-stack-frame] > h6 {\n    margin-top: 0;\n    margin-bottom: var(--size-gap);\n    font-family: var(--font-stack-monospace);\n    color: #222;\n  }\n  [data-nextjs-call-stack-frame] > h6[data-nextjs-frame-expanded='false'] {\n    color: #666;\n  }\n  [data-nextjs-call-stack-frame] > div {\n    display: flex;\n    align-items: center;\n    padding-left: calc(var(--size-gap) + var(--size-gap-half));\n    font-size: var(--size-font-small);\n    color: #999;\n  }\n  [data-nextjs-call-stack-frame] > div > svg {\n    width: auto;\n    height: var(--size-font-small);\n    margin-left: var(--size-gap);\n\n    display: none;\n  }\n\n  [data-nextjs-call-stack-frame] > div[data-has-source] {\n    cursor: pointer;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source]:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source] > svg {\n    display: unset;\n  }\n"],["\n  button[data-nextjs-data-runtime-error-collapsed-action] {\n    background: none;\n    border: none;\n    padding: 0;\n    font-size: var(--size-font-small);\n    line-height: var(--size-font-bigger);\n    color: var(--color-accents-3);\n  }\n\n  [data-nextjs-call-stack-frame]:not(:last-child) {\n    margin-bottom: var(--size-gap-double);\n  }\n\n  [data-nextjs-call-stack-frame] > h6 {\n    margin-top: 0;\n    margin-bottom: var(--size-gap);\n    font-family: var(--font-stack-monospace);\n    color: #222;\n  }\n  [data-nextjs-call-stack-frame] > h6[data-nextjs-frame-expanded='false'] {\n    color: #666;\n  }\n  [data-nextjs-call-stack-frame] > div {\n    display: flex;\n    align-items: center;\n    padding-left: calc(var(--size-gap) + var(--size-gap-half));\n    font-size: var(--size-font-small);\n    color: #999;\n  }\n  [data-nextjs-call-stack-frame] > div > svg {\n    width: auto;\n    height: var(--size-font-small);\n    margin-left: var(--size-gap);\n\n    display: none;\n  }\n\n  [data-nextjs-call-stack-frame] > div[data-has-source] {\n    cursor: pointer;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source]:hover {\n    text-decoration: underline dotted;\n  }\n  [data-nextjs-call-stack-frame] > div[data-has-source] > svg {\n    display: unset;\n  }\n"])));var v},758:function(e,n,t){"use strict";var r=this&&this.__awaiter||function(e,n,t,r){function adopt(e){return e instanceof t?e:new t((function(n){n(e)}))}return new(t||(t=Promise))((function(t,a){function fulfilled(e){try{step(r.next(e))}catch(e){a(e)}}function rejected(e){try{step(r["throw"](e))}catch(e){a(e)}}function step(e){e.done?t(e.value):adopt(e.value).then(fulfilled,rejected)}step((r=r.apply(e,n||[])).next())}))};var a=this&&this.__generator||function(e,n){var t={label:0,sent:function(){if(i[0]&1)throw i[1];return i[1]},trys:[],ops:[]},r,a,i,o;return o={next:verb(0),throw:verb(1),return:verb(2)},typeof Symbol==="function"&&(o[Symbol.iterator]=function(){return this}),o;function verb(e){return function(n){return step([e,n])}}function step(o){if(r)throw new TypeError("Generator is already executing.");while(t)try{if(r=1,a&&(i=o[0]&2?a["return"]:o[0]?a["throw"]||((i=a["return"])&&i.call(a),0):a.next)&&!(i=i.call(a,o[1])).done)return i;if(a=0,i)o=[o[0]&2,i.value];switch(o[0]){case 0:case 1:i=o;break;case 4:t.label++;return{value:o[1],done:false};case 5:t.label++;a=o[1];o=[0];continue;case 7:o=t.ops.pop();t.trys.pop();continue;default:if(!(i=t.trys,i=i.length>0&&i[i.length-1])&&(o[0]===6||o[0]===2)){t=0;continue}if(o[0]===3&&(!i||o[1]>i[0]&&o[1]<i[3])){t.label=o[1];break}if(o[0]===6&&t.label<i[1]){t.label=i[1];i=o;break}if(i&&t.label<i[2]){t.label=i[2];t.ops.push(o);break}if(i[2])t.ops.pop();t.trys.pop();continue}o=n.call(e,t)}catch(e){o=[6,e];a=0}finally{r=i=0}if(o[0]&5)throw o[1];return{value:o[0]?o[1]:void 0,done:true}}};n.__esModule=true;n.getErrorByType=void 0;var i=t(591);var o=t(915);var l=t(392);function getErrorByType(e){return r(this,void 0,void 0,(function(){var n,t,r,s;var u;return a(this,(function(a){switch(a.label){case 0:n=e.id,t=e.event;r=t.type;switch(r){case i.TYPE_UNHANDLED_ERROR:return[3,1];case i.TYPE_UNHANDLED_REJECTION:return[3,1]}return[3,3];case 1:u={id:n,runtime:true,error:t.reason};return[4,(0,l.getOriginalStackFrames)((0,o.isNodeError)(t.reason),t.frames)];case 2:return[2,(u.frames=a.sent(),u)];case 3:{return[3,4]}a.label=4;case 4:s=t;throw new Error("type system invariant violation")}}))}))}n.getErrorByType=getErrorByType},915:function(e,n,t){"use strict";var r=this&&this.__assign||function(){r=Object.assign||function(e){for(var n,t=1,r=arguments.length;t<r;t++){n=arguments[t];for(var a in n)if(Object.prototype.hasOwnProperty.call(n,a))e[a]=n[a]}return e};return r.apply(this,arguments)};n.__esModule=true;n.getNodeError=n.isNodeError=n.getFilesystemFrame=void 0;var a=t(922);function getFilesystemFrame(e){var n=r({},e);if(typeof n.file==="string"){if(n.file.startsWith("/")||/^[a-z]:\\/i.test(n.file)||n.file.startsWith("\\\\")){n.file="file://"+n.file}}return n}n.getFilesystemFrame=getFilesystemFrame;var i=Symbol("NextjsNodeError");function isNodeError(e){return i in e}n.isNodeError=isNodeError;function getNodeError(e){var n;try{throw new Error(e.message)}catch(e){n=e}n.name=e.name;try{n.stack=(0,a.parse)(e.stack).map(getFilesystemFrame).map((function(e){var n="    at "+e.methodName;if(e.file){var t=e.file;if(e.lineNumber){t+=":"+e.lineNumber;if(e.column){t+=":"+e.column}}n+=" ("+t+")"}return n})).join("\n")}catch(t){n.stack=e.stack}Object.defineProperty(n,i,{writable:false,enumerable:false,configurable:false});return n}n.getNodeError=getNodeError},373:function(e,n){"use strict";n.__esModule=true;n.noop=void 0;function noop(e){var n=[];for(var t=1;t<arguments.length;t++){n[t-1]=arguments[t]}var r=e.length-1;return e.slice(0,r).reduce((function(e,t,r){return e+t+n[r]}),"")+e[r]}n.noop=noop},525:function(e,n,t){"use strict";n.__esModule=true;n.parseStack=void 0;var r=t(922);var a=/\/_next(\/static\/.+)/g;function parseStack(e){var n=(0,r.parse)(e);return n.map((function(e){var n,t;try{var r=new URL(e.file);var i=a.exec(r.pathname);if(i){var o=(t=(n=process.env.__NEXT_DIST_DIR)===null||n===void 0?void 0:n.replace(/\\/g,"/"))===null||t===void 0?void 0:t.replace(/\/$/,"");if(o){e.file="file://"+o.concat(i.pop())}}}catch(e){}return e}))}n.parseStack=parseStack},392:function(e,n){"use strict";var t=this&&this.__awaiter||function(e,n,t,r){function adopt(e){return e instanceof t?e:new t((function(n){n(e)}))}return new(t||(t=Promise))((function(t,a){function fulfilled(e){try{step(r.next(e))}catch(e){a(e)}}function rejected(e){try{step(r["throw"](e))}catch(e){a(e)}}function step(e){e.done?t(e.value):adopt(e.value).then(fulfilled,rejected)}step((r=r.apply(e,n||[])).next())}))};var r=this&&this.__generator||function(e,n){var t={label:0,sent:function(){if(i[0]&1)throw i[1];return i[1]},trys:[],ops:[]},r,a,i,o;return o={next:verb(0),throw:verb(1),return:verb(2)},typeof Symbol==="function"&&(o[Symbol.iterator]=function(){return this}),o;function verb(e){return function(n){return step([e,n])}}function step(o){if(r)throw new TypeError("Generator is already executing.");while(t)try{if(r=1,a&&(i=o[0]&2?a["return"]:o[0]?a["throw"]||((i=a["return"])&&i.call(a),0):a.next)&&!(i=i.call(a,o[1])).done)return i;if(a=0,i)o=[o[0]&2,i.value];switch(o[0]){case 0:case 1:i=o;break;case 4:t.label++;return{value:o[1],done:false};case 5:t.label++;a=o[1];o=[0];continue;case 7:o=t.ops.pop();t.trys.pop();continue;default:if(!(i=t.trys,i=i.length>0&&i[i.length-1])&&(o[0]===6||o[0]===2)){t=0;continue}if(o[0]===3&&(!i||o[1]>i[0]&&o[1]<i[3])){t.label=o[1];break}if(o[0]===6&&t.label<i[1]){t.label=i[1];i=o;break}if(i&&t.label<i[2]){t.label=i[2];t.ops.push(o);break}if(i[2])t.ops.pop();t.trys.pop();continue}o=n.call(e,t)}catch(e){o=[6,e];a=0}finally{r=i=0}if(o[0]&5)throw o[1];return{value:o[0]?o[1]:void 0,done:true}}};n.__esModule=true;n.getFrameSource=n.getOriginalStackFrame=n.getOriginalStackFrames=void 0;function getOriginalStackFrames(e,n){return Promise.all(n.map((function(n){return getOriginalStackFrame(e,n)})))}n.getOriginalStackFrames=getOriginalStackFrames;function getOriginalStackFrame(e,n){var a,i;function _getOriginalStackFrame(){var a,i,o,l,s;return t(this,void 0,void 0,(function(){var t,u,c,f,d,p,v,b,m;return r(this,(function(r){switch(r.label){case 0:t=new URLSearchParams;t.append("isServerSide",String(e));for(u in n){t.append(u,((a=n[u])!==null&&a!==void 0?a:"").toString())}c=new AbortController;f=setTimeout((function(){return c.abort()}),3e3);return[4,self.fetch((process.env.__NEXT_ROUTER_BASEPATH||"")+"/__nextjs_original-stack-frame?"+t.toString(),{signal:c.signal})["finally"]((function(){clearTimeout(f)}))];case 1:d=r.sent();if(!(!d.ok||d.status===204))return[3,3];v=(p=Promise).reject;b=Error.bind;return[4,d.text()];case 2:return[2,v.apply(p,[new(b.apply(Error,[void 0,r.sent()]))])];case 3:return[4,d.json()];case 4:m=r.sent();return[2,{error:false,reason:null,external:false,expanded:!Boolean((s=((i=n.file)===null||i===void 0?void 0:i.includes("node_modules"))||((l=(o=m.originalStackFrame)===null||o===void 0?void 0:o.file)===null||l===void 0?void 0:l.includes("node_modules")))!==null&&s!==void 0?s:true),sourceStackFrame:n,originalStackFrame:m.originalStackFrame,originalCodeFrame:m.originalCodeFrame||null}]}}))}))}if(!(((a=n.file)===null||a===void 0?void 0:a.startsWith("webpack-internal:"))||((i=n.file)===null||i===void 0?void 0:i.startsWith("file:")))){return Promise.resolve({error:false,reason:null,external:true,expanded:false,sourceStackFrame:n,originalStackFrame:null,originalCodeFrame:null})}return _getOriginalStackFrame()["catch"]((function(e){var t,r;return{error:true,reason:(r=(t=e===null||e===void 0?void 0:e.message)!==null&&t!==void 0?t:e===null||e===void 0?void 0:e.toString())!==null&&r!==void 0?r:"Unknown Error",external:false,expanded:false,sourceStackFrame:n,originalStackFrame:null,originalCodeFrame:null}}))}n.getOriginalStackFrame=getOriginalStackFrame;function getFrameSource(e){var n;var t="";try{var r=new URL(e.file);if(typeof globalThis!=="undefined"&&((n=globalThis.location)===null||n===void 0?void 0:n.origin)!==r.origin){if(r.origin==="null"){t+=r.protocol}else{t+=r.origin}}t+=r.pathname;t+=" "}catch(n){t+=(e.file||"(unknown)")+" "}if(e.lineNumber!=null){if(e.column!=null){t+="("+e.lineNumber+":"+e.column+") "}else{t+="("+e.lineNumber+") "}}return t.slice(0,-1)}n.getFrameSource=getFrameSource},814:function(e,n,t){"use strict";var r=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var a=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var i=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))r(n,e,t);a(n,e);return n};n.__esModule=true;n.useOnClickOutside=void 0;var o=i(t(522));function useOnClickOutside(e,n){o.useEffect((function(){if(e==null||n==null){return}var listener=function(t){if(!e||e.contains(t.target)){return}n(t)};var t=e.getRootNode();t.addEventListener("mousedown",listener);t.addEventListener("touchstart",listener);return function(){t.removeEventListener("mousedown",listener);t.removeEventListener("touchstart",listener)}}),[n,e])}n.useOnClickOutside=useOnClickOutside},796:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.Base=void 0;var l=o(t(522));var s=t(373);function Base(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,s.noop)(u||(u=r(["\n          :host {\n            --size-gap-half: 4px;\n            --size-gap: 8px;\n            --size-gap-double: 16px;\n            --size-gap-quad: 32px;\n\n            --size-font-small: 14px;\n            --size-font: 16px;\n            --size-font-big: 20px;\n            --size-font-bigger: 24px;\n\n            --color-accents-1: #808080;\n            --color-accents-2: #222222;\n            --color-accents-3: #404040;\n\n            --font-stack-monospace: 'SFMono-Regular', Consolas,\n              'Liberation Mono', Menlo, Courier, monospace;\n\n            --color-ansi-selection: rgba(95, 126, 151, 0.48);\n            --color-ansi-bg: #111111;\n            --color-ansi-fg: #cccccc;\n\n            --color-ansi-white: #777777;\n            --color-ansi-black: #141414;\n            --color-ansi-blue: #00aaff;\n            --color-ansi-cyan: #88ddff;\n            --color-ansi-green: #98ec65;\n            --color-ansi-magenta: #aa88ff;\n            --color-ansi-red: #ff5555;\n            --color-ansi-yellow: #ffcc33;\n            --color-ansi-bright-white: #ffffff;\n            --color-ansi-bright-black: #777777;\n            --color-ansi-bright-blue: #33bbff;\n            --color-ansi-bright-cyan: #bbecff;\n            --color-ansi-bright-green: #b6f292;\n            --color-ansi-bright-magenta: #cebbff;\n            --color-ansi-bright-red: #ff8888;\n            --color-ansi-bright-yellow: #ffd966;\n          }\n\n          .mono {\n            font-family: var(--font-stack-monospace);\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-bottom: var(--size-gap);\n            font-weight: 500;\n            line-height: 1.5;\n          }\n\n          h1 {\n            font-size: 40px;\n          }\n          h2 {\n            font-size: 32px;\n          }\n          h3 {\n            font-size: 28px;\n          }\n          h4 {\n            font-size: 24px;\n          }\n          h5 {\n            font-size: 20px;\n          }\n          h6 {\n            font-size: 16px;\n          }\n        "],["\n          :host {\n            --size-gap-half: 4px;\n            --size-gap: 8px;\n            --size-gap-double: 16px;\n            --size-gap-quad: 32px;\n\n            --size-font-small: 14px;\n            --size-font: 16px;\n            --size-font-big: 20px;\n            --size-font-bigger: 24px;\n\n            --color-accents-1: #808080;\n            --color-accents-2: #222222;\n            --color-accents-3: #404040;\n\n            --font-stack-monospace: 'SFMono-Regular', Consolas,\n              'Liberation Mono', Menlo, Courier, monospace;\n\n            --color-ansi-selection: rgba(95, 126, 151, 0.48);\n            --color-ansi-bg: #111111;\n            --color-ansi-fg: #cccccc;\n\n            --color-ansi-white: #777777;\n            --color-ansi-black: #141414;\n            --color-ansi-blue: #00aaff;\n            --color-ansi-cyan: #88ddff;\n            --color-ansi-green: #98ec65;\n            --color-ansi-magenta: #aa88ff;\n            --color-ansi-red: #ff5555;\n            --color-ansi-yellow: #ffcc33;\n            --color-ansi-bright-white: #ffffff;\n            --color-ansi-bright-black: #777777;\n            --color-ansi-bright-blue: #33bbff;\n            --color-ansi-bright-cyan: #bbecff;\n            --color-ansi-bright-green: #b6f292;\n            --color-ansi-bright-magenta: #cebbff;\n            --color-ansi-bright-red: #ff8888;\n            --color-ansi-bright-yellow: #ffd966;\n          }\n\n          .mono {\n            font-family: var(--font-stack-monospace);\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-bottom: var(--size-gap);\n            font-weight: 500;\n            line-height: 1.5;\n          }\n\n          h1 {\n            font-size: 40px;\n          }\n          h2 {\n            font-size: 32px;\n          }\n          h3 {\n            font-size: 28px;\n          }\n          h4 {\n            font-size: 24px;\n          }\n          h5 {\n            font-size: 20px;\n          }\n          h6 {\n            font-size: 16px;\n          }\n        "])))}})}n.Base=Base;var u},174:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.ComponentStyles=void 0;var l=o(t(522));var s=t(761);var u=t(824);var c=t(448);var f=t(173);var d=t(492);var p=t(626);var v=t(71);var b=t(610);var m=t(710);var h=t(157);var g=t(373);function ComponentStyles(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,g.noop)(y||(y=r(["\n          ","\n          ","\n          ","\n          ","\n          ","\n          ","\n\n          ","\n          ","\n          ","\n          ","\n        "],["\n          ","\n          ","\n          ","\n          ","\n          ","\n          ","\n\n          ","\n          ","\n          ","\n          ","\n        "])),f.styles,p.styles,u.styles,c.styles,s.styles,d.styles,v.styles,b.styles,h.styles,m.styles)}})}n.ComponentStyles=ComponentStyles;var y},803:function(e,n,t){"use strict";var r=this&&this.__makeTemplateObject||function(e,n){if(Object.defineProperty){Object.defineProperty(e,"raw",{value:n})}else{e.raw=n}return e};var a=this&&this.__createBinding||(Object.create?function(e,n,t,r){if(r===undefined)r=t;Object.defineProperty(e,r,{enumerable:true,get:function(){return n[t]}})}:function(e,n,t,r){if(r===undefined)r=t;e[r]=n[t]});var i=this&&this.__setModuleDefault||(Object.create?function(e,n){Object.defineProperty(e,"default",{enumerable:true,value:n})}:function(e,n){e["default"]=n});var o=this&&this.__importStar||function(e){if(e&&e.__esModule)return e;var n={};if(e!=null)for(var t in e)if(t!=="default"&&Object.prototype.hasOwnProperty.call(e,t))a(n,e,t);i(n,e);return n};n.__esModule=true;n.CssReset=void 0;var l=o(t(522));var s=t(373);function CssReset(){return l.createElement("style",{dangerouslySetInnerHTML:{__html:(0,s.noop)(u||(u=r(["\n          :host {\n            all: initial;\n\n            /* the direction property is not reset by 'all' */\n            direction: ltr;\n          }\n\n          /*!\n           * Bootstrap Reboot v4.4.1 (https://getbootstrap.com/)\n           * Copyright 2011-2019 The Bootstrap Authors\n           * Copyright 2011-2019 Twitter, Inc.\n           * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)\n           * Forked from Normalize.css, licensed MIT (https://github.com/necolas/normalize.css/blob/master/LICENSE.md)\n           */\n          *,\n          *::before,\n          *::after {\n            box-sizing: border-box;\n          }\n\n          :host {\n            font-family: sans-serif;\n            line-height: 1.15;\n            -webkit-text-size-adjust: 100%;\n            -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n          }\n\n          article,\n          aside,\n          figcaption,\n          figure,\n          footer,\n          header,\n          hgroup,\n          main,\n          nav,\n          section {\n            display: block;\n          }\n\n          :host {\n            margin: 0;\n            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,\n              'Helvetica Neue', Arial, 'Noto Sans', sans-serif,\n              'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',\n              'Noto Color Emoji';\n            font-size: 16px;\n            font-weight: 400;\n            line-height: 1.5;\n            color: #212529;\n            text-align: left;\n            background-color: #fff;\n          }\n\n          [tabindex='-1']:focus:not(:focus-visible) {\n            outline: 0 !important;\n          }\n\n          hr {\n            box-sizing: content-box;\n            height: 0;\n            overflow: visible;\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-top: 0;\n            margin-bottom: 8px;\n          }\n\n          p {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          abbr[title],\n          abbr[data-original-title] {\n            text-decoration: underline;\n            -webkit-text-decoration: underline dotted;\n            text-decoration: underline dotted;\n            cursor: help;\n            border-bottom: 0;\n            -webkit-text-decoration-skip-ink: none;\n            text-decoration-skip-ink: none;\n          }\n\n          address {\n            margin-bottom: 16px;\n            font-style: normal;\n            line-height: inherit;\n          }\n\n          ol,\n          ul,\n          dl {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          ol ol,\n          ul ul,\n          ol ul,\n          ul ol {\n            margin-bottom: 0;\n          }\n\n          dt {\n            font-weight: 700;\n          }\n\n          dd {\n            margin-bottom: 8px;\n            margin-left: 0;\n          }\n\n          blockquote {\n            margin: 0 0 16px;\n          }\n\n          b,\n          strong {\n            font-weight: bolder;\n          }\n\n          small {\n            font-size: 80%;\n          }\n\n          sub,\n          sup {\n            position: relative;\n            font-size: 75%;\n            line-height: 0;\n            vertical-align: baseline;\n          }\n\n          sub {\n            bottom: -0.25em;\n          }\n\n          sup {\n            top: -0.5em;\n          }\n\n          a {\n            color: #007bff;\n            text-decoration: none;\n            background-color: transparent;\n          }\n\n          a:hover {\n            color: #0056b3;\n            text-decoration: underline;\n          }\n\n          a:not([href]) {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          a:not([href]):hover {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          pre,\n          code,\n          kbd,\n          samp {\n            font-family: SFMono-Regular, Menlo, Monaco, Consolas,\n              'Liberation Mono', 'Courier New', monospace;\n            font-size: 1em;\n          }\n\n          pre {\n            margin-top: 0;\n            margin-bottom: 16px;\n            overflow: auto;\n          }\n\n          figure {\n            margin: 0 0 16px;\n          }\n\n          img {\n            vertical-align: middle;\n            border-style: none;\n          }\n\n          svg {\n            overflow: hidden;\n            vertical-align: middle;\n          }\n\n          table {\n            border-collapse: collapse;\n          }\n\n          caption {\n            padding-top: 12px;\n            padding-bottom: 12px;\n            color: #6c757d;\n            text-align: left;\n            caption-side: bottom;\n          }\n\n          th {\n            text-align: inherit;\n          }\n\n          label {\n            display: inline-block;\n            margin-bottom: 8px;\n          }\n\n          button {\n            border-radius: 0;\n          }\n\n          button:focus {\n            outline: 1px dotted;\n            outline: 5px auto -webkit-focus-ring-color;\n          }\n\n          input,\n          button,\n          select,\n          optgroup,\n          textarea {\n            margin: 0;\n            font-family: inherit;\n            font-size: inherit;\n            line-height: inherit;\n          }\n\n          button,\n          input {\n            overflow: visible;\n          }\n\n          button,\n          select {\n            text-transform: none;\n          }\n\n          select {\n            word-wrap: normal;\n          }\n\n          button,\n          [type='button'],\n          [type='reset'],\n          [type='submit'] {\n            -webkit-appearance: button;\n          }\n\n          button:not(:disabled),\n          [type='button']:not(:disabled),\n          [type='reset']:not(:disabled),\n          [type='submit']:not(:disabled) {\n            cursor: pointer;\n          }\n\n          button::-moz-focus-inner,\n          [type='button']::-moz-focus-inner,\n          [type='reset']::-moz-focus-inner,\n          [type='submit']::-moz-focus-inner {\n            padding: 0;\n            border-style: none;\n          }\n\n          input[type='radio'],\n          input[type='checkbox'] {\n            box-sizing: border-box;\n            padding: 0;\n          }\n\n          input[type='date'],\n          input[type='time'],\n          input[type='datetime-local'],\n          input[type='month'] {\n            -webkit-appearance: listbox;\n          }\n\n          textarea {\n            overflow: auto;\n            resize: vertical;\n          }\n\n          fieldset {\n            min-width: 0;\n            padding: 0;\n            margin: 0;\n            border: 0;\n          }\n\n          legend {\n            display: block;\n            width: 100%;\n            max-width: 100%;\n            padding: 0;\n            margin-bottom: 8px;\n            font-size: 24px;\n            line-height: inherit;\n            color: inherit;\n            white-space: normal;\n          }\n\n          progress {\n            vertical-align: baseline;\n          }\n\n          [type='number']::-webkit-inner-spin-button,\n          [type='number']::-webkit-outer-spin-button {\n            height: auto;\n          }\n\n          [type='search'] {\n            outline-offset: -2px;\n            -webkit-appearance: none;\n          }\n\n          [type='search']::-webkit-search-decoration {\n            -webkit-appearance: none;\n          }\n\n          ::-webkit-file-upload-button {\n            font: inherit;\n            -webkit-appearance: button;\n          }\n\n          output {\n            display: inline-block;\n          }\n\n          summary {\n            display: list-item;\n            cursor: pointer;\n          }\n\n          template {\n            display: none;\n          }\n\n          [hidden] {\n            display: none !important;\n          }\n        "],["\n          :host {\n            all: initial;\n\n            /* the direction property is not reset by 'all' */\n            direction: ltr;\n          }\n\n          /*!\n           * Bootstrap Reboot v4.4.1 (https://getbootstrap.com/)\n           * Copyright 2011-2019 The Bootstrap Authors\n           * Copyright 2011-2019 Twitter, Inc.\n           * Licensed under MIT (https://github.com/twbs/bootstrap/blob/master/LICENSE)\n           * Forked from Normalize.css, licensed MIT (https://github.com/necolas/normalize.css/blob/master/LICENSE.md)\n           */\n          *,\n          *::before,\n          *::after {\n            box-sizing: border-box;\n          }\n\n          :host {\n            font-family: sans-serif;\n            line-height: 1.15;\n            -webkit-text-size-adjust: 100%;\n            -webkit-tap-highlight-color: rgba(0, 0, 0, 0);\n          }\n\n          article,\n          aside,\n          figcaption,\n          figure,\n          footer,\n          header,\n          hgroup,\n          main,\n          nav,\n          section {\n            display: block;\n          }\n\n          :host {\n            margin: 0;\n            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,\n              'Helvetica Neue', Arial, 'Noto Sans', sans-serif,\n              'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',\n              'Noto Color Emoji';\n            font-size: 16px;\n            font-weight: 400;\n            line-height: 1.5;\n            color: #212529;\n            text-align: left;\n            background-color: #fff;\n          }\n\n          [tabindex='-1']:focus:not(:focus-visible) {\n            outline: 0 !important;\n          }\n\n          hr {\n            box-sizing: content-box;\n            height: 0;\n            overflow: visible;\n          }\n\n          h1,\n          h2,\n          h3,\n          h4,\n          h5,\n          h6 {\n            margin-top: 0;\n            margin-bottom: 8px;\n          }\n\n          p {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          abbr[title],\n          abbr[data-original-title] {\n            text-decoration: underline;\n            -webkit-text-decoration: underline dotted;\n            text-decoration: underline dotted;\n            cursor: help;\n            border-bottom: 0;\n            -webkit-text-decoration-skip-ink: none;\n            text-decoration-skip-ink: none;\n          }\n\n          address {\n            margin-bottom: 16px;\n            font-style: normal;\n            line-height: inherit;\n          }\n\n          ol,\n          ul,\n          dl {\n            margin-top: 0;\n            margin-bottom: 16px;\n          }\n\n          ol ol,\n          ul ul,\n          ol ul,\n          ul ol {\n            margin-bottom: 0;\n          }\n\n          dt {\n            font-weight: 700;\n          }\n\n          dd {\n            margin-bottom: 8px;\n            margin-left: 0;\n          }\n\n          blockquote {\n            margin: 0 0 16px;\n          }\n\n          b,\n          strong {\n            font-weight: bolder;\n          }\n\n          small {\n            font-size: 80%;\n          }\n\n          sub,\n          sup {\n            position: relative;\n            font-size: 75%;\n            line-height: 0;\n            vertical-align: baseline;\n          }\n\n          sub {\n            bottom: -0.25em;\n          }\n\n          sup {\n            top: -0.5em;\n          }\n\n          a {\n            color: #007bff;\n            text-decoration: none;\n            background-color: transparent;\n          }\n\n          a:hover {\n            color: #0056b3;\n            text-decoration: underline;\n          }\n\n          a:not([href]) {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          a:not([href]):hover {\n            color: inherit;\n            text-decoration: none;\n          }\n\n          pre,\n          code,\n          kbd,\n          samp {\n            font-family: SFMono-Regular, Menlo, Monaco, Consolas,\n              'Liberation Mono', 'Courier New', monospace;\n            font-size: 1em;\n          }\n\n          pre {\n            margin-top: 0;\n            margin-bottom: 16px;\n            overflow: auto;\n          }\n\n          figure {\n            margin: 0 0 16px;\n          }\n\n          img {\n            vertical-align: middle;\n            border-style: none;\n          }\n\n          svg {\n            overflow: hidden;\n            vertical-align: middle;\n          }\n\n          table {\n            border-collapse: collapse;\n          }\n\n          caption {\n            padding-top: 12px;\n            padding-bottom: 12px;\n            color: #6c757d;\n            text-align: left;\n            caption-side: bottom;\n          }\n\n          th {\n            text-align: inherit;\n          }\n\n          label {\n            display: inline-block;\n            margin-bottom: 8px;\n          }\n\n          button {\n            border-radius: 0;\n          }\n\n          button:focus {\n            outline: 1px dotted;\n            outline: 5px auto -webkit-focus-ring-color;\n          }\n\n          input,\n          button,\n          select,\n          optgroup,\n          textarea {\n            margin: 0;\n            font-family: inherit;\n            font-size: inherit;\n            line-height: inherit;\n          }\n\n          button,\n          input {\n            overflow: visible;\n          }\n\n          button,\n          select {\n            text-transform: none;\n          }\n\n          select {\n            word-wrap: normal;\n          }\n\n          button,\n          [type='button'],\n          [type='reset'],\n          [type='submit'] {\n            -webkit-appearance: button;\n          }\n\n          button:not(:disabled),\n          [type='button']:not(:disabled),\n          [type='reset']:not(:disabled),\n          [type='submit']:not(:disabled) {\n            cursor: pointer;\n          }\n\n          button::-moz-focus-inner,\n          [type='button']::-moz-focus-inner,\n          [type='reset']::-moz-focus-inner,\n          [type='submit']::-moz-focus-inner {\n            padding: 0;\n            border-style: none;\n          }\n\n          input[type='radio'],\n          input[type='checkbox'] {\n            box-sizing: border-box;\n            padding: 0;\n          }\n\n          input[type='date'],\n          input[type='time'],\n          input[type='datetime-local'],\n          input[type='month'] {\n            -webkit-appearance: listbox;\n          }\n\n          textarea {\n            overflow: auto;\n            resize: vertical;\n          }\n\n          fieldset {\n            min-width: 0;\n            padding: 0;\n            margin: 0;\n            border: 0;\n          }\n\n          legend {\n            display: block;\n            width: 100%;\n            max-width: 100%;\n            padding: 0;\n            margin-bottom: 8px;\n            font-size: 24px;\n            line-height: inherit;\n            color: inherit;\n            white-space: normal;\n          }\n\n          progress {\n            vertical-align: baseline;\n          }\n\n          [type='number']::-webkit-inner-spin-button,\n          [type='number']::-webkit-outer-spin-button {\n            height: auto;\n          }\n\n          [type='search'] {\n            outline-offset: -2px;\n            -webkit-appearance: none;\n          }\n\n          [type='search']::-webkit-search-decoration {\n            -webkit-appearance: none;\n          }\n\n          ::-webkit-file-upload-button {\n            font: inherit;\n            -webkit-appearance: button;\n          }\n\n          output {\n            display: inline-block;\n          }\n\n          summary {\n            display: list-item;\n            cursor: pointer;\n          }\n\n          template {\n            display: none;\n          }\n\n          [hidden] {\n            display: none !important;\n          }\n        "])))}})}n.CssReset=CssReset;var u},518:function(e){"use strict";e.exports=require("next/dist/compiled/strip-ansi")},522:function(e){"use strict";e.exports=require("react")},255:function(e){"use strict";e.exports=require("react-dom")}};var n={};function __nccwpck_require__(t){var r=n[t];if(r!==undefined){return r.exports}var a=n[t]={id:t,loaded:false,exports:{}};var i=true;try{e[t].call(a.exports,a,a.exports,__nccwpck_require__);i=false}finally{if(i)delete n[t]}a.loaded=true;return a.exports}!function(){__nccwpck_require__.nmd=function(e){e.paths=[];if(!e.children)e.children=[];return e}}();if(typeof __nccwpck_require__!=="undefined")__nccwpck_require__.ab=__dirname+"/";var t=__nccwpck_require__(977);module.exports=t})();
